Arima_tidiers {broom}	R Documentation

Tidying methods for ARIMA modeling of time series

Description

These methods tidy the coefficients of ARIMA models of univariate time series.

Usage

## S3 method for class 'Arima'
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

## S3 method for class 'Arima'
glance(x, ...)

Arguments

x
An object of class "Arima"

conf.int
whether to include a confidence interval

conf.level
confidence level of the interval, used only if conf.int=TRUE

...
extra arguments (not used)

Details

augment is not currently implemented, as it is not clear whether ARIMA predictions can or should be merged with the original data frame.

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy returns one row for each coefficient in the model, with five columns:

term
The term in the nonlinear model being estimated and tested

estimate
The estimated coefficient

std.error
The standard error from the linear model

If conf.int = TRUE, also returns

conf.low
low end of confidence interval

conf.high
high end of confidence interval

glance returns one row with the columns

sigma
the square root of the estimated residual variance

logLik
the data's log-likelihood under the model

AIC
the Akaike Information Criterion

BIC
the Bayesian Information Criterion

See Also

arima

Examples

fit <- arima(lh, order = c(1, 0, 0))
tidy(fit)
glance(fit)

[Package broom version 0.4.5 Index]