aareg_tidiers {broom}	R Documentation

Tidiers for aareg objects

Description

These tidy the coefficients of Aalen additive regression objects.

Usage

## S3 method for class 'aareg'
tidy(x, ...)

## S3 method for class 'aareg'
glance(x, ...)

Arguments

x
an "aareg" object

...
extra arguments (not used)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.aareg returns one row for each coefficient, with the columns

term
name of coefficient

estimate
estimate of the slope

statistic
test statistic for coefficient

std.error
standard error of statistic

robust.se
robust version of standard error estimate (only when x was called with dfbeta = TRUE)

z
z score

p.value
p-value

glance returns a one-row data frame containing

statistic
chi-squared statistic

p.value
p-value based on chi-squared statistic

df
degrees of freedom used by coefficients

Examples

if (require("survival", quietly = TRUE)) {
    afit <- aareg(Surv(time, status) ~ age + sex + ph.ecog, data=lung,
                  dfbeta=TRUE)
    summary(afit)
    tidy(afit)
}

[Package broom version 0.4.5 Index]