Tidying methods for anova and AOV objects

Tidies the result of an analysis of variance into an ANOVA table. Only a tidy method is provided, not an augment or glance method.

# S3 method for anova
tidy(x, ...)

# S3 method for aov
tidy(x, ...)

# S3 method for aovlist
tidy(x, ...)

Arguments

x
An object of class "anova", "aov", or "aovlist"

...
extra arguments (not used)

Value

A data.frame with columns

term

Term within the model, or "Residuals"

df

Degrees of freedom used by this term in the model

sumsq

Sum of squares explained by this term

meansq

Mean of sum of squares among degrees of freedom

statistic

F statistic

p.value

P-value from F test

In the case of an "aovlist" object, there is also a stratum column describing the error stratum
Details

Note that the "term" column of an ANOVA table can come with leading or trailing whitespace, which this tidying method trims.

Note that Anova from car (perhaps counter-intuively) outputs an object of class anova for generalized linear models. These objects are also supported.

Examples

a <- anova(lm(mpg ~ wt + qsec + disp, mtcars))
tidy(a)

#>        term df        sumsq       meansq    statistic      p.value
#> 1        wt  1 8.477252e+02 8.477252e+02 1.214366e+02 1.081618e-11
#> 2      qsec  1 8.285831e+01 8.285831e+01 1.186944e+01 1.817334e-03
#> 3      disp  1 1.023935e-03 1.023935e-03 1.466785e-04 9.904229e-01
#> 4 Residuals 28 1.954626e+02 6.980807e+00           NA           NA

a <- aov(mpg ~ wt + qsec + disp, mtcars)
tidy(a)

#>        term df        sumsq       meansq    statistic      p.value
#> 1        wt  1 8.477252e+02 8.477252e+02 1.214366e+02 1.081618e-11
#> 2      qsec  1 8.285831e+01 8.285831e+01 1.186944e+01 1.817334e-03
#> 3      disp  1 1.023935e-03 1.023935e-03 1.466785e-04 9.904229e-01
#> 4 Residuals 28 1.954626e+02 6.980807e+00           NA           NA

al <- aov(mpg ~ wt + qsec + Error(disp / am), mtcars)
tidy(al)

#> Warning: ANOVA F-tests on an essentially perfect fit are unreliable

#> Warning: ANOVA F-tests on an essentially perfect fit are unreliable

#>        term df     sumsq    meansq statistic    p.value stratum
#> 1 Residuals  0   0.00000       NaN        NA         NA    disp
#> 2 Residuals  0   0.00000       NaN        NA         NA disp:am
#> 3      qsec  1  34.16791 34.167910  4.720651 0.03874957  Within
#> 4 Residuals 27 195.42508  7.237966        NA         NA  Within