augment {broom}	R Documentation

Augment data according to a tidied model

Description

Given an R statistical model or other non-tidy object, add columns to the original dataset such as predictions, residuals and cluster assignments.

Usage

augment(x, ...)

Arguments

x
model or other R object to convert to data frame

...
other arguments passed to methods

Details

Note that by convention the first argument is almost always data, which specifies the original data object. This is not part of the S3 signature, partly because it prevents rowwise_df_tidiers from taking a column name as the first argument.

This generic originated in the ggplot2 package, where it was called "fortify."

See Also

augment.lm

[Package broom version 0.4.5 Index]