Augment data with information from a(n) drc object

Augment accepts a model object and a dataset and adds information about each observation in the dataset. Most commonly, this includes predicted values in the .fitted column, residuals in the .resid column, and standard errors for the fitted values in a .se.fit column. New columns always begin with a . prefix to avoid overwriting columns in the original dataset.

Users may pass data to augment via either the data argument or the newdata argument. If the user passes data to the data argument, it must be exactly the data that was used to fit the model object. Pass datasets to newdata to augment data that was not used during model fitting. This still requires that all columns used to fit the model are present.

Augment will often behavior different depending on whether data or newdata is specified. This is because there is often information associated with training observations (such as influences or related) measures that is not meaningfully defined for new observations.

For convenience, many augment methods provide default data arguments, so that augment(fit) will return the augmented training data. In these cases augment tries to reconstruct the original data based on the model object, with some varying degrees of success.

The augmented dataset is always returned as a tibble::tibble with the same number of rows as the passed dataset. This means that the passed data must be coercible to a tibble. At this time, tibbles do not support matrix-columns. This means you should not specify a matrix of covariates in a model formula during the original model fitting process, and that splines::ns(), stats::poly() and survival::Surv() objects are not supported in input data. If you encounter errors, try explicitly passing a tibble, or fitting the original model on data in a tibble.

We are in the process of defining behaviors for models fit with various na.action arguments, but make no guarantees about behavior when data is missing at this time.

# S3 method for drc
augment(x, data = NULL, newdata = NULL, se_fit = FALSE,
  conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x
A drc object produced by a call to drc::drm().

data
A data.frame() or tibble::tibble() containing the original data that was used to produce the object x. Defaults to stats::model.frame(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata
A data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

se_fit
Logical indicating whether or not a .se.fit column should be added to the augmented output. For some models, this calculation can be somwhat time-consuming. Defaults to FALSE.

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

.conf.high

Upper bound on confidence interval for fitted values.

.conf.low

Lower bound on confidence interval for fitted values.

.cooksd

Cooks distance.

.fitted

Fitted or predicted value.

.resid

The difference between fitted and observed values.

.se.fit

Standard errors of fitted values.

Examples

library(drc)

#> Loading required package: MASS

#> 
#> Attaching package: ‘MASS’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     select

#> 
#> 'drc' has been loaded.

#> Please cite R and 'drc' if used for a publication,

#> for references type 'citation()' and 'citation('drc')'.

#> 
#> Attaching package: ‘drc’

#> The following objects are masked from ‘package:stats’:
#> 
#>     gaussian, getInitial

mod <- drm(dead/total~conc, type,
   weights = total, data = selenium, fct = LL.2(), type = "binomial")

tidy(mod)

#> # A tibble: 8 x 6
#>   term  curve estimate std.error statistic  p.value
#>   <chr> <chr>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 b     1       -1.50      0.155     -9.67 2.01e-22
#> 2 b     2       -0.843     0.139     -6.06 1.35e- 9
#> 3 b     3       -2.16      0.138    -15.7  1.65e-55
#> 4 b     4       -1.45      0.169     -8.62 3.41e-18
#> 5 e     1      252.       13.8       18.2  1.16e-74
#> 6 e     2      378.       39.4        9.61 3.53e-22
#> 7 e     3      120.        5.91      20.3  1.14e-91
#> 8 e     4       88.8       8.62      10.3  3.28e-25

tidy(mod, conf.int = TRUE)

#> # A tibble: 8 x 8
#>   term  curve estimate std.error statistic  p.value conf.low conf.high
#>   <chr> <chr>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 b     1       -1.50      0.155     -9.67 2.01e-22       NA        NA
#> 2 b     2       -0.843     0.139     -6.06 1.35e- 9       NA        NA
#> 3 b     3       -2.16      0.138    -15.7  1.65e-55       NA        NA
#> 4 b     4       -1.45      0.169     -8.62 3.41e-18       NA        NA
#> 5 e     1      252.       13.8       18.2  1.16e-74       NA        NA
#> 6 e     2      378.       39.4        9.61 3.53e-22       NA        NA
#> 7 e     3      120.        5.91      20.3  1.14e-91       NA        NA
#> 8 e     4       88.8       8.62      10.3  3.28e-25       NA        NA

glance(mod)

#> # A tibble: 1 x 4
#>     AIC   BIC logLik df.residual
#>   <dbl> <dbl>  <dbl>       <int>
#> 1  768.  778.  -376.          17

augment(mod, selenium)

#> # A tibble: 25 x 7
#>     type  conc total  dead .fitted  .resid    .cooksd
#>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl>      <dbl>
#>  1     1     0   151     3   0      0.0199 0         
#>  2     1   100   146    40   0.199  0.0748 0.0000909 
#>  3     1   200   116    31   0.414 -0.146  0.000104  
#>  4     1   300   159    85   0.565 -0.0302 0.00000516
#>  5     1   400   150   102   0.667  0.0133 0.00000220
#>  6     1   500   140   112   0.737  0.0633 0.0000720 
#>  7     2     0   141     2   0      0.0142 0         
#>  8     2   100   153    30   0.246 -0.0495 0.000168  
#>  9     2   200   142    59   0.369  0.0468 0.0000347 
#> 10     2   300   139    82   0.451  0.139  0.0000430 
#> # … with 15 more rows