Augment data with information from a(n) fixest object

Augment accepts a model object and a dataset and adds information about each observation in the dataset. Most commonly, this includes predicted values in the .fitted column, residuals in the .resid column, and standard errors for the fitted values in a .se.fit column. New columns always begin with a . prefix to avoid overwriting columns in the original dataset.

Users may pass data to augment via either the data argument or the newdata argument. If the user passes data to the data argument, it must be exactly the data that was used to fit the model object. Pass datasets to newdata to augment data that was not used during model fitting. This still requires that at least all predictor variable columns used to fit the model are present. If the original outcome variable used to fit the model is not included in newdata, then no .resid column will be included in the output.

Augment will often behave differently depending on whether data or newdata is given. This is because there is often information associated with training observations (such as influences or related) measures that is not meaningfully defined for new observations.

For convenience, many augment methods provide default data arguments, so that augment(fit) will return the augmented training data. In these cases, augment tries to reconstruct the original data based on the model object with varying degrees of success.

The augmented dataset is always returned as a tibble::tibble with the same number of rows as the passed dataset. This means that the passed data must be coercible to a tibble. At this time, tibbles do not support matrix-columns. This means you should not specify a matrix of covariates in a model formula during the original model fitting process, and that splines::ns(), stats::poly() and survival::Surv() objects are not supported in input data. If you encounter errors, try explicitly passing a tibble, or fitting the original model on data in a tibble.

We are in the process of defining behaviors for models fit with various na.action arguments, but make no guarantees about behavior when data is missing at this time.

# S3 method for fixest
augment(
  x,
  data = NULL,
  newdata = NULL,
  type.predict = c("link", "response"),
  type.residuals = c("response", "deviance", "pearson", "working"),
  ...
)

Arguments

x
A fixest object returned from any of the fixest estimators

data
A base::data.frame or tibble::tibble() containing the original data that was used to produce the object x. Defaults to stats::model.frame(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata
A base::data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

type.predict
Passed to predict.fixest type argument. Defaults to "link" (like predict.glm).

type.residuals
Passed to predict.fixest type argument. Defaults to "response" (like residuals.lm, but unlike residuals.glm).

...
Additional arguments passed to summary and confint. Important arguments are se and cluster. Other arguments are dof, exact_dof, forceCovariance, and keepBounded. See summary.fixest.

Note

Important note: fixest models do not include a copy of the input data, so you must provide it manually.

augment.fixest only works for fixest::feols(), fixest::feglm(), and fixest::femlm() models. It does not work with results from fixest::fenegbin(), fixest::feNmlm(), or fixest::fepois().

See also

Value

A tibble::tibble() with columns:

.fitted

Fitted or predicted value.

.resid

The difference between observed and fitted values.

Examples

# \donttest{
library(fixest)
gravity <- feols(log(Euros) ~ log(dist_km) | Origin + Destination + Product + Year, trade)

tidy(gravity)

#> # A tibble: 1 x 5
#>   term         estimate std.error statistic       p.value
#>   <chr>           <dbl>     <dbl>     <dbl>         <dbl>
#> 1 log(dist_km)    -2.17     0.154     -14.1 0.00000000119

glance(gravity)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared within.r.squared pseudo.r.squared sigma  nobs     AIC
#>       <dbl>         <dbl>            <dbl>            <dbl> <dbl> <int>   <dbl>
#> 1     0.706         0.705            0.219               NA  1.74 38325 151481.
#> # … with 2 more variables: BIC <dbl>, logLik <dbl>

augment(gravity, trade)

#> # A tibble: 38,325 x 9
#>    .rownames Destination Origin Product  Year dist_km    Euros .fitted  .resid
#>    <chr>     <fct>       <fct>    <int> <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 1         LU          BE           1  2007    140.  2966697    14.1  0.812 
#>  2 2         BE          LU           1  2007    140.  6755030    13.0  2.75  
#>  3 3         LU          BE           2  2007    140. 57078782    16.9  0.924 
#>  4 4         BE          LU           2  2007    140.  7117406    15.8 -0.0470
#>  5 5         LU          BE           3  2007    140. 17379821    16.3  0.378 
#>  6 6         BE          LU           3  2007    140.  2622254    15.2 -0.402 
#>  7 7         LU          BE           4  2007    140. 64867588    17.4  0.595 
#>  8 8         BE          LU           4  2007    140. 10731757    16.3 -0.0937
#>  9 9         LU          BE           5  2007    140.   330702    14.1 -1.37  
#> 10 10        BE          LU           5  2007    140.     7706    13.0 -4.02  
#> # … with 38,315 more rows

## To get robust or clustered SEs, users can either:
# 1) Or, specify the arguments directly in the tidy() call
tidy(gravity, conf.int = TRUE, cluster = c("Product", "Year"))

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic  p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17    0.0760     -28.5 3.88e-10    -2.32     -2.02

tidy(gravity, conf.int = TRUE, se = "threeway")

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic       p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>         <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17     0.175     -12.4 0.00000000608    -2.51     -1.83

# 2) Feed tidy() a summary.fixest object that has already accepted these arguments
gravity_summ <- summary(gravity, cluster = c("Product", "Year"))
tidy(gravity_summ, conf.int = TRUE)

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic  p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17    0.0760     -28.5 3.88e-10    -2.32     -2.02

# Approach (1) is preferred.

## The other fixest methods all work similarly. For example:
gravity_pois <- feglm(Euros ~ log(dist_km) | Origin + Destination + Product + Year, trade)
tidy(gravity_pois)

#> # A tibble: 1 x 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 log(dist_km)    -1.53     0.116     -13.2 7.89e-40

glance(gravity_pois)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared within.r.squared pseudo.r.squared sigma  nobs     AIC
#>       <dbl>         <dbl>            <dbl>            <dbl> <dbl> <int>   <dbl>
#> 1        NA            NA               NA            0.764    NA 38325 1.40e12
#> # … with 2 more variables: BIC <dbl>, logLik <dbl>

augment(gravity_pois, trade)

#> # A tibble: 38,325 x 9
#>    .rownames Destination Origin Product  Year dist_km    Euros .fitted    .resid
#>    <chr>     <fct>       <fct>    <int> <dbl>   <dbl>    <dbl>   <dbl>     <dbl>
#>  1 1         LU          BE           1  2007    140.  2966697    16.0   -6.06e6
#>  2 2         BE          LU           1  2007    140.  6755030    15.4    1.97e6
#>  3 3         LU          BE           2  2007    140. 57078782    17.4    2.00e7
#>  4 4         BE          LU           2  2007    140.  7117406    16.8   -1.26e7
#>  5 5         LU          BE           3  2007    140. 17379821    16.7   -1.00e4
#>  6 6         BE          LU           3  2007    140.  2622254    16.0   -6.60e6
#>  7 7         LU          BE           4  2007    140. 64867588    17.5    2.64e7
#>  8 8         BE          LU           4  2007    140. 10731757    16.8   -9.64e6
#>  9 9         LU          BE           5  2007    140.   330702    14.5   -1.64e6
#> 10 10        BE          LU           5  2007    140.     7706    13.9   -1.04e6
#> # … with 38,315 more rows

# }