augment.nlrq {broom}	R Documentation

Tidy a(n) nlrq object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'nlrq'
augment(x, ...)

Arguments

x
A nlrq object returned from quantreg::nlrq().

...
Arguments passed on to augment.nls

data

A data.frame() or tibble::tibble() containing the original data that was used to produce the object x. Defaults to stats::model.frame(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata

A data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

x

An nls object returned from stats::nls().

Value

A tibble::tibble() containing the data passed to augment, and additional columns:

.fitted
The predicted response for that observation.

.resid
The residual for a particular point. Present only when data has been passed to augment via the data argument.

See Also

augment(), quantreg::nlrq()

Other quantreg tidiers: augment.rqs, augment.rq, glance.nlrq, glance.rq, tidy.nlrq, tidy.rqs, tidy.rq

[Package broom version 0.5.0 Index]