Tidying method for a bootstrapped temporal exponential random graph model

This method tidies the coefficients of a bootstrapped temporal exponential random graph model estimated with the xergm. It simply returns the coefficients and their confidence intervals.

# S3 method for btergm
tidy(x, conf.level = 0.95, exponentiate = FALSE,
  quick = FALSE, ...)

Arguments

x
a btergm object

conf.level
confidence level of the bootstrapped interval

exponentiate
whether to exponentiate the coefficient estimates and confidence intervals

quick
whether to compute a smaller and faster version, containing only the term and estimate columns.

...
extra arguments (currently not used)

Value

A data.frame without rownames.

tidy.btergm returns one row for each coefficient, with four columns:

term

The term in the model being estimated and tested

estimate

The estimated coefficient

conf.low

The lower bound of the confidence interval

conf.high

The lower bound of the confidence interval

Details

There is no augment or glance method for ergm objects.

See also

Examples

if (require("xergm")) {
    # Using the same simulated example as the xergm package
    # Create 10 random networks with 10 actors
    networks <- list()
    for(i in 1:10){
        mat <- matrix(rbinom(100, 1, .25), nrow = 10, ncol = 10)
        diag(mat) <- 0
        nw <- network::network(mat)
        networks[[i]] <- nw
    }
    # Create 10 matrices as covariates
    covariates <- list()
    for (i in 1:10) {
        mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
        covariates[[i]] <- mat
    }
    # Fit a model where the propensity to form ties depends
    # on the edge covariates, controlling for the number of
    # in-stars
    btfit <- btergm(networks ~ edges + istar(2) +
                      edgecov(covariates), R = 100)

    # Show terms, coefficient estimates and errors
    tidy(btfit)

    # Show coefficients as odds ratios with a 99% CI
    tidy(btfit, exponentiate = TRUE, conf.level = 0.99)
}

#> Loading required package: xergm

#> Warning: package 'xergm' was built under R version 3.4.3

#> Loading required package: xergm.common

#> Warning: package 'xergm.common' was built under R version 3.4.3

#> Loading required package: ergm

#> Warning: package 'ergm' was built under R version 3.4.3

#> Loading required package: statnet.common

#> 
#> Attaching package: 'statnet.common'

#> The following object is masked from 'package:base':
#> 
#>     order

#> Loading required package: network

#> Warning: package 'network' was built under R version 3.4.4

#> network: Classes for Relational Data
#> Version 1.13.0.1 created on 2015-08-31.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> ergm: version 3.8.0, created on 2017-08-18
#> Copyright (c) 2017, Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Carter T. Butts, University of California -- Irvine
#>                     Steven M. Goodreau, University of Washington
#>                     Pavel N. Krivitsky, University of Wollongong
#>                     Martina Morris, University of Washington
#>                     with contributions from
#>                     Li Wang
#>                     Kirk Li, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#> Based on "statnet" project software (statnet.org).
#> For license and citation information see statnet.org/attribution
#> or type citation("ergm").

#> NOTE: Versions before 3.6.1 had a bug in the implementation of the bd()
#> constriant which distorted the sampled distribution somewhat. In
#> addition, Sampson's Monks datasets had mislabeled vertices. See the
#> NEWS and the documentation for more details.

#> 
#> Attaching package: 'xergm.common'

#> The following object is masked from 'package:ergm':
#> 
#>     gof

#> Loading required package: btergm

#> Warning: package 'btergm' was built under R version 3.4.4

#> Package:  btergm
#> Version:  1.9.1
#> Date:     2018-02-14
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)
#>           Bruce A. Desmarais (Pennsylvania State University)

#> Loading required package: tnam

#> Warning: package 'tnam' was built under R version 3.4.3

#> Package:  tnam
#> Version:  1.6.5
#> Date:     2017-03-31
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)

#> Loading required package: rem

#> Warning: package 'rem' was built under R version 3.4.3

#> Loading required package: GERGM

#> Warning: package 'GERGM' was built under R version 3.4.3

#> GERGM: Generalized Exponential Random Graph Models
#> Version 0.11.2 created on 2017-03-14.
#> copyright (c) 2017, Matthew J. Denny, Penn State University
#>                     James D. Wilson, University of San Francisco
#>                     Skyler Cranmer, Ohio State University
#>                     Bruce A. Desmarais, Penn State University
#>                     Shankar Bhamidi, University of North Carolina
#> Type help('gergm') to get started.
#> Development website: https://github.com/matthewjdenny/GERGM

#> Package:  xergm
#> Version:  1.8.2
#> Date:     2017-04-01
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)
#>           Bruce A. Desmarais (Pennsylvania State University)
#> 
#> Please cite the xergm package in your publications -- see citation("xergm").

#> 
#> Initial dimensions of the network and covariates:

#>                  t=1 t=2 t=3 t=4 t=5 t=6 t=7 t=8 t=9 t=10
#> networks (row)    10  10  10  10  10  10  10  10  10   10
#> networks (col)    10  10  10  10  10  10  10  10  10   10
#> covariates (row)  10  10  10  10  10  10  10  10  10   10
#> covariates (col)  10  10  10  10  10  10  10  10  10   10

#> 
#> All networks are conformable.

#> 
#> Dimensions of the network and covariates after adjustment:

#>                  t=1 t=2 t=3 t=4 t=5 t=6 t=7 t=8 t=9 t=10
#> networks (row)    10  10  10  10  10  10  10  10  10   10
#> networks (col)    10  10  10  10  10  10  10  10  10   10
#> covariates (row)  10  10  10  10  10  10  10  10  10   10
#> covariates (col)  10  10  10  10  10  10  10  10  10   10

#> 
#> Starting pseudolikelihood estimation with 100 bootstrapping replications on a single computing core...

#> Done.

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#>                      term  estimate  conf.low conf.high
#> 1                   edges 0.2943735 0.1682938 0.8107676
#> 2                  istar2 1.0760679 0.7581794 1.2903869
#> 3 edgecov.covariates[[i]] 1.0442169 0.8496848 1.4029402