confint.geeglm {broom}	R Documentation

Confidence interval for geeglm objects

Description

Generate confidence intervals for GEE analyses

Usage

## S3 method for class 'geeglm'
confint(object, parm, level = 0.95, ...)

Arguments

object
The 'geeglm' object

parm
The parameter to calculate the confidence interval for. If not specified, the default is to calculate a confidence interval on all parameters (all variables in the model).

level
confidence level of the interval, used only if conf.int=TRUE

...
Additional parameters

Details

This function was taken from http://stackoverflow.com/a/21221995/2632184.

Value

Returns the upper and lower confidence intervals

[Package broom version 0.4.5 Index]