confint_tidy {broom}	R Documentation

Calculate confidence interval as a tidy data frame

Description

Return a confidence interval as a tidy data frame. This directly wraps the confint function, but ensures it folllows broom conventions: column names of conf.low and conf.high, and no row names

Usage

confint_tidy(x, conf.level = 0.95, func = stats::confint, ...)

Arguments

x
a model object for which confint can be calculated

conf.level
confidence level

func
Function to use for computing confint

...
extra arguments passed on to confint

Value

A data frame with two columns: conf.low and conf.high.

See Also

confint

[Package broom version 0.4.5 Index]