finish_glance {broom}	R Documentation

Add logLik, AIC, BIC, and other common measurements to a glance of a prediction

Description

A helper function for several functions in the glance generic. Methods such as logLik, AIC, and BIC are defined for many prediction objects, such as lm, glm, and nls. This is a helper function that adds them to a glance data.frame can be performed. If any of them cannot be computed, it fails quietly.

Usage

finish_glance(ret, x)

Arguments

ret
a one-row data frame (a partially complete glance)

x
the prediction model

Details

In one special case, deviance for objects of the lmerMod class from lme4 is computed with deviance(x, REML=FALSE).

Value

a one-row data frame with additional columns added, such as

logLik
log likelihoods

AIC
Akaike Information Criterion

BIC
Bayesian Information Criterion

deviance
deviance

df.residual
residual degrees of freedom

Each of these are produced by the corresponding generics

[Package broom version 0.5.0 Index]