fix_data_frame {broom}	R Documentation

Ensure an object is a data frame, with rownames moved into a column

Description

Ensure an object is a data frame, with rownames moved into a column

Usage

fix_data_frame(x, newnames = NULL, newcol = "term")

Arguments

x
a data.frame or matrix

newnames
new column names, not including the rownames

newcol
the name of the new rownames column

Value

a data.frame, with rownames moved into a column and new column names assigned

[Package broom version 0.5.0 Index]