gamlss_tidiers {broom}	R Documentation

Tidying methods for gamlss objects

Description

Tidying methods for "gamlss" objects from the gamlss package.

Usage

## S3 method for class 'gamlss'
tidy(x, quick = FALSE, ...)

Arguments

x
A "gamlss" object

quick
Whether to perform a fast version, and return only the coefficients

...
Extra arguments (not used)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

A data.frame with one row for each coefficient, containing columns

parameter
Type of coefficient being estimated: mu, sigma, nu, or tau

term
The term in the model being estimated and tested

estimate
The estimated coefficient

std.error
The standard error from the linear model

statistic
t-statistic

p.value
two-sided p-value

if (requireNamespace("gamlss", quietly = TRUE)) data(abdom) mod<-gamlss(y~pb(x),sigma.fo=~pb(x),family=BCT, data=abdom, method=mixed(1,20))

tidy(mod)

[Package broom version 0.4.5 Index]