Tidying methods for a GARCH model (tseries package)

This tidies the result of a coefficient of the GARCH model implemented in tseries package.

# S3 method for garch
tidy(x, ...)

# S3 method for garch
glance(x, method = c("box-ljung-test", "jarque-bera-test"),
  ...)

Arguments

x
garch object

...
extra arguments (not used)

method
character which specifies the hypothesis test to be shown in glance. The garch function reports 2 hypothesis tests: Jarque-Bera to residuals and Box-Ljung to squared residuals.

data
original data (used with augment)

newdata
new data provided for predition use (used with augment)

Value

A data.frame with one row for each coefficient, with five columns:

term

The term in the linear model being estimated and tested

estimate

The estimated coefficient

std.error

The standard error

statistic

test statistic

p.value

p-value

A data.frame with one row, with seven columns:
statistic

Test statistic used to compute the p-value

p.value

P-value

parameter

Parameter field in the htest, typically degrees of freedom

method

Method used to compute the statistic as a string

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

Examples

if (require("tseries", quietly = TRUE)) {
    data(EuStockMarkets)
    dax <- diff(log(EuStockMarkets))[,"DAX"]
    dax.garch <- garch(dax)
    dax.garch
    tidy(dax.garch)
    glance(dax.garch)
    head(augment(dax.garch, dax))
    smp1 = window(dax, end = c(1997, frequency(dax)))
    smp2 = window(dax, start = c(1998, 1))
    dax.garch <- garch(smp1)
    augment(dax.garch, newdata = smp2)
}

#> 
#>  ***** ESTIMATION WITH ANALYTICAL GRADIENT ***** 
#> 
#> 
#>      I     INITIAL X(I)        D(I)
#> 
#>      1     9.549651e-05     1.000e+00
#>      2     5.000000e-02     1.000e+00
#>      3     5.000000e-02     1.000e+00
#> 
#>     IT   NF      F         RELDF    PRELDF    RELDX   STPPAR   D*STEP   NPRELDF
#>      0    1 -7.584e+03
#>      1    8 -7.585e+03  1.45e-05  2.60e-05  1.4e-05  1.0e+11  1.4e-06  1.35e+06
#>      2    9 -7.585e+03  1.88e-07  1.97e-07  1.3e-05  2.0e+00  1.4e-06  1.50e+00
#>      3   18 -7.589e+03  6.22e-04  1.10e-03  3.5e-01  2.0e+00  5.5e-02  1.50e+00
#>      4   21 -7.601e+03  1.58e-03  1.81e-03  6.2e-01  1.9e+00  2.2e-01  3.07e-01
#>      5   23 -7.634e+03  4.22e-03  3.55e-03  4.3e-01  9.6e-01  4.4e-01  3.06e-02
#>      6   25 -7.646e+03  1.61e-03  1.85e-03  2.9e-02  2.0e+00  4.4e-02  5.43e-02
#>      7   27 -7.646e+03  3.82e-05  5.23e-04  1.3e-02  2.0e+00  2.0e-02  1.46e-02
#>      8   28 -7.648e+03  1.86e-04  1.46e-04  6.5e-03  2.0e+00  9.9e-03  1.54e-03
#>      9   29 -7.648e+03  3.12e-05  4.83e-05  6.4e-03  2.0e+00  9.9e-03  3.34e-03
#>     10   30 -7.648e+03  1.39e-05  6.31e-05  6.2e-03  1.9e+00  9.9e-03  1.86e-03
#>     11   31 -7.650e+03  2.70e-04  3.24e-04  6.0e-03  1.9e+00  9.9e-03  4.99e-03
#>     12   34 -7.656e+03  8.42e-04  8.57e-04  2.2e-02  1.7e-01  3.9e-02  2.22e-03
#>     13   36 -7.661e+03  6.12e-04  6.40e-04  1.9e-02  4.2e-01  3.9e-02  2.09e-03
#>     14   38 -7.665e+03  4.87e-04  8.63e-04  4.9e-02  4.1e-01  9.6e-02  9.69e-04
#>     15   48 -7.666e+03  1.02e-04  1.86e-04  1.9e-07  4.5e+00  3.5e-07  3.94e-04
#>     16   49 -7.666e+03  1.12e-07  1.01e-07  1.9e-07  2.0e+00  3.5e-07  6.22e-05
#>     17   57 -7.666e+03  1.60e-05  2.70e-05  2.0e-03  9.3e-01  3.7e-03  6.10e-05
#>     18   59 -7.666e+03  5.23e-06  7.01e-06  3.7e-03  3.9e-01  8.0e-03  7.77e-06
#>     19   60 -7.666e+03  4.08e-08  3.74e-08  1.4e-04  0.0e+00  3.1e-04  3.74e-08
#>     20   61 -7.666e+03  2.31e-09  8.57e-10  8.6e-06  0.0e+00  2.0e-05  8.57e-10
#>     21   62 -7.666e+03  5.35e-11  2.25e-13  7.6e-07  0.0e+00  1.6e-06  2.25e-13
#>     22   63 -7.666e+03  1.81e-12  7.06e-16  1.7e-08  0.0e+00  3.4e-08  7.06e-16
#>     23   64 -7.666e+03  6.98e-14  1.69e-17  1.0e-09  0.0e+00  2.4e-09  1.69e-17
#>     24   65 -7.666e+03 -1.16e-14  1.76e-20  1.9e-10  0.0e+00  4.0e-10  1.76e-20
#> 
#>  ***** X- AND RELATIVE FUNCTION CONVERGENCE *****
#> 
#>  FUNCTION    -7.665775e+03   RELDX        1.874e-10
#>  FUNC. EVALS      65         GRAD. EVALS      24
#>  PRELDF       1.760e-20      NPRELDF      1.760e-20
#> 
#>      I      FINAL X(I)        D(I)          G(I)
#> 
#>      1    4.639289e-06     1.000e+00    -2.337e-02
#>      2    6.832875e-02     1.000e+00    -8.294e-07
#>      3    8.890666e-01     1.000e+00    -2.230e-06
#> 

#> Error: augment doesn't know how to deal with data of class garch