glance.Mclust {broom}	R Documentation

Glance at a(n) Mclust object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Usage

## S3 method for class 'Mclust'
glance(x, ...)

Arguments

x
An Mclust object return from mclust::Mclust().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

model
A character string denoting the model at which the optimal BIC occurs

n
The number of observations in the data

G
The optimal number of mixture components

BIC
The optimal BIC value

logLik
The log-likelihood corresponding to the optimal BIC

df
The number of estimated parameters

hypvol
If the other model contains a noise component, the value of the hypervolume parameter. Otherwise NA.

[Package broom version 0.5.0 Index]