Glance at a(n) ergm object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for ergm
glance(x, deviance = FALSE, mcmc = FALSE, ...)

Arguments

x
An ergm object returned from a call to ergm::ergm().

deviance
Logical indicating whether or not to report null and residual deviance for the model, as well as degrees of freedom. Defaults to FALSE.

mcmc
Logical indicating whether or not to report MCMC interval, burn-in and sample size used to estimate the model. Defaults to FALSE.

...
Additional arguments to pass to ergm::summary.ergm(). Cautionary note: Mispecified arguments may be silently ignored.

Value

glance.ergm returns a one-row data.frame with the columns

independence

Whether the model assumed dyadic independence

iterations

The number of iterations performed before convergence

logLik

If applicable, the log-likelihood associated with the model

AIC

The Akaike Information Criterion

BIC

The Bayesian Information Criterion

If deviance = TRUE, and if the model supports it, the data frame will also contain the columns
null.deviance

The null deviance of the model

df.null

The degrees of freedom of the null deviance

residual.deviance

The residual deviance of the model

df.residual

The degrees of freedom of the residual deviance

Last, if mcmc = TRUE, the data frame will also contain the columns
MCMC.interval

The interval used during MCMC estimation

MCMC.burnin

The burn-in period of the MCMC estimation

MCMC.samplesize

The sample size used during MCMC estimation

See also