glance.garch {broom}	R Documentation

Tidy a(n) garch object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'garch'
glance(x, test = c("box-ljung-test", "jarque-bera-test"),
  ...)

Arguments

x
A garch object returned by tseries::garch().

test
Character specification of which hypothesis test to use. The garch function reports 2 hypothesis tests: Jarque-Bera to residuals and Box-Ljung to squared residuals.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

statistic
Test statistic used to compute the p-value

p.value
P-value

parameter
Parameter field in the htest, typically degrees of freedom

method
Method used to compute the statistic as a string

logLik
the data's log-likelihood under the model

AIC
the Akaike Information Criterion

BIC
the Bayesian Information Criterion

See Also

glance(), tseries::garch(), []

Other garch tidiers: tidy.garch

[Package broom version 0.5.0 Index]