Glance at a(n) glm object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for glm
glance(x, ...)

Arguments

x
A glm object returned from stats::glm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

null.deviance

the deviance of the null model

df.null

the residual degrees of freedom for the null model

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

deviance

deviance

df.residual

residual degrees of freedom

See also

Examples

g <- glm(am ~ mpg, mtcars, family = "binomial")
glance(g)

#> # A tibble: 1 x 7
#>   null.deviance df.null logLik   AIC   BIC deviance df.residual
#>           <dbl>   <int>  <dbl> <dbl> <dbl>    <dbl>       <int>
#> 1          43.2      31  -14.8  33.7  36.6     29.7          30