Glance at a(n) gmm object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for gmm
glance(x, ...)

Arguments

x
A gmm object returned from gmm::gmm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with exactly one row and columns:

df

Degrees of freedom used by the model.

df.residual

Residual degrees of freedom.

nobs

Number of observations used.

p.value

P-value corresponding to the test statistic.

statistic

Test statistic.

Examples

library(gmm)

#> Error: package or namespace load failed for ‘gmm’ in dyn.load(file, DLLpath = DLLpath, ...):
#>  unable to load shared object '/Users/runner/runners/2.263.0/work/_temp/Library/gmm/libs/gmm.so':
#>   dlopen(/Users/runner/runners/2.263.0/work/_temp/Library/gmm/libs/gmm.so, 6): Library not loaded: /usr/local/gfortran/lib/libgfortran.3.dylib
#>   Referenced from: /Users/runner/runners/2.263.0/work/_temp/Library/gmm/libs/gmm.so
#>   Reason: image not found

# examples come from the "gmm" package
## CAPM test with GMM
data(Finance)

#> Warning: data set ‘Finance’ not found

r <- Finance[1:300, 1:10]

#> Error in eval(expr, envir, enclos): object 'Finance' not found

rm <- Finance[1:300, "rm"]

#> Error in eval(expr, envir, enclos): object 'Finance' not found

rf <- Finance[1:300, "rf"]

#> Error in eval(expr, envir, enclos): object 'Finance' not found

z <- as.matrix(r - rf)

#> Error in as.matrix(r - rf): object 'r' not found

t <- nrow(z)

#> Error in nrow(z): object 'z' not found

zm <- rm - rf

#> Error in rm - rf: non-numeric argument to binary operator

h <- matrix(zm, t, 1)

#> Error in matrix(zm, t, 1): object 'zm' not found

res <- gmm(z ~ zm, x = h)

#> Error in gmm(z ~ zm, x = h): could not find function "gmm"

# tidy result
tidy(res)

#> Error in tidy(res): object 'res' not found

tidy(res, conf.int = TRUE)

#> Error in tidy(res, conf.int = TRUE): object 'res' not found

tidy(res, conf.int = TRUE, conf.level = .99)

#> Error in tidy(res, conf.int = TRUE, conf.level = 0.99): object 'res' not found

# coefficient plot
library(ggplot2)
library(dplyr)

tidy(res, conf.int = TRUE) %>%
  mutate(variable = reorder(term, estimate)) %>%
  ggplot(aes(estimate, variable)) +
  geom_point() +
  geom_errorbarh(aes(xmin = conf.low, xmax = conf.high)) +
  geom_vline(xintercept = 0, color = "red", lty = 2)

#> Error in tidy(res, conf.int = TRUE): object 'res' not found

# from a function instead of a matrix
g <- function(theta, x) {
  e <- x[, 2:11] - theta[1] - (x[, 1] - theta[1]) %*% matrix(theta[2:11], 1, 10)
  gmat <- cbind(e, e * c(x[, 1]))
  return(gmat)
}

x <- as.matrix(cbind(rm, r))

#> Error in cbind(rm, r): object 'r' not found

res_black <- gmm(g, x = x, t0 = rep(0, 11))

#> Error in gmm(g, x = x, t0 = rep(0, 11)): could not find function "gmm"

tidy(res_black)

#> Error in tidy(res_black): object 'res_black' not found

tidy(res_black, conf.int = TRUE)

#> Error in tidy(res_black, conf.int = TRUE): object 'res_black' not found

## APT test with Fama-French factors and GMM

f1 <- zm

#> Error in eval(expr, envir, enclos): object 'zm' not found

f2 <- Finance[1:300, "hml"] - rf

#> Error in eval(expr, envir, enclos): object 'Finance' not found

f3 <- Finance[1:300, "smb"] - rf

#> Error in eval(expr, envir, enclos): object 'Finance' not found

h <- cbind(f1, f2, f3)

#> Error in cbind(f1, f2, f3): object 'f1' not found

res2 <- gmm(z ~ f1 + f2 + f3, x = h)

#> Error in gmm(z ~ f1 + f2 + f3, x = h): could not find function "gmm"

td2 <- tidy(res2, conf.int = TRUE)

#> Error in tidy(res2, conf.int = TRUE): object 'res2' not found

td2

#> Error in eval(expr, envir, enclos): object 'td2' not found

# coefficient plot
td2 %>%
  mutate(variable = reorder(term, estimate)) %>%
  ggplot(aes(estimate, variable)) +
  geom_point() +
  geom_errorbarh(aes(xmin = conf.low, xmax = conf.high)) +
  geom_vline(xintercept = 0, color = "red", lty = 2)

#> Error in eval(lhs, parent, parent): object 'td2' not found