glance.ivreg {broom}	R Documentation

Glance at a(n) ivreg object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Usage

## S3 method for class 'ivreg'
glance(x, diagnostics = FALSE, ...)

Arguments

x
An ivreg object created by a call to AER::ivreg().

diagnostics
Logical indicating whether to include statistics and p-values for Sargan, Wu-Hausman and weak instrument tests. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble with columns

r.squared
The percent of variance explained by the model

adj.r.squared
r.squared adjusted based on the degrees of freedom

sigma
The square root of the estimated residual variance

statistic
Wald test statistic

p.value
p-value from the Wald test

df
Degrees of freedom used by the coefficients

df.residual
residual degrees of freedom

If diagnostics = TRUE, will also return the following columns:

statistic.Sargan
Statistic for Sargan test

p.value.Sargan
P-value for Sargan test

statistic.Wu.Hausman
Statistic for Wu-Hausman test

p.value.Wu.Hausman
P-value for Wu-Hausman test

statistic.weakinst
Statistic for Wu-Hausman test

p.value.weakinst
P-value for weak instruments test

See Also

glance(), AER::ivreg()

Other ivreg tidiers: augment.ivreg, tidy.ivreg

Examples

library(AER)

data("CigarettesSW", package = "AER")
ivr <- ivreg(
  log(packs) ~ income | population,
  data = CigarettesSW,
  subset = year == "1995"
)

summary(ivr)

tidy(ivr)
tidy(ivr, conf.int = TRUE)
tidy(ivr, conf.int = TRUE, exponentiate = TRUE)

augment(ivr)

glance(ivr)

[Package broom version 0.5.0 Index]