Glance at a(n) lavaan object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for lavaan
glance(x, ...)

Arguments

x
A lavaan object, such as those return from lavaan::cfa(), and lavaan::sem().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

chisq

Model chi squared

npar

Number of parameters in the model

rmsea

Root mean square error of approximation

rmsea.conf.high

95 percent upper bound on RMSEA

srmr

Standardised root mean residual

agfi

Adjusted goodness of fit

cfi

Comparative fit index

tli

Tucker Lewis index

aic

Akaike information criterion

bic

Bayesian information criterion

ngroups

Number of groups in model

nobs

Number of observations included

norig

Number of observation in the original dataset

nexcluded

Number of excluded observations

converged

Logical - Did the model converge

estimator

Estimator used

missing_method

Method for eliminating missing data

For further recommendations on reporting SEM and CFA models see Schreiber, J. B. (2017). Update to core reporting practices in structural equation modeling. Research in Social and Administrative Pharmacy, 13(3), 634-643. https://doi.org/10.1016/j.sapharm.2016.06.006
See also

Examples

if (require("lavaan", quietly = TRUE)) {

 library(lavaan)

 cfa.fit <- cfa(
   'F =~ x1 + x2 + x3 + x4 + x5',
   data = HolzingerSwineford1939, group = "school"
 )
 glance(cfa.fit)

}

#> This is lavaan 0.6-1

#> lavaan is BETA software! Please report any bugs.

#> # A tibble: 1 x 17
#>    agfi   aic   bic   cfi chisq  npar rmsea rmsea.conf.high  srmr   tli
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>           <dbl> <dbl> <dbl>
#> 1 0.971 4473. 4584. 0.766  99.3    30 0.244           0.288 0.115 0.533
#> # ... with 7 more variables: converged <lgl>, estimator <chr>, ngroups <int>,
#> #   missing_method <chr>, nobs <int>, norig <int>, nexcluded <int>