glance.lmRob {broom}	R Documentation

Glance at a(n) lmRob object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'lmRob'
glance(x, ...)

Arguments

x
A lmRob object returned from robust::lmRob().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

Value

A tibble::tibble() with exactly one row and columns:

deviance
Deviance of the model.

df.residual
Residual degrees of freedom.

nobs
Number of observations used.

r.squared
R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

sigma
Estimated standard error of the residuals.

See Also

robust::lmRob()

Other robust tidiers: augment.lmRob(), glance.glmRob(), tidy.glmRob(), tidy.lmRob()

Examples

 # load modeling library
 library(robust)
  
 # fit model
 m <- lmRob(mpg ~ wt, data = mtcars)

 # summarize model fit with tidiers
 tidy(m)
 augment(m)
 glance(m)

[Package broom version 1.0.0 Index]