Glance at a(n) margins object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

# S3 method for margins
glance(x, ...)

Arguments

x

A margins object returned from margins::margins().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

adj.r.squared

Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.

df

Degrees of freedom used by the model.

df.residual

Residual degrees of freedom.

nobs

Number of observations used.

p.value

P-value corresponding to the test statistic.

r.squared

R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

sigma

Estimated standard error of the residuals.

statistic

Test statistic.

Examples

if (requireNamespace("margins", quietly = TRUE)) {

library(margins)

## Example 1: Logit model ##

mod_log <- glm(am ~ cyl + hp + wt, data = mtcars, family = binomial)

# Get tidied "naive" model coefficients
tidy(mod_log)

# Convert to marginal effects with margins::margins()
marg_log <- margins(mod_log)

# Get tidied marginal effects
tidy(marg_log)
tidy(marg_log, conf.int = TRUE)
glance(marg_log) ## Requires running the underlying model again. Quick for this example.
if (FALSE) augment(marg_log) ## Not supported.
augment(mod_log) ## But can get the same info by running on the underlying model.

## Example 2: Threeway interaction terms ##

mod_ie <- lm(mpg ~ wt * cyl * disp, data = mtcars)

# Get tidied "naive" model coefficients
tidy(mod_ie)

# Convert to marginal effects with margins::margins()
marg_ie0 <- margins(mod_ie)
# Get tidied marginal effects 
tidy(marg_ie0)
glance(marg_ie0)

# Marginal effects evaluated at specific values of a variable (here: cyl)
marg_ie1 <- margins(mod_ie, at = list(cyl = c(4,6,8)))
tidy(marg_ie1)

# Marginal effects of one interaction variable (here: wt), modulated at 
# specific values of the two other interaction variables (here: cyl and drat)
marg_ie2 <- margins(mod_ie,
                    variables = "wt", ## Main var
                    at = list(cyl = c(4,6,8), drat = c(3, 3.5, 4))) ## Modulating vars
tidy(marg_ie2)

}
#> # A tibble: 18 × 7
#>    term  at.variable at.value estimate std.error statistic p.value
#>    <chr> <chr>          <dbl>    <dbl>     <dbl>     <dbl>   <dbl>
#>  1 wt    cyl              4       7.91      5.06     1.56    0.118
#>  2 wt    drat             3       7.91      5.06     1.56    0.118
#>  3 wt    cyl              4       7.91      5.06     1.56    0.118
#>  4 wt    drat             3.5     7.91      5.06     1.56    0.118
#>  5 wt    cyl              4       7.91      5.06     1.56    0.118
#>  6 wt    drat             4       7.91      5.06     1.56    0.118
#>  7 wt    cyl              6       2.96      2.52     1.18    0.239
#>  8 wt    drat             3       2.96      2.52     1.18    0.239
#>  9 wt    cyl              6       2.96      2.52     1.18    0.239
#> 10 wt    drat             3.5     2.96      2.52     1.18    0.239
#> 11 wt    cyl              6       2.96      2.52     1.18    0.239
#> 12 wt    drat             4       2.96      2.52     1.18    0.239
#> 13 wt    cyl              8      -1.98      2.40    -0.825   0.409
#> 14 wt    drat             3      -1.98      2.40    -0.825   0.409
#> 15 wt    cyl              8      -1.98      2.40    -0.825   0.409
#> 16 wt    drat             3.5    -1.98      2.40    -0.825   0.409
#> 17 wt    cyl              8      -1.98      2.40    -0.825   0.409
#> 18 wt    drat             4      -1.98      2.40    -0.825   0.409