Glance at a(n) orcutt object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for orcutt
glance(x, ...)

Arguments

x
An orcutt object returned from orcutt::cochrane.orcutt().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with exactly one row and columns:

adj.r.squared

Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.

dw.original

Durbin-Watson statistic of original fit

dw.transformed

Durbin-Watson statistic of transformed fit

nobs

Number of observations used.

number.interaction

Number of interactions

p.value.original

P-value of original Durbin-Watson statistic

p.value.transformed

P-value of autocorrelation after transformation

r.squared

R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

rho

Spearman's rho autocorrelation

Examples

library(orcutt)

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

#> # A tibble: 4 x 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept) 19.8         5.94      3.33   0.00244 
#> 2 wt          -5.03        1.22     -4.11   0.000310
#> 3 qsec         0.927       0.342     2.71   0.0114  
#> 4 disp        -0.000128    0.0106   -0.0121 0.990   

co <- cochrane.orcutt(reg)
co

#> Cochrane-orcutt estimation for first order autocorrelation 
#>  
#> Call:
#> lm(formula = mpg ~ wt + qsec + disp, data = mtcars)
#> 
#>  number of interaction: 7
#>  rho 0.26819
#> 
#> Durbin-Watson statistic 
#> (original):    1.49575 , p-value: 4.063e-02
#> (transformed): 2.05696 , p-value: 5.21e-01
#>  
#>  coefficients: 
#> (Intercept)          wt        qsec        disp 
#>   21.814858   -4.852590    0.797029   -0.001359 

tidy(co)

#> Warning: deal with tidy.orcutt conf.int nonsense

#> Warning: Assigning non-quosure objects to quosure lists is deprecated as of rlang 0.3.0.
#> Please coerce to a bare list beforehand with `as.list()`
#> This warning is displayed once per session.

#> # A tibble: 4 x 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept) 21.8        6.63       3.29  0.00279
#> 2 wt          -4.85       1.33      -3.65  0.00112
#> 3 qsec         0.797      0.370      2.15  0.0402 
#> 4 disp        -0.00136    0.0110    -0.123 0.903  

glance(co)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared   rho number.interact… dw.original p.value.original
#>       <dbl>         <dbl> <dbl>            <dbl>       <dbl>            <dbl>
#> 1     0.799         0.777 0.268                7        1.50           0.0406
#> # … with 3 more variables: dw.transformed <dbl>, p.value.transformed <dbl>,
#> #   nobs <int>