Glance at a(n) rma object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for rma
glance(x, ...)

Arguments

x
An rma object such as those created by metafor::rma(), metafor::rma.uni(), metafor::rma.glmm(), metafor::rma.mh(), metafor::rma.mv(), or metafor::rma.peto().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

cochran.qe

In meta-analysis, test statistic for the Cochran's Q_e test of residual heterogeneity.

cochran.qm

In meta-analysis, test statistic for the Cochran's Q_m omnibus test of coefficients.

h.squared

Value of the H-Squared statistic

i.squared

Value of the I-Squared statistic

measure

the measure used in the meta-analysis

method

Which method was used.

nobs

Number of observations used.

p.value.cochran.qe

In meta-analysis, p-value for the Cochran's Q_e test of residual heterogeneity.

p.value.cochran.qm

In meta-analysis, p-value for the Cochran's Q_m omnibus test of coefficients.

tau.squared

In meta-analysis, estimated amount of residual heterogeneity.

tau.squared.se

In meta-analysis, standard error of residual heterogeneity.

Examples

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

glance(meta_analysis)

#> # A tibble: 1 x 14
#>   i.squared h.squared tau.squared tau.squared.se cochran.qe p.value.cochran…
#>       <dbl>     <dbl>       <dbl>          <dbl>      <dbl>            <dbl>
#> 1      92.3      13.0       0.318          0.174       152.         2.00e-26
#> # … with 8 more variables: cochran.qm <dbl>, p.value.cochran.qm <dbl>,
#> #   logLik <dbl>, deviance <dbl>, AIC <dbl>, BIC <dbl>, AICc <dbl>, nobs <int>