Glance at a(n) survdiff object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

# S3 method for survdiff
glance(x, ...)

Arguments

x
An survdiff object returned from survival::survdiff().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with exactly one row and columns:

df

Degrees of freedom used by the model.

p.value

P-value corresponding to the test statistic.

statistic

Test statistic.

Examples

library(survival)

s <- survdiff(
  Surv(time, status) ~ pat.karno + strata(inst),
  data = lung
)

tidy(s)

#> # A tibble: 8 x 4
#>   pat.karno     N   obs    exp
#>   <chr>     <dbl> <dbl>  <dbl>
#> 1 30            2     1  0.692
#> 2 40            2     1  1.10 
#> 3 50            4     4  1.17 
#> 4 60           30    27 16.3  
#> 5 70           41    31 26.4  
#> 6 80           50    38 41.9  
#> 7 90           60    38 47.2  
#> 8 100          35    21 26.2  

glance(s)

#> # A tibble: 1 x 3
#>   statistic    df p.value
#>       <dbl> <dbl>   <dbl>
#> 1      21.4     7 0.00326