Glance at a(n) svyolr object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for svyolr
glance(x, ...)

Arguments

x
A svyolr object returned from survey::svyolr().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with exactly one row and columns:

df.residual

Residual degrees of freedom.

edf

The effective degrees of freedom

nobs

Number of observations used.

Examples

library(MASS)

fit <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)

tidy(fit, exponentiate = TRUE, conf.int = TRUE)

#> 
#> Re-fitting to get Hessian

#> Waiting for profiling to be done...

#> 
#> Re-fitting to get Hessian

#> # A tibble: 8 x 7
#>   term          estimate std.error statistic conf.low conf.high coef.type  
#>   <chr>            <dbl>     <dbl>     <dbl>    <dbl>     <dbl> <chr>      
#> 1 InflMedium       1.76     0.105       5.41    1.44      2.16  coefficient
#> 2 InflHigh         3.63     0.127      10.1     2.83      4.66  coefficient
#> 3 TypeApartment    0.564    0.119      -4.80    0.446     0.712 coefficient
#> 4 TypeAtrium       0.693    0.155      -2.36    0.511     0.940 coefficient
#> 5 TypeTerrace      0.336    0.151      -7.20    0.249     0.451 coefficient
#> 6 ContHigh         1.43     0.0955      3.77    1.19      1.73  coefficient
#> 7 Low|Medium       0.609    0.125      -3.97   NA        NA     scale      
#> 8 Medium|High      2.00     0.125       5.50   NA        NA     scale      

glance(fit)

#> # A tibble: 1 x 7
#>     edf logLik   AIC   BIC deviance df.residual  nobs
#>   <int>  <dbl> <dbl> <dbl>    <dbl>       <int> <int>
#> 1     8 -1740. 3495. 3539.    3479.        1673  1681