glm_tidiers {broom}	R Documentation

Tidying methods for a glm object

Description

Tidy a glm object. The tidy and augment methods are handled by lm_tidiers.

Usage

## S3 method for class 'glm'
glance(x, ...)

Arguments

x
glm object

...
extra arguments, not used

Value

tidy and augment return the same values as do tidy.lm and augment.lm.

glance returns a one-row data.frame with the columns

null.deviance
the deviance of the null model

df.null
the residual degrees of freedom for the null model

logLik
the data's log-likelihood under the model

AIC
the Akaike Information Criterion

BIC
the Bayesian Information Criterion

deviance
deviance

df.residual
residual degrees of freedom

See Also

tidy.lm and augment.lm. Also glm, which computes the values reported by the glance method.

Examples

g <- glm(am ~ mpg, mtcars, family = "binomial")
glance(g)

[Package broom version 0.4.5 Index]