Tidying methods for an htest object

Tidies hypothesis test objects, such as those from stats::cor.test(), stats::t.test(), stats::wilcox.test(), and stats::chisq.test(), into a one-row data frame.

# S3 method for htest
tidy(x, ...)

# S3 method for htest
glance(x, ...)

# S3 method for htest
augment(x, ...)

Arguments

x
An object of class "htest"

...
extra arguments (not used)

Value

Both tidy and glance return the same output, a one-row data frame with one or more of the following columns:

estimate

Estimate of the effect size

statistic

Test statistic used to compute the p-value

p.value

P-value

parameter

Parameter field in the htest, typically degrees of freedom

conf.low

Lower bound on a confidence interval

conf.high

Upper bound on a confidence interval

estimate1

Sometimes two estimates are computed, such as in a two-sample t-test

estimate2

Sometimes two estimates are computed, such as in a two-sample t-test

method

Method used to compute the statistic as a string

alternative

Alternative hypothesis as a string

Which columns are included depends on the hypothesis test used. For chi-squared tests, augment.htest will returns, for each cell of the tested table, the additional columns:
.observed

Observed count

.prop

Proportion of the total

.row.prop

Row proportion (2 dimensions table only)

.col.prop

Column proportion (2 dimensions table only)

.expected

Expected count under the null hypothesis

.residuals

Pearson residual

.stdres

Standardized residual

See stats::chisq.test() for more details on how residuals are computed.
Details

augment method is defined only for chi-squared tests, since there is no sense, for other tests, in which a hypothesis test generates one value for each observation.

Examples

tt <- t.test(rnorm(10))
tidy(tt)

#>     estimate  statistic   p.value parameter  conf.low conf.high
#> 1 -0.2178381 -0.5332314 0.6067881         9 -1.141985 0.7063086
#>              method alternative
#> 1 One Sample t-test   two.sided

glance(tt)  # same output for all htests

#>     estimate  statistic   p.value parameter  conf.low conf.high
#> 1 -0.2178381 -0.5332314 0.6067881         9 -1.141985 0.7063086
#>              method alternative
#> 1 One Sample t-test   two.sided

tt <- t.test(mpg ~ am, data = mtcars)
tidy(tt)

#>    estimate estimate1 estimate2 statistic     p.value parameter  conf.low
#> 1 -7.244939  17.14737  24.39231 -3.767123 0.001373638  18.33225 -11.28019
#>   conf.high                  method alternative
#> 1 -3.209684 Welch Two Sample t-test   two.sided

wt <- wilcox.test(mpg ~ am, data = mtcars, conf.int = TRUE, exact = FALSE)
tidy(wt)

#>    estimate statistic     p.value  conf.low conf.high
#> 1 -6.799963        42 0.001871391 -11.69994 -2.900042
#>                                              method alternative
#> 1 Wilcoxon rank sum test with continuity correction   two.sided

ct <- cor.test(mtcars$wt, mtcars$mpg)
tidy(ct)

#>     estimate statistic      p.value parameter   conf.low  conf.high
#> 1 -0.8676594 -9.559044 1.293959e-10        30 -0.9338264 -0.7440872
#>                                 method alternative
#> 1 Pearson's product-moment correlation   two.sided

chit <- chisq.test(xtabs(Freq ~ Sex + Class, data = as.data.frame(Titanic)))
tidy(chit)

#>   statistic      p.value parameter                     method
#> 1  349.9145 1.556637e-75         3 Pearson's Chi-squared test

augment(chit)

#>      Sex Class .observed      .prop  .row.prop .col.prop .expected .residuals
#> 1   Male   1st       180 0.08178101 0.10398614 0.5538462 255.59973  -4.728681
#> 2 Female   1st       145 0.06587915 0.30851064 0.4461538  69.40027   9.074854
#> 3   Male   2nd       179 0.08132667 0.10340843 0.6280702 224.14130  -3.015179
#> 4 Female   2nd       106 0.04815993 0.22553191 0.3719298  60.85870   5.786457
#> 5   Male   3rd       510 0.23171286 0.29462738 0.7223796 555.24125  -1.919967
#> 6 Female   3rd       196 0.08905043 0.41702128 0.2776204 150.75875   3.684626
#> 7   Male  Crew       862 0.39164016 0.49797805 0.9740113 696.01772   6.291462
#> 8 Female  Crew        23 0.01044980 0.04893617 0.0259887 188.98228 -12.074000
#>      .stdres
#> 1 -11.083954
#> 2  11.083954
#> 3  -6.993369
#> 4   6.993369
#> 5  -5.041320
#> 6   5.041320
#> 7  17.607380
#> 8 -17.607380