inflate {broom}	R Documentation

Expand a dataset to include all factorial combinations of one or more variables

Description

This function is deprecated: use tidyr::crossing instead

Usage

inflate(df, ..., stringsAsFactors = FALSE)

Arguments

df
a tbl

...
arguments

stringsAsFactors
logical specifying if character vectors are converted to factors.

Value

A tbl

[Package broom version 0.4.5 Index]