insert a row of NAs into a data frame wherever another data frame has NAs

insert a row of NAs into a data frame wherever another data frame has NAs

insert_NAs(x, original)

Arguments

x
data frame that has one row for each non-NA row in original

original
data frame with NAs