Tidying methods for lavaan models

These methods tidy the coefficients of lavaan CFA and SEM models.

# S3 method for lavaan
tidy(x, conf.int = TRUE, conf.level = 0.95, ...)

# S3 method for lavaan
glance(x, ...)

Arguments

x
An object of class lavaan, such as those from lavaan::cfa(), or lavaan::sem()

conf.int
Logical indicating if a confidence intervals are desired.

conf.level
Confidence level to use. Default is 0.95.

...
For tidy, additional arguments passed to lavaan::parameterEstimates(). Ignored for glance.

Value

The tidy method returns a tibble with one row for each estimated parameter and columns:

term

The result of paste(lhs, op, rhs)

op

The operator in the model syntax (e.g. ~~ for covariances, or ~ for regression parameters)

group

The group (if specified) in the lavaan model

estimate

The parameter estimate (may be standardized)

std.error

statistic

The z value returned by lavaan::parameterEstimates()

p.value

conf.low

conf.high

std.lv

Standardized estimates based on the variances of the (continuous) latent variables only

std.all

Standardized estimates based on both the variances of both (continuous) observed and latent variables.

std.nox

Standardized estimates based on both the variances of both (continuous) observed and latent variables, but not the variances of exogenous covariates.

The glance method returns a single row dataset giving model fit and related information, including:
chisq

Model chi squared

npar

Number of parameters in the model

rmsea

Root mean square error of approximation

rmsea.conf.high

95 percent upper bound on RMSEA

srmr

Standardised root mean residual

agfi

Adjusted goodness of fit

cfi

Comparative fit index

tli

Tucker Lewis index

aic

Akaike information criterion

bic

Bayesian information criterion

ngroups

Number of groups in model

nobs

Number of observations included

norig

Number of observation in the original dataset

nexcluded

Number of excluded observations

converged

Logical - Did the model converge

estimator

Estimator used

missing_method

Method for eliminating missing data

For further recommendations on reporting SEM and CFA models see Schreiber, J. B. (2017). Update to core reporting practices in structural equation modeling. Research in Social and Administrative Pharmacy, 13(3), 634-643. https://doi.org/10.1016/j.sapharm.2016.06.006
Examples

if (require("lavaan")) {
 cfa.fit <- cfa('F =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9',
                data = HolzingerSwineford1939, group = "school")
 tidy(cfa.fit)
}

#> Loading required package: lavaan

#> This is lavaan 0.6-1

#> lavaan is BETA software! Please report any bugs.

#> 
#> Attaching package: 'lavaan'

#> The following object is masked from 'package:ks':
#> 
#>     vech

#> The following object is masked from 'package:psych':
#> 
#>     cor2cov

#> # A tibble: 58 x 13
#>    term     op    block group estimate std.error statistic  p.value conf.low
#>    <chr>    <chr> <int> <int>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>
#>  1 F =~ x1  =~        1     1    1         0         NA    NA         1     
#>  2 F =~ x2  =~        1     1    0.333     0.190      1.76  7.89e-2  -0.0385
#>  3 F =~ x3  =~        1     1    0.400     0.182      2.20  2.80e-2   0.0432
#>  4 F =~ x4  =~        1     1    1.66      0.280      5.92  3.28e-9   1.11  
#>  5 F =~ x5  =~        1     1    1.92      0.323      5.95  2.60e-9   1.29  
#>  6 F =~ x6  =~        1     1    1.48      0.247      5.98  2.23e-9   0.992 
#>  7 F =~ x7  =~        1     1    0.453     0.173      2.61  8.96e-3   0.113 
#>  8 F =~ x8  =~        1     1    0.376     0.155      2.43  1.51e-2   0.0726
#>  9 F =~ x9  =~        1     1    0.422     0.159      2.66  7.80e-3   0.111 
#> 10 x1 ~~ x1 ~~        1     1    1.07      0.127      8.47  0.        0.825 
#> # ... with 48 more rows, and 4 more variables: conf.high <dbl>, std.lv <dbl>,
#> #   std.all <dbl>, std.nox <dbl>

if (require("lavaan", quietly = TRUE)) {

 cfa.fit <- cfa(
   'F =~ x1 + x2 + x3 + x4 + x5',
   data = HolzingerSwineford1939, group = "school"
 )
 glance(cfa.fit)

}

#> # A tibble: 1 x 17
#>    agfi   aic   bic   cfi chisq  npar rmsea rmsea.conf.high  srmr   tli
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>           <dbl> <dbl> <dbl>
#> 1 0.971 4473. 4584. 0.766  99.3    30 0.244           0.288 0.115 0.533
#> # ... with 7 more variables: converged <lgl>, estimator <chr>, ngroups <int>,
#> #   missing_method <chr>, nobs <int>, norig <int>, nexcluded <int>