leveneTest_tidiers {broom}	R Documentation

Tidy/glance a(n) leveneTest object

Description

For models that have only a single component, the tidy() and glance() methods are identical. Please see the documentation for both of those methods.

Usage

## S3 method for class 'leveneTest'
tidy(x, ...)

Arguments

x
An object of class anova created by a call to car::leveneTest().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

df
Degrees of freedom used by this term in the model.

p.value
The two-sided p-value associated with the observed statistic.

statistic
The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

df.residual
Residual degrees of freedom.

See Also

tidy(), glance(), car::leveneTest()

Other car tidiers: durbinWatsonTest_tidiers

Examples

if (requireNamespace("car", quietly = TRUE)) {

library(car)
data(Moore)
lt <- with(Moore, leveneTest(conformity, fcategory))
tidy(lt)
glance(lt) # same output for all leveneTest

}

[Package broom version 0.7.12 Index]