list_tidiers {broom}	R Documentation

Tidiers for return values from functions that aren't S3 objects

Description

This method handles the return values of functions that return lists rather than S3 objects, such as optim, svd, or interp, and therefore cannot be handled by S3 dispatch.

Usage

## S3 method for class 'list'
tidy(x, ...)

## S3 method for class 'list'
glance(x, ...)

Arguments

x
list object

...
extra arguments, passed to the tidying function

Details

Those tiders themselves are implemented as functions of the form tidy_<function> or glance_<function> that are not exported.

See Also

optim_tidiers, xyz_tidiers, svd_tidiers, orcutt_tidiers

[Package broom version 0.4.5 Index]