Augmenting methods for loess models

This method augments the original data with information on the fitted values and residuals, and optionally the standard errors.

# S3 method for loess
augment(x, data = stats::model.frame(x), newdata, ...)

Arguments

x
A "loess" object

data
Original data, defaults to the extracting it from the model

newdata
If provided, performs predictions on the new data

...
extra arguments

Value

When newdata is not supplied augment.loess returns one row for each observation with three columns added to the original data:

.fitted

Fitted values of model

.se.fit

Standard errors of the fitted values

.resid

Residuals of the fitted values

When newdata is supplied augment.loess returns one row for each observation with one additional column:
.fitted

Fitted values of model

.se.fit

Standard errors of the fitted values

Details

When the modeling was performed with na.action = "na.omit" (as is the typical default), rows with NA in the initial data are omitted entirely from the augmented data frame. When the modeling was performed with na.action = "na.exclude", one should provide the original data as a second argument, at which point the augmented data will contain those rows (typically with NAs in place of the new columns). If the original data is not provided to augment and na.action = "na.exclude", a warning is raised and the incomplete rows are dropped.

See also

Examples

lo <- loess(mpg ~ wt, mtcars)
augment(lo)

#>              .rownames  mpg    wt  .fitted   .se.fit      .resid
#> 1            Mazda RX4 21.0 2.620 22.23500 1.0051037 -1.23500287
#> 2        Mazda RX4 Wag 21.0 2.875 21.07864 1.0699785 -0.07864039
#> 3           Datsun 710 22.8 2.320 24.64036 1.0279114 -1.84035946
#> 4       Hornet 4 Drive 21.4 3.215 19.48575 0.8613926  1.91424768
#> 5    Hornet Sportabout 18.7 3.440 17.92139 0.7545629  0.77860889
#> 6              Valiant 18.1 3.460 17.81047 0.7469703  0.28953295
#> 7           Duster 360 14.3 3.570 17.19469 0.7209080 -2.89468859
#> 8            Merc 240D 24.4 3.190 19.62190 0.8659118  4.77809534
#> 9             Merc 230 22.8 3.150 19.82365 0.8675160  2.97634956
#> 10            Merc 280 19.2 3.440 17.92139 0.7545629  1.27860889
#> 11           Merc 280C 17.8 3.440 17.92139 0.7545629 -0.12139111
#> 12          Merc 450SE 16.4 4.070 15.06835 1.4945931  1.33165499
#> 13          Merc 450SL 17.3 3.730 16.36124 0.8082683  0.93876153
#> 14         Merc 450SLC 15.2 3.780 16.12469 0.8819110 -0.92469036
#> 15  Cadillac Fleetwood 10.4 5.250 12.00567 1.4311979 -1.60567233
#> 16 Lincoln Continental 10.4 5.424 11.79784 1.7288616 -1.39784402
#> 17   Chrysler Imperial 14.7 5.345 11.88216 1.5648626  2.81784400
#> 18            Fiat 128 32.4 2.200 25.87891 0.9990827  6.52108816
#> 19         Honda Civic 30.4 1.615 31.25696 1.5340390 -0.85695536
#> 20      Toyota Corolla 33.9 1.835 29.39104 1.0520751  4.50896465
#> 21       Toyota Corona 21.5 2.465 23.26502 1.0547677 -1.76501983
#> 22    Dodge Challenger 15.5 3.520 17.46743 0.7287936 -1.96742941
#> 23         AMC Javelin 15.2 3.435 17.94910 0.7565093 -2.74910498
#> 24          Camaro Z28 13.3 3.840 15.87862 1.0004840 -2.57862456
#> 25    Pontiac Firebird 19.2 3.845 15.86046 1.0118611  3.33954248
#> 26           Fiat X1-9 27.3 1.935 28.51253 0.9854662 -1.21252674
#> 27       Porsche 914-2 26.0 2.140 26.50478 0.9870267 -0.50477911
#> 28        Lotus Europa 30.4 1.513 32.08897 1.9183213 -1.68897234
#> 29      Ford Pantera L 15.8 3.170 19.72501 0.8670235 -3.92500705
#> 30        Ferrari Dino 19.7 2.770 21.57074 1.0341080 -1.87073637
#> 31       Maserati Bora 15.0 3.570 17.19469 0.7209080 -2.19468859
#> 32          Volvo 142E 21.4 2.780 21.52790 1.0395396 -0.12790190

# with all columns of original data
augment(lo, mtcars)

#>              .rownames  mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1            Mazda RX4 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2        Mazda RX4 Wag 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3           Datsun 710 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4       Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5    Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6              Valiant 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7           Duster 360 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8            Merc 240D 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9             Merc 230 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10            Merc 280 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 11           Merc 280C 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 12          Merc 450SE 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 13          Merc 450SL 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 14         Merc 450SLC 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15  Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 16 Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 17   Chrysler Imperial 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 18            Fiat 128 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 19         Honda Civic 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 20      Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 21       Toyota Corona 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22    Dodge Challenger 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 23         AMC Javelin 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24          Camaro Z28 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 25    Pontiac Firebird 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 26           Fiat X1-9 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 27       Porsche 914-2 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 28        Lotus Europa 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 29      Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 30        Ferrari Dino 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 31       Maserati Bora 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 32          Volvo 142E 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#>     .fitted   .se.fit      .resid
#> 1  22.23500 1.0051037 -1.23500287
#> 2  21.07864 1.0699785 -0.07864039
#> 3  24.64036 1.0279114 -1.84035946
#> 4  19.48575 0.8613926  1.91424768
#> 5  17.92139 0.7545629  0.77860889
#> 6  17.81047 0.7469703  0.28953295
#> 7  17.19469 0.7209080 -2.89468859
#> 8  19.62190 0.8659118  4.77809534
#> 9  19.82365 0.8675160  2.97634956
#> 10 17.92139 0.7545629  1.27860889
#> 11 17.92139 0.7545629 -0.12139111
#> 12 15.06835 1.4945931  1.33165499
#> 13 16.36124 0.8082683  0.93876153
#> 14 16.12469 0.8819110 -0.92469036
#> 15 12.00567 1.4311979 -1.60567233
#> 16 11.79784 1.7288616 -1.39784402
#> 17 11.88216 1.5648626  2.81784400
#> 18 25.87891 0.9990827  6.52108816
#> 19 31.25696 1.5340390 -0.85695536
#> 20 29.39104 1.0520751  4.50896465
#> 21 23.26502 1.0547677 -1.76501983
#> 22 17.46743 0.7287936 -1.96742941
#> 23 17.94910 0.7565093 -2.74910498
#> 24 15.87862 1.0004840 -2.57862456
#> 25 15.86046 1.0118611  3.33954248
#> 26 28.51253 0.9854662 -1.21252674
#> 27 26.50478 0.9870267 -0.50477911
#> 28 32.08897 1.9183213 -1.68897234
#> 29 19.72501 0.8670235 -3.92500705
#> 30 21.57074 1.0341080 -1.87073637
#> 31 17.19469 0.7209080 -2.19468859
#> 32 21.52790 1.0395396 -0.12790190

# with a new dataset
augment(lo, newdata = head(mtcars))

#>           .rownames  mpg cyl disp  hp drat    wt  qsec vs am gear carb  .fitted
#> 1         Mazda RX4 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4 22.23500
#> 2     Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4 21.07864
#> 3        Datsun 710 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1 24.64036
#> 4    Hornet 4 Drive 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1 19.48575
#> 5 Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2 17.92139
#> 6           Valiant 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1 17.81047
#>     .se.fit
#> 1 1.0051037
#> 2 1.0699785
#> 3 1.0279114
#> 4 0.8613926
#> 5 0.7545629
#> 6 0.7469703