Tidying methods for Mclust objects

These methods summarize the results of Mclust clustering into three tidy forms. tidy describes the size, mixing probability, mean and variabilty of each class, augment adds the class assignments and their probabilities to the original data, and glance summarizes the model parameters of the clustering.

# S3 method for Mclust
tidy(x, ...)

# S3 method for Mclust
augment(x, data, ...)

# S3 method for Mclust
glance(x, ...)

Arguments

x
Mclust object

...
extra arguments, not used

data
Original data (required for augment)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen. tidy returns one row per component, with

component

A factor describing the cluster from 1:k (or 0:k in presence of a noise term in x)

size

The size of each component

proportion

The mixing proportion of each component

variance

In case of one-dimensional and spherical models, the variance for each component, omitted otherwise. NA for noise component

mean

The mean for each component. In case of two- or more dimensional models, a column with the mean is added for each dimension. NA for noise component

augment returns the original data with two extra columns:
.class

The class assigned by the Mclust algorithm

.uncertainty

The uncertainty associated with the classification

glance returns a one-row data.frame with the columns
model

A character string denoting the model at which the optimal BIC occurs

n

The number of observations in the data

G

The optimal number of mixture components

BIC

The optimal BIC value

logLik

The log-likelihood corresponding to the optimal BIC

df

The number of estimated parameters

hypvol

The hypervolume parameter for the noise component if required, otherwise set to NA

See also

Mclust

Examples

library(dplyr)
library(ggplot2)
library(mclust)

set.seed(2016)
centers <- data.frame(cluster=factor(1:3), size=c(100, 150, 50),
                      x1=c(5, 0, -3), x2=c(-1, 1, -2))
points <- centers %>% group_by(cluster) %>%
 do(data.frame(x1=rnorm(.$size[1], .$x1[1]),
               x2=rnorm(.$size[1], .$x2[1]))) %>%
 ungroup()

m = Mclust(points %>% dplyr::select(x1, x2))

#> fitting ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

tidy(m)

#>   component size proportion variance     mean.x1    mean.x2
#> 1         1  100  0.3331737 0.985107  4.88752710 -0.8407367
#> 2         2  151  0.5027053 0.985107  0.06634404  1.0124152
#> 3         3   49  0.1641210 0.985107 -3.07061853 -2.0233387

head(augment(m, points))

#>   cluster       x1          x2 .class .uncertainty
#> 1       1 4.085258 -0.37102709      1 2.441540e-04
#> 2       1 6.001248 -2.31481105      1 5.329488e-10
#> 3       1 4.943577 -2.02473082      1 1.628193e-07
#> 4       1 5.296645  0.53447981      1 3.576154e-06
#> 5       1 2.208529 -0.67466136      2 4.267338e-01
#> 6       1 4.717260 -0.01225955      1 2.176909e-05

glance(m)

#>   model   n G       BIC   logLik df hypvol
#> 1   EII 300 3 -2323.814 -1136.24  9     NA