Tidy a(n) rma object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for rma
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, include_studies = TRUE, measure = "GEN", ...)

Arguments

x
An rma object such as those created by metafor::rma(), metafor::rma.uni(), metafor::rma.glmm(), metafor::rma.mh(), metafor::rma.mv(), or metafor::rma.peto().

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

include_studies
Logical. Should individual studies be included in the output? Defaults to TRUE.

measure
Measure type. See metafor::escalc()

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

conf.high

Upper bound on the confidence interval for the estimate.

conf.low

Lower bound on the confidence interval for the estimate.

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

study

The name of the individual study

type

The estimate type (summary vs individual study)

Examples

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

tidy(meta_analysis)

#> # A tibble: 14 x 6
#>    study   type    estimate std.error statistic    p.value
#>    <chr>   <chr>      <dbl>     <dbl>     <dbl>      <dbl>
#>  1 1       study    -0.889     0.571    -1.56   NA        
#>  2 2       study    -1.59      0.441    -3.59   NA        
#>  3 3       study    -1.35      0.644    -2.09   NA        
#>  4 4       study    -1.44      0.141   -10.2    NA        
#>  5 5       study    -0.218     0.226    -0.961  NA        
#>  6 6       study    -0.786     0.0831   -9.46   NA        
#>  7 7       study    -1.62      0.472    -3.43   NA        
#>  8 8       study     0.0120    0.0629    0.190  NA        
#>  9 9       study    -0.469     0.238    -1.98   NA        
#> 10 10      study    -1.37      0.270    -5.07   NA        
#> 11 11      study    -0.339     0.111    -3.05   NA        
#> 12 12      study     0.446     0.730     0.611  NA        
#> 13 13      study    -0.0173    0.267    -0.0648 NA        
#> 14 overall summary  -0.715     0.181    -3.95    0.0000774