tidy.gam {broom}	R Documentation

Tidy a(n) gam object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'gam'
tidy(x, parametric = FALSE, ...)

Arguments

x
A gam object returned from a call to mgcv::gam().

parametric
Logical indicating if parametric or smooth terms should be tidied. Defaults to FALSE, meaning that smooth terms are tidied by default.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

To tidy Gam objects created by calls to gam::gam(), see tidy.Gam().

See Also

tidy(), mgcv::gam(), tidy.Gam()

Other mgcv tidiers: glance.gam

Examples

g <- mgcv::gam(mpg ~ s(hp) + am + qsec, data = mtcars)
  
tidy(g)
tidy(g, parametric = TRUE)
glance(g)

[Package broom version 0.5.0 Index]