muhaz_tidiers {broom}	R Documentation

Tidying methods for kernel based hazard rate estimates

Description

These methods tidy the output of muhaz objects as returned by the muhaz function, which provides kernel based non-parametric hazard rate estimators.

Usage

## S3 method for class 'muhaz'
tidy(x, ...)

## S3 method for class 'muhaz'
glance(x, ...)

Arguments

x
muhaz object

...
extra arguments (not used)

Details

The "augment" method is not useful and therefore not available for muhaz objects.

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.muhaz returns a tibble containing two columns: time at which the hazard rate was estimated and estimate.

glance.muhaz returns a one-row data.frame with the columns

nobs
Number of observations used for estimation

min.time
The minimum observed event or censoring time

max.time
The maximum observed event or censoring time

min.harzard
Minimal estimated hazard

max.hazard
Maximal estimated hazard

Examples

if (require("muhaz", quietly = TRUE)) {
  data(ovarian, package="survival")
  x <- muhaz(ovarian$futime, ovarian$fustat)
  tidy(x)
  glance(x)
}

[Package broom version 0.4.5 Index]