Tidiers for NULL inputs

tidy(NULL), glance(NULL) and augment(NULL) all return an empty tibble::tibble. This empty tibble can be treated a tibble with zero rows, making it convenient to combine with other tibbles using functions like purrr::map_df() on lists of potentially NULL objects.

# S3 method for NULL
tidy(x, ...)

# S3 method for NULL
glance(x, ...)

# S3 method for NULL
augment(x, ...)

Arguments

x
The value NULL.

...
Additional arguments (not used).

Value

An empty tibble::tibble.

See also