orcutt_tidiers {broom}	R Documentation

Tidiers for Cochrane Orcutt object

Description

Tidies a Cochrane Orcutt object, which estimates autocorrelation and beta coefficients in a linear fit.

Usage

## S3 method for class 'orcutt'
tidy(x, ...)

## S3 method for class 'orcutt'
glance(x, ...)

Arguments

x
An "orcutt" object returned by cochrane.orcutt

...
Extra arguments passed on to tidy.lm

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy returns the same information as tidy.lm, though without confidence interval options.

glance

glance returns a one-row data frame with the following columns:

r.squared
R-squared

adj.r.squared
Adjusted R-squared

rho
Spearman's rho autocorrelation

number.interaction
Number of interactions

dw.original
Durbin-Watson statistic of original fit

p.value.original
P-value of original Durbin-Watson statistic

dw.transformed
Durbin-Watson statistic of transformed fit

p.value.transformed
P-value of autocorrelation after transformation

Examples

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

if (require("orcutt", quietly = TRUE)) {
  co <- cochrane.orcutt(reg)
  co
  
  tidy(co)
  glance(co)
}

[Package broom version 0.4.5 Index]