Tidying methods for a randomForest model

These methods tidy the variable importance of a random forest model summary, augment the original data with information on the fitted values/classifications and error, and construct a one-row glance of the model's statistics.

# S3 method for randomForest
tidy(x, ...)

# S3 method for randomForest
augment(x, data = NULL, ...)

# S3 method for randomForest
glance(x, ...)

Arguments

x
randomForest object

...
Additional arguments (ignored)

data
Model data for use by augment.randomForest.

Value

All tidying methods return a data.frame without rownames. The structure depends on the method chosen. tidy.randomForest returns one row for each model term, with the following columns:

term

The term in the randomForest model

class_*

One column for each model term; the relative importance of each term per class. Only present if the model was created with importance = TRUE

MeanDecreaseAccuracy

A measure of variable importance. See randomForest for more information. Only present if the model was created with importance = TRUE

MeanDecreaseGini

A measure of variable importance. See randomForest for more information.

sd_*

Sandard deviations for the preceding statistics. Only present if the model was created with importance = TRUE

augment.randomForest returns the original data with additional columns:
.oob_times

The number of trees for which the given case was "out of bag". See randomForest for more details.

.fitted

The fitted value or class.

.li_*

The casewise variable importance for each term. Only present if the model was created with importance = TRUE

In addition, augment returns additional columns for classification and usupervised trees:
.votes_*

For each case, the voting results, with one column per class.

glance.randomForest returns a one-row data.frame with the following columns: For regression trees, the following additional columns are present:
mse

The average mean squared error across all trees.

rsq

The average pesudo-R-squared across all trees. See randomForest for more information.

For classification trees, the following columns are present for each class
*_precision

*_recall

*_accuracy

*_f_measure