robust_tidiers {broom}	R Documentation

Tidiers for lmRob and glmRob objects

Description

Tidying robust regression objects from the robust package. The tidy and augment methods simply pass it on to the linear model tidiers.

Usage

## S3 method for class 'lmRob'
tidy(x, ...)

## S3 method for class 'lmRob'
augment(x, ...)

## S3 method for class 'lmRob'
glance(x, ...)

## S3 method for class 'glmRob'
tidy(x, ...)

## S3 method for class 'glmRob'
augment(x, ...)

## S3 method for class 'glmRob'
glance(x, ...)

Arguments

x
An lmRob or glmRob object with a robust regression

...
Extra arguments, not used

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy and augment return the same results as lm_tidiers.

On an lmRob glance returns a one-row data frame with the following columns:

r.squared
R-squared

deviance
Robust deviance

sigma
Residual scale estimate

df.residual
Number of residual degrees of freedom

On an glmRob glance returns a one-row data frame with the following columns:

deviance
Robust deviance

null.deviance
Deviance under the null model

df.residual
Number of residual degrees of freedom

See Also

lm_tidiers, lmRob, glmRob

Examples

if (require("robust", quietly = TRUE)) {
  m <- lmRob(mpg ~ wt, data = mtcars)
  
  tidy(m)
  augment(m)
  glance(m)
  
  gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
  glance(gm)
}

[Package broom version 0.4.5 Index]