Tidying methods for rowwise_dfs from dplyr, for tidying each row and recombining the results

These tidy, augment and glance methods are for performing tidying on each row of a rowwise data frame created by dplyr's group_by and do operations. They first group a rowwise data frame based on all columns that are not lists, then perform the tidying operation on the specified column. This greatly shortens a common idiom of extracting tidy/augment/glance outputs after a do statement.

# S3 method for rowwise_df
tidy(x, object, ...)

# S3 method for rowwise_df
tidy_(x, object, ...)

# S3 method for rowwise_df
augment(x, object, ...)

# S3 method for rowwise_df
augment_(x, object, ...)

# S3 method for rowwise_df
glance(x, object, ...)

# S3 method for rowwise_df
glance_(x, object, ...)

# S3 method for tbl_df
tidy(x, ...)

# S3 method for tbl_df
augment(x, ...)

# S3 method for tbl_df
glance(x, ...)

Arguments

x
a rowwise_df

object
the column name of the column containing the models to be tidied. For tidy, augment, and glance it should be the bare name; for _ methods it should be quoted.

...
additional arguments to pass on to the respective tidying method

Value

A "grouped_df", where the non-list columns of the original are used as grouping columns alongside the tidied outputs.

Details

Note that this functionality is not currently implemented for data.tables, since the result of the do operation is difficult to distinguish from a regular data.table.

Examples

library(dplyr)
regressions <- mtcars %>%
    group_by(cyl) %>%
    do(mod = lm(mpg ~ wt, .))

regressions

#> Source: local data frame [3 x 2]
#> Groups: <by row>
#> 
#> # A tibble: 3 x 2
#>     cyl      mod
#> * <dbl>   <list>
#> 1     4 <S3: lm>
#> 2     6 <S3: lm>
#> 3     8 <S3: lm>

regressions %>% tidy(mod)

#> # A tibble: 6 x 6
#> # Groups:   cyl [3]
#>     cyl        term  estimate std.error statistic      p.value
#>   <dbl>       <chr>     <dbl>     <dbl>     <dbl>        <dbl>
#> 1     4 (Intercept) 39.571196 4.3465820  9.103980 7.771511e-06
#> 2     4          wt -5.647025 1.8501185 -3.052251 1.374278e-02
#> 3     6 (Intercept) 28.408845 4.1843688  6.789278 1.054844e-03
#> 4     6          wt -2.780106 1.3349173 -2.082605 9.175766e-02
#> 5     8 (Intercept) 23.868029 3.0054619  7.941551 4.052705e-06
#> 6     8          wt -2.192438 0.7392393 -2.965803 1.179281e-02

regressions %>% augment(mod)

#> # A tibble: 32 x 10
#> # Groups:   cyl [3]
#>      cyl   mpg    wt  .fitted  .se.fit      .resid       .hat   .sigma
#>    <dbl> <dbl> <dbl>    <dbl>    <dbl>       <dbl>      <dbl>    <dbl>
#>  1     4  22.8 2.320 26.47010 1.006720 -3.67009741 0.09127118 3.261796
#>  2     4  24.4 3.190 21.55719 1.951521  2.84281457 0.34297508 3.309771
#>  3     4  22.8 3.150 21.78307 1.888462  1.01693356 0.32116829 3.507377
#>  4     4  32.4 2.200 27.14774 1.017163  5.25225956 0.09317454 2.947803
#>  5     4  30.4 1.615 30.45125 1.596671 -0.05125022 0.22958701 3.534359
#>  6     4  33.9 1.835 29.20890 1.305700  4.69109534 0.15353342 3.040129
#>  7     4  21.5 2.465 25.65128 1.058052 -4.15127874 0.10081613 3.177493
#>  8     4  27.3 1.935 28.64420 1.196043 -1.34420213 0.12882788 3.497551
#>  9     4  26.0 2.140 27.48656 1.040267 -1.48656195 0.09745541 3.490854
#> 10     4  30.4 1.513 31.02725 1.747377 -0.62724679 0.27497267 3.524811
#> # ... with 22 more rows, and 2 more variables: .cooksd <dbl>, .std.resid <dbl>

regressions %>% glance(mod)

#> # A tibble: 3 x 12
#> # Groups:   cyl [3]
#>     cyl r.squared adj.r.squared    sigma statistic    p.value    df    logLik
#>   <dbl>     <dbl>         <dbl>    <dbl>     <dbl>      <dbl> <int>     <dbl>
#> 1     4 0.5086326     0.4540362 3.332283  9.316233 0.01374278     2 -27.74487
#> 2     6 0.4645102     0.3574122 1.165202  4.337245 0.09175766     2  -9.82518
#> 3     8 0.4229655     0.3748793 2.024091  8.795985 0.01179281     2 -28.65778
#> # ... with 4 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>

# we can provide additional arguments to the tidying function
regressions %>% tidy(mod, conf.int = TRUE)

#> # A tibble: 6 x 8
#> # Groups:   cyl [3]
#>     cyl        term  estimate std.error statistic      p.value  conf.low
#>   <dbl>       <chr>     <dbl>     <dbl>     <dbl>        <dbl>     <dbl>
#> 1     4 (Intercept) 39.571196 4.3465820  9.103980 7.771511e-06 29.738544
#> 2     4          wt -5.647025 1.8501185 -3.052251 1.374278e-02 -9.832284
#> 3     6 (Intercept) 28.408845 4.1843688  6.789278 1.054844e-03 17.652582
#> 4     6          wt -2.780106 1.3349173 -2.082605 9.175766e-02 -6.211620
#> 5     8 (Intercept) 23.868029 3.0054619  7.941551 4.052705e-06 17.319690
#> 6     8          wt -2.192438 0.7392393 -2.965803 1.179281e-02 -3.803102
#> # ... with 1 more variables: conf.high <dbl>

# we can also include the original dataset as a "data" argument
# to augment:
regressions <- mtcars %>%
    group_by(cyl) %>%
    do(mod = lm(mpg ~ wt, .), original = (.))

# this allows all the original columns to be included:
regressions %>% augment(mod)  # doesn't include all original

#> # A tibble: 32 x 10
#> # Groups:   cyl [3]
#>      cyl   mpg    wt  .fitted  .se.fit      .resid       .hat   .sigma
#>    <dbl> <dbl> <dbl>    <dbl>    <dbl>       <dbl>      <dbl>    <dbl>
#>  1     4  22.8 2.320 26.47010 1.006720 -3.67009741 0.09127118 3.261796
#>  2     4  24.4 3.190 21.55719 1.951521  2.84281457 0.34297508 3.309771
#>  3     4  22.8 3.150 21.78307 1.888462  1.01693356 0.32116829 3.507377
#>  4     4  32.4 2.200 27.14774 1.017163  5.25225956 0.09317454 2.947803
#>  5     4  30.4 1.615 30.45125 1.596671 -0.05125022 0.22958701 3.534359
#>  6     4  33.9 1.835 29.20890 1.305700  4.69109534 0.15353342 3.040129
#>  7     4  21.5 2.465 25.65128 1.058052 -4.15127874 0.10081613 3.177493
#>  8     4  27.3 1.935 28.64420 1.196043 -1.34420213 0.12882788 3.497551
#>  9     4  26.0 2.140 27.48656 1.040267 -1.48656195 0.09745541 3.490854
#> 10     4  30.4 1.513 31.02725 1.747377 -0.62724679 0.27497267 3.524811
#> # ... with 22 more rows, and 2 more variables: .cooksd <dbl>, .std.resid <dbl>

regressions %>% augment(mod, data = original)  # includes all original

#> # A tibble: 32 x 18
#> # Groups:   cyl [3]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb  .fitted
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>
#>  1  22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1 26.47010
#>  2  24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2 21.55719
#>  3  22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2 21.78307
#>  4  32.4     4  78.7    66  4.08 2.200 19.47     1     1     4     1 27.14774
#>  5  30.4     4  75.7    52  4.93 1.615 18.52     1     1     4     2 30.45125
#>  6  33.9     4  71.1    65  4.22 1.835 19.90     1     1     4     1 29.20890
#>  7  21.5     4 120.1    97  3.70 2.465 20.01     1     0     3     1 25.65128
#>  8  27.3     4  79.0    66  4.08 1.935 18.90     1     1     4     1 28.64420
#>  9  26.0     4 120.3    91  4.43 2.140 16.70     0     1     5     2 27.48656
#> 10  30.4     4  95.1   113  3.77 1.513 16.90     1     1     5     2 31.02725
#> # ... with 22 more rows, and 6 more variables: .se.fit <dbl>, .resid <dbl>,
#> #   .hat <dbl>, .sigma <dbl>, .cooksd <dbl>, .std.resid <dbl>