sparse_tidiers {broom}	R Documentation

Tidy a sparseMatrix object from the Matrix package

Description

sparseMatrix tidiers are deprecated and will be removed from an upcoming version of broom.

Usage

## S3 method for class 'dgTMatrix'
tidy(x, ...)

## S3 method for class 'dgCMatrix'
tidy(x, ...)

## S3 method for class 'sparseMatrix'
tidy(x, ...)

Arguments

x
A Matrix object

...
Extra arguments, not used

Details

Tidy a sparseMatrix object from the Matrix package into a three-column data frame, row, column, and value (with zeros missing). If there are row names or column names, use those, otherwise use indices

[Package broom version 0.5.0 Index]