Tidying methods for a speedlm model

These methods tidy the coefficients of a "speedlm" object into a summary, augment the original data with information on the fitted values and residuals, and construct a one-row glance of the model's statistics.

# S3 method for speedlm
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)

# S3 method for speedlm
glance(x, ...)

# S3 method for speedlm
augment(x, data = stats::model.frame(x), newdata = data,
  ...)

Arguments

x
speedlm object

conf.int
whether to include a confidence interval

conf.level
confidence level of the interval, used only if conf.int=TRUE

exponentiate
whether to exponentiate the coefficient estimates and confidence intervals (typical for logistic regression)

quick
whether to compute a smaller and faster version, containing only the term and estimate columns.

...
extra arguments (not used)

data
data frame to augment

newdata
new data to use for predictions, optional

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.speedlm returns the tidied output of the lm with one row for each term in the formula. The columns match those in lm_tidiers.

glance.speedlm returns a one-row data.frame with the columns

r.squared

The percent of variance explained by the model

adj.r.squared

r.squared adjusted based on the degrees of freedom

statistic

F-statistic

p.value

p-value from the F test, describing whether the full regression is significant

df

Degrees of freedom used by the coefficients

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

deviance

deviance

df.residual

residual degrees of freedom

augment.speedlm returns one row for each observation, with just one column added:
.fitted

Fitted values of model

See also

Examples

if (require("speedglm", quietly = TRUE)) {
    mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars)
    tidy(mod)
    glance(mod)
    augment(mod)
}

#> Warning: package 'speedglm' was built under R version 3.4.3

#> 
#> Attaching package: 'speedglm'

#> The following object is masked from 'package:boot':
#> 
#>     control

#>              .rownames  mpg    wt  qsec   .fitted
#> 1            Mazda RX4 21.0 2.620 16.46 21.815109
#> 2        Mazda RX4 Wag 21.0 2.875 17.02 21.048224
#> 3           Datsun 710 22.8 2.320 18.61 25.327279
#> 4       Hornet 4 Drive 21.4 3.215 19.44 21.580569
#> 5    Hornet Sportabout 18.7 3.440 17.02 18.196114
#> 6              Valiant 18.1 3.460 20.22 21.068588
#> 7           Duster 360 14.3 3.570 15.84 16.443423
#> 8            Merc 240D 24.4 3.190 20.00 22.227120
#> 9             Merc 230 22.8 3.150 22.90 25.123713
#> 10            Merc 280 19.2 3.440 18.30 19.385488
#> 11           Merc 280C 17.8 3.440 18.90 19.943006
#> 12          Merc 450SE 16.4 4.070 17.40 15.368981
#> 13          Merc 450SL 17.3 3.730 17.60 17.271134
#> 14         Merc 450SLC 15.2 3.780 18.00 17.390414
#> 15  Cadillac Fleetwood 10.4 5.250 17.98  9.951297
#> 16 Lincoln Continental 10.4 5.424 17.82  8.924276
#> 17   Chrysler Imperial 14.7 5.345 17.42  8.951388
#> 18            Fiat 128 32.4 2.200 19.47 26.732147
#> 19         Honda Civic 30.4 1.615 18.52 28.802478
#> 20      Toyota Corolla 33.9 1.835 19.90 28.974215
#> 21       Toyota Corona 21.5 2.465 20.01 25.896199
#> 22    Dodge Challenger 15.5 3.520 16.87 17.652896
#> 23         AMC Javelin 15.2 3.435 17.30 18.481530
#> 24          Camaro Z28 13.3 3.840 15.41 14.680913
#> 25    Pontiac Firebird 19.2 3.845 17.05 16.179557
#> 26           Fiat X1-9 27.3 1.935 18.90 27.540219
#> 27       Porsche 914-2 26.0 2.140 16.70 24.461147
#> 28        Lotus Europa 30.4 1.513 16.90 27.812072
#> 29      Ford Pantera L 15.8 3.170 14.50 17.217490
#> 30        Ferrari Dino 19.7 2.770 15.50 20.165881
#> 31       Maserati Bora 15.0 3.570 14.60 15.291217
#> 32          Volvo 142E 21.4 2.780 18.60 22.995915