summary_tidiers {broom}	R Documentation

Tidy/glance a(n) summaryDefault object

Description

For models that have only a single component, the tidy() and glance() methods are identical. Please see the documentation for both of those methods.

Usage

## S3 method for class 'summaryDefault'
tidy(x, ...)

## S3 method for class 'summaryDefault'
glance(x, ...)

Arguments

x
A summaryDefault object, created by calling summary() on a vector.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

minimum
Minimum value in original vector.

q1
First quartile of original vector.

median
Median of original vector.

mean
Mean of original vector.

q3
Third quartile of original vector.

maximum
Maximum value in original vector.

na
Number of NA values in original vector. Column present only when original vector had at least one NA entry.

See Also

tidy(), summary()

Examples

v <- rnorm(1000)
s <- summary(v)
s

tidy(s)
glance(s)

v2 <- c(v,NA)
tidy(summary(v2))

[Package broom version 0.5.0 Index]