Tidy a(n) anova object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

# S3 method for anova
tidy(x, ...)

Arguments

x

An anova object, such as those created by stats::anova(), car::Anova(), or car::leveneTest().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

Details

The term column of an ANOVA table can come with leading or trailing whitespace, which this tidying method trims.

For documentation on the tidier for car::leveneTest() output, see tidy.leveneTest()

See also

Value

A tibble::tibble() with columns:

df

Degrees of freedom used by this term in the model.

meansq

Mean sum of squares. Equal to total sum of squares divided by degrees of freedom.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

sumsq

Sum of squares explained by this term.

term

The name of the regression term.

Examples

# fit models
a <- lm(mpg ~ wt + qsec + disp, mtcars)
b <- lm(mpg ~ wt + qsec, mtcars)

mod <- anova(a, b)

# summarize model fit with tidiers
tidy(mod)
#> # A tibble: 2 × 6
#>   res.df   rss    df    sumsq statistic p.value
#>    <dbl> <dbl> <dbl>    <dbl>     <dbl>   <dbl>
#> 1     28  195.    NA NA       NA         NA    
#> 2     29  195.    -1 -0.00102  0.000147   0.990