Tidy a(n) aovlist object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for aovlist
tidy(x, ...)

Arguments

x

An aovlist objects, such as those created by stats::aov().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

The term column of an ANOVA table can come with leading or trailing whitespace, which this tidying method trims.

For documentation on the tidier for car::leveneTest() output, see tidy.leveneTest()

See also

Value

A tibble::tibble() with columns:

df

Degrees of freedom used by this term in the model.

meansq

Mean sum of squares. Equal to total sum of squares divided by degrees of freedom.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

stratum

The error stratum.

sumsq

Sum of squares explained by this term.

term

The name of the regression term.

Examples

a <- aov(mpg ~ wt + qsec + Error(disp / am), mtcars)
tidy(a)
#> # A tibble: 5 × 7
#>   stratum term         df   sumsq  meansq statistic  p.value
#>   <chr>   <chr>     <dbl>   <dbl>   <dbl>     <dbl>    <dbl>
#> 1 disp    wt            1 809.    809.        NA    NA      
#> 2 disp:am wt            1   0.389   0.389     NA    NA      
#> 3 Within  wt            1  87.2    87.2       12.0   0.00176
#> 4 Within  qsec          1  34.2    34.2        4.72  0.0387 
#> 5 Within  Residuals    27 195.      7.24      NA    NA