tidy.binWidth {broom}	R Documentation

Tidy a(n) binWidth object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'binWidth'
tidy(x, ...)

Arguments

x
A binGroup::binWidth() object.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

alternative
Alternative hypothesis (character).

ci.width
Expected width of confidence interval.

p
True proportion.

n
Total sample size

See Also

tidy(), binGroup::binWidth()

Other bingroup tidiers: glance.binDesign(), tidy.binDesign()

Examples

if (requireNamespace("binGroup", quietly = TRUE)) {

library(binGroup)
library(dplyr)
library(ggplot2)

bw <- binWidth(100, .1)
bw
tidy(bw)

}

[Package broom version 0.7.12 Index]