tidy.coeftest {broom}	R Documentation

Tidying methods for coeftest objects

Description

This tidies the result of a coefficient test, from the coeftest function in the lmtest package.

Usage

## S3 method for class 'coeftest'
tidy(x, ...)

Arguments

x
coeftest object

...
extra arguments (not used)

Value

A data.frame with one row for each coefficient, with five columns:

term
The term in the linear model being estimated and tested

estimate
The estimated coefficient

std.error
The standard error

statistic
test statistic

p.value
p-value

Examples

if (require("lmtest", quietly = TRUE)) {
    data(Mandible)
    fm <- lm(length ~ age, data=Mandible, subset=(age <= 28))
    
    coeftest(fm)
    tidy(coeftest(fm))
}

[Package broom version 0.4.5 Index]