tidy.confint.glht {broom}	R Documentation

Tidy a(n) confint.glht object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'confint.glht'
tidy(x, ...)

Arguments

x
A confint.glht object created by calling multcomp::confint.glht() on a glht object created with multcomp::glht().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See Also

tidy(), multcomp::confint.glht(), multcomp::glht()

Other multcomp tidiers: tidy.cld, tidy.glht, tidy.summary.glht

[Package broom version 0.5.0 Index]