tidy.default {broom}	R Documentation

Default tidying method

Description

By default, tidy uses as.data.frame to convert its output. This is dangerous, as it may fail with an uninformative error message. Generally tidy is intended to be used on structured model objects such as lm or htest for which a specific S3 object exists.

Usage

## Default S3 method:
tidy(x, ...)

Arguments

x
an object to be tidied

...
extra arguments (not used)

Details

If you know that you want to use as.data.frame on your untidy object, just use it directly.

Value

A data frame, from as.data.frame applied to the input x.

[Package broom version 0.4.5 Index]