Tidy a distance matrix

Tidy a distance matrix, such as that computed by the dist function, into a one-row-per-pair table. If the distance matrix does not include an upper triangle and/or diagonal, this will not either.

# S3 method for dist
tidy(x, diag = attr(x, "Diag"), upper = attr(x, "Upper"),
  ...)

Arguments

x
A "dist" object

diag
Whether to include the diagonal of the distance matrix. Defaults to whether the distance matrix includes it

upper
Whether to include the upper right triangle of the distance matrix. Defaults to whether the distance matrix includes it

...
Extra arguments, not used

Value

A data frame with one row for each pair of item distances, with columns:

item1

First item

item2

Second item

distance

Distance between items

Examples

iris_dist <- dist(t(iris[, 1:4]))
iris_dist

#>              Sepal.Length Sepal.Width Petal.Length
#> Sepal.Width      36.15785                         
#> Petal.Length     28.96619    25.77809             
#> Petal.Width      57.18304    25.86407     33.86473

tidy(iris_dist)

#>          item1        item2 distance
#> 1  Sepal.Width Sepal.Length 36.15785
#> 2 Petal.Length Sepal.Length 28.96619
#> 3  Petal.Width Sepal.Length 57.18304
#> 4 Petal.Length  Sepal.Width 25.77809
#> 5  Petal.Width  Sepal.Width 25.86407
#> 6  Petal.Width Petal.Length 33.86473

tidy(iris_dist, upper = TRUE)

#>           item1        item2 distance
#> 1   Sepal.Width Sepal.Length 36.15785
#> 2  Petal.Length Sepal.Length 28.96619
#> 3   Petal.Width Sepal.Length 57.18304
#> 4  Sepal.Length  Sepal.Width 36.15785
#> 5  Petal.Length  Sepal.Width 25.77809
#> 6   Petal.Width  Sepal.Width 25.86407
#> 7  Sepal.Length Petal.Length 28.96619
#> 8   Sepal.Width Petal.Length 25.77809
#> 9   Petal.Width Petal.Length 33.86473
#> 10 Sepal.Length  Petal.Width 57.18304
#> 11  Sepal.Width  Petal.Width 25.86407
#> 12 Petal.Length  Petal.Width 33.86473

tidy(iris_dist, diag = TRUE)

#>           item1        item2 distance
#> 1  Sepal.Length Sepal.Length  0.00000
#> 2   Sepal.Width Sepal.Length 36.15785
#> 3  Petal.Length Sepal.Length 28.96619
#> 4   Petal.Width Sepal.Length 57.18304
#> 6   Sepal.Width  Sepal.Width  0.00000
#> 7  Petal.Length  Sepal.Width 25.77809
#> 8   Petal.Width  Sepal.Width 25.86407
#> 11 Petal.Length Petal.Length  0.00000
#> 12  Petal.Width Petal.Length 33.86473
#> 16  Petal.Width  Petal.Width  0.00000