Tidy a(n) fitdistr object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for fitdistr
tidy(x, ...)

Arguments

x
A fitdistr object returned by MASS::fitdistr().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row for estimated parameter, with columns:

term

The term that was estimated

estimate

Estimated value

std.error

Standard error of estimate

See also

Examples

set.seed(2015)
x <- rnorm(100, 5, 2)

library(MASS)
fit <- fitdistr(x, dnorm, list(mean = 3, sd = 1))

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

tidy(fit)

#> # A tibble: 2 x 3
#>   term  estimate std.error
#>   <chr>    <dbl>     <dbl>
#> 1 mean      4.90     0.201
#> 2 sd        2.01     0.142

glance(fit)

#> # A tibble: 1 x 4
#>       n logLik   AIC   BIC
#>   <int>  <dbl> <dbl> <dbl>
#> 1   100  -212.  427.  433.