Tidy a(n) fixest object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for fixest
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x
A fixest object returned from any of the fixest estimators

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments passed to summary and confint. Important arguments are se and cluster. Other arguments are dof, exact_dof, forceCovariance, and keepBounded. See summary.fixest.

Details

The fixest package provides a family of functions for estimating models with arbitrary numbers of fixed-effects, in both an OLS and a GLM context. The package also supports robust (i.e. White) and clustered standard error reporting via the generic summary.fixest() command. In a similar vein, the tidy() method for these models allows users to specify a desired standard error correction either 1) implicitly via the supplied fixest object, or 2) explicitly as part of the tidy call. See examples below.

Note that fixest confidence intervals are calculated assuming a normal distribution -- this assumes infinite degrees of freedom for the CI. (This assumption is distinct from the degrees of freedom used to calculate the standard errors. For more on degrees of freedom with clusters and fixed effects, see https://github.com/lrberge/fixest/issues/6 and https://github.com/sgaure/lfe/issues/1#issuecomment-530646990)

See also

Value

A tibble::tibble() with columns:

conf.high

Upper bound on the confidence interval for the estimate.

conf.low

Lower bound on the confidence interval for the estimate.

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

Examples

# \donttest{
library(fixest)
gravity <- feols(log(Euros) ~ log(dist_km) | Origin + Destination + Product + Year, trade)

tidy(gravity)

#> # A tibble: 1 x 5
#>   term         estimate std.error statistic       p.value
#>   <chr>           <dbl>     <dbl>     <dbl>         <dbl>
#> 1 log(dist_km)    -2.17     0.154     -14.1 0.00000000119

glance(gravity)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared within.r.squared pseudo.r.squared sigma  nobs     AIC
#>       <dbl>         <dbl>            <dbl>            <dbl> <dbl> <int>   <dbl>
#> 1     0.706         0.705            0.219               NA  1.74 38325 151481.
#> # … with 2 more variables: BIC <dbl>, logLik <dbl>

augment(gravity, trade)

#> # A tibble: 38,325 x 9
#>    .rownames Destination Origin Product  Year dist_km    Euros .fitted  .resid
#>    <chr>     <fct>       <fct>    <int> <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 1         LU          BE           1  2007    140.  2966697    14.1  0.812 
#>  2 2         BE          LU           1  2007    140.  6755030    13.0  2.75  
#>  3 3         LU          BE           2  2007    140. 57078782    16.9  0.924 
#>  4 4         BE          LU           2  2007    140.  7117406    15.8 -0.0470
#>  5 5         LU          BE           3  2007    140. 17379821    16.3  0.378 
#>  6 6         BE          LU           3  2007    140.  2622254    15.2 -0.402 
#>  7 7         LU          BE           4  2007    140. 64867588    17.4  0.595 
#>  8 8         BE          LU           4  2007    140. 10731757    16.3 -0.0937
#>  9 9         LU          BE           5  2007    140.   330702    14.1 -1.37  
#> 10 10        BE          LU           5  2007    140.     7706    13.0 -4.02  
#> # … with 38,315 more rows

## To get robust or clustered SEs, users can either:
# 1) Or, specify the arguments directly in the tidy() call
tidy(gravity, conf.int = TRUE, cluster = c("Product", "Year"))

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic  p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17    0.0760     -28.5 3.88e-10    -2.32     -2.02

tidy(gravity, conf.int = TRUE, se = "threeway")

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic       p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>         <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17     0.175     -12.4 0.00000000608    -2.51     -1.83

# 2) Feed tidy() a summary.fixest object that has already accepted these arguments
gravity_summ <- summary(gravity, cluster = c("Product", "Year"))
tidy(gravity_summ, conf.int = TRUE)

#> # A tibble: 1 x 7
#>   term         estimate std.error statistic  p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 log(dist_km)    -2.17    0.0760     -28.5 3.88e-10    -2.32     -2.02

# Approach (1) is preferred.

## The other fixest methods all work similarly. For example:
gravity_pois <- feglm(Euros ~ log(dist_km) | Origin + Destination + Product + Year, trade)
tidy(gravity_pois)

#> # A tibble: 1 x 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 log(dist_km)    -1.53     0.116     -13.2 7.89e-40

glance(gravity_pois)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared within.r.squared pseudo.r.squared sigma  nobs     AIC
#>       <dbl>         <dbl>            <dbl>            <dbl> <dbl> <int>   <dbl>
#> 1        NA            NA               NA            0.764    NA 38325 1.40e12
#> # … with 2 more variables: BIC <dbl>, logLik <dbl>

augment(gravity_pois, trade)

#> # A tibble: 38,325 x 9
#>    .rownames Destination Origin Product  Year dist_km    Euros .fitted    .resid
#>    <chr>     <fct>       <fct>    <int> <dbl>   <dbl>    <dbl>   <dbl>     <dbl>
#>  1 1         LU          BE           1  2007    140.  2966697    16.0   -6.06e6
#>  2 2         BE          LU           1  2007    140.  6755030    15.4    1.97e6
#>  3 3         LU          BE           2  2007    140. 57078782    17.4    2.00e7
#>  4 4         BE          LU           2  2007    140.  7117406    16.8   -1.26e7
#>  5 5         LU          BE           3  2007    140. 17379821    16.7   -1.00e4
#>  6 6         BE          LU           3  2007    140.  2622254    16.0   -6.60e6
#>  7 7         LU          BE           4  2007    140. 64867588    17.5    2.64e7
#>  8 8         BE          LU           4  2007    140. 10731757    16.8   -9.64e6
#>  9 9         LU          BE           5  2007    140.   330702    14.5   -1.64e6
#> 10 10        BE          LU           5  2007    140.     7706    13.9   -1.04e6
#> # … with 38,315 more rows

# }