Tidy a(n) gam object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for gam
tidy(x, parametric = FALSE, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x
A gam object returned from a call to mgcv::gam().

parametric
Logical indicating if parametric or smooth terms should be tidied. Defaults to FALSE, meaning that smooth terms are tidied by default.

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

When parametric = FALSE return columns edf and ref.df rather than estimate and std.error.

See also

Value

A tibble::tibble() with columns:

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

edf

The effective degrees of freedom. Only reported when `parametric = FALSE`

ref.df

The reference degrees of freedom. Only reported when `parametric = FALSE`

Examples

g <- mgcv::gam(mpg ~ s(hp) + am + qsec, data = mtcars)

tidy(g)

#> # A tibble: 1 x 5
#>   term    edf ref.df statistic p.value
#>   <chr> <dbl>  <dbl>     <dbl>   <dbl>
#> 1 s(hp)  2.36   3.02      6.34 0.00218

tidy(g, parametric = TRUE)

#> # A tibble: 3 x 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept)  16.7        9.83      1.70  0.101  
#> 2 am            4.37       1.56      2.81  0.00918
#> 3 qsec          0.0904     0.525     0.172 0.865  

glance(g)

#> # A tibble: 1 x 7
#>      df logLik   AIC   BIC deviance df.residual  nobs
#>   <dbl>  <dbl> <dbl> <dbl>    <dbl>       <dbl> <int>
#> 1  5.36  -74.4  162.  171.     196.        26.6    32

augment(g)

#> # A tibble: 32 x 11
#>    .rownames          mpg    am  qsec    hp .fitted .se.fit .resid   .hat .sigma
#>    <chr>            <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl>  <dbl>  <dbl> <lgl> 
#>  1 Mazda RX4         21       1  16.5   110    24.3   1.03  -3.25  0.145  NA    
#>  2 Mazda RX4 Wag     21       1  17.0   110    24.3   0.925 -3.30  0.116  NA    
#>  3 Datsun 710        22.8     1  18.6    93    26.0   0.894 -3.22  0.109  NA    
#>  4 Hornet 4 Drive    21.4     0  19.4   110    20.2   0.827  1.25  0.0930 NA    
#>  5 Hornet Sportabo…  18.7     0  17.0   175    15.7   0.815  3.02  0.0902 NA    
#>  6 Valiant           18.1     0  20.2   105    20.7   0.914 -2.56  0.113  NA    
#>  7 Duster 360        14.3     0  15.8   245    12.7   1.11   1.63  0.167  NA    
#>  8 Merc 240D         24.4     0  20      62    25.0   1.45  -0.618 0.287  NA    
#>  9 Merc 230          22.8     0  22.9    95    21.8   1.81   0.959 0.446  NA    
#> 10 Merc 280          19.2     0  18.3   123    19.0   0.864  0.211 0.102  NA    
#> # … with 22 more rows, and 1 more variable: .cooksd <dbl>