Tidy a(n) gamlss object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for gamlss
tidy(x, ...)

Arguments

x
A gamlss object returned from gamlss::gamlss().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

parameter

Type of coefficient being estimated: `mu`, `sigma`, `nu`, or `tau`.

Examples

library(gamlss)

#> Loading required package: splines

#> Loading required package: gamlss.data

#> 
#> Attaching package: ‘gamlss.data’

#> The following object is masked from ‘package:boot’:
#> 
#>     aids

#> The following object is masked from ‘package:mclust’:
#> 
#>     acidity

#> The following object is masked from ‘package:datasets’:
#> 
#>     sleep

#> Loading required package: gamlss.dist

#> Loading required package: nlme

#> 
#> Attaching package: ‘nlme’

#> The following objects are masked from ‘package:ordinal’:
#> 
#>     ranef, VarCorr

#> The following object is masked from ‘package:dplyr’:
#> 
#>     collapse

#> Loading required package: parallel

#>  **********   GAMLSS Version 5.1-7  ********** 

#> For more on GAMLSS look at http://www.gamlss.com/

#> Type gamlssNews() to see new features/changes/bug fixes.

#> 
#> Attaching package: ‘gamlss’

#> The following object is masked from ‘package:caret’:
#> 
#>     calibration

g <- gamlss(
  y ~ pb(x),
  sigma.fo = ~ pb(x),
  family = BCT,
  data = abdom,
  method = mixed(1, 20)
)

#> GAMLSS-RS iteration 1: Global Deviance = 4771.925 
#> GAMLSS-CG iteration 1: Global Deviance = 4771.013 
#> GAMLSS-CG iteration 2: Global Deviance = 4770.994 
#> GAMLSS-CG iteration 3: Global Deviance = 4770.994 

tidy(g)

#> # A tibble: 6 x 6
#>   parameter term        estimate std.error statistic   p.value
#>   <chr>     <chr>          <dbl>     <dbl>     <dbl>     <dbl>
#> 1 mu        (Intercept) -64.4      1.33      -48.5   1.89e-210
#> 2 mu        pb(x)        10.7      0.0578    185.    0.       
#> 3 sigma     (Intercept)  -2.65     0.108     -24.5   8.09e- 93
#> 4 sigma     pb(x)        -0.0100   0.00378    -2.65  8.29e-  3
#> 5 nu        (Intercept)  -0.107    0.557      -0.192 8.48e-  1
#> 6 tau       (Intercept)   2.49     0.301       8.28  7.77e- 16