tidy.glmRob {broom}	R Documentation

Tidy a(n) glmRob object

Description

This method wraps tidy.lm().

Usage

## S3 method for class 'glmRob'
tidy(x, ...)

Arguments

x
A glmRob object returned from robust::glmRob().

...
Arguments passed on to tidy.lm

x

An lm object created by stats::lm().

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

quick

Logical indiciating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

exponentiate

Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

Details

For tidiers for robust models from the MASS package see tidy.rlm().

Value

A tibble::tibble() with one row for each term in the regression. The tibble has columns:

term
The name of the regression term.

estimate
The estimated value of the regression term.

std.error
The standard error of the regression term.

statistic
The value of a statistic, almost always a T-statistic, to use in a hypothesis that the regression term is non-zero.

p.value
The two-sided p-value associated with the observed statistic.

conf.low
The low end of a confidence interval for the regression term. Included only if conf.int = TRUE.

conf.high
The high end of a confidence interval for the regression term. Included only if conf.int = TRUE.

If the linear model is an mlm object (multiple linear model), there is an additional column:

response
Which response column the coefficients correspond to (typically Y1, Y2, etc)

See Also

tidy(), tidy.lm()

robust::glmRob()

Other robust tidiers: augment.glmRob, augment.lmRob, glance.glmRob, glance.lmRob, tidy.lmRob

Examples

library(robust)
m <- lmRob(mpg ~ wt, data = mtcars)

tidy(m)
augment(m)
glance(m)

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
glance(gm)

[Package broom version 0.5.0 Index]