Tidy a(n) lm.beta object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for lm.beta
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

# S3 method for summary.lm.beta
tidy(x, ...)

Arguments

x
An lm.beta object created by lm.beta::lm.beta.

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

If the linear model is an mlm object (multiple linear model), there is an additional column response.

If you have missing values in your model data, you may need to refit the model with na.action = na.exclude.

See also

Value

A tibble::tibble() with columns:

conf.high

Upper bound on the confidence interval for the estimate.

conf.low

Lower bound on the confidence interval for the estimate.

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

Examples

library(lm.beta)

mod <- stats::lm(speed ~ ., data = cars)

# standardize
mod_standardized <- lm.beta::lm.beta(mod)
tidy(mod_standardized)

#> # A tibble: 2 x 6
#>   term        estimate std_estimate std.error statistic  p.value
#>   <chr>          <dbl>        <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)    8.28         0        0.874       9.47 1.44e-12
#> 2 dist           0.166        0.807    0.0175      9.46 1.49e-12

## Taken from lm/ lm.beta help
##
## Annette Dobson (1990) "An Introduction to Generalized Linear Models".
## Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- stats::lm(weight ~ group)

# standardize
lm.D9.beta <- lm.beta::lm.beta(lm.D9)
tidy(lm.D9.beta)

#> # A tibble: 2 x 6
#>   term        estimate std_estimate std.error statistic  p.value
#>   <chr>          <dbl>        <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)    5.03         0         0.220     22.9  9.55e-15
#> 2 groupTrt      -0.371       -0.270     0.311     -1.19 2.49e- 1