tidy.manova {broom}	R Documentation

Tidy a(n) manova object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'manova'
tidy(x, test = "Pillai", ...)

Arguments

x
A manova object return from stats::manova().

test
One of "Pillai" (Pillai's trace), "Wilks" (Wilk's lambda), "Hotelling-Lawley" (Hotelling-Lawley trace) or "Roy" (Roy's greatest root) indicating which test statistic should be used. Defaults to "Pillai".

...
Arguments passed on to stats::summary.manova

object

An object of class "manova" or an aov object with multiple responses.

test

The name of the test statistic to be used. Partial matching is used so the name can be abbreviated.

intercept

logical. If TRUE, the intercept term is included in the table.

tol

tolerance to be used in deciding if the residuals are rank-deficient: see qr.

Value

A tibble::tibble with columns:

\item{term}{Term in design}
\item{statistic}{Approximate F statistic}
\item{num.df}{Degrees of freedom}
\item{p.value}{P-value}

Depending on which test statistic is specified, one of the following columns is also included:

\item{pillai}{Pillai's trace}
\item{wilks}{Wilk's lambda}
\item{hl}{Hotelling-Lawley trace}
\item{roy}{Roy's greatest root}

See Also

tidy(), stats::summary.manova()

Other anova tidiers: tidy.TukeyHSD, tidy.anova, tidy.aovlist, tidy.aov

Examples

npk2 <- within(npk, foo <- rnorm(24))
m <- manova(cbind(yield, foo) ~ block + N * P * K, npk2)
tidy(m) 

[Package broom version 0.5.0 Index]