tidy.mlogit {broom}	R Documentation

Tidying methods for logit models

Description

These methods tidy the coefficients of mnl and nl models generated by the functions of the mlogit package.

Usage

## S3 method for class 'mlogit'
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x
an object returned from mlogit::mlogit().

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

conf.high
Upper bound on the confidence interval for the estimate.

conf.low
Lower bound on the confidence interval for the estimate.

estimate
The estimated value of the regression term.

p.value
The two-sided p-value associated with the observed statistic.

statistic
The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error
The standard error of the regression term.

term
The name of the regression term.

See Also

tidy(), mlogit::mlogit()

Other mlogit tidiers: augment.mlogit(), glance.mlogit()

Examples

if (requireNamespace("mlogit", quietly = TRUE)) {

## Not run: 
library(mlogit)
data("Fishing", package = "mlogit")
Fish <- dfidx(Fishing, varying = 2:9, shape = "wide", choice = "mode")
m <- mlogit(mode ~ price + catch | income, data = Fish)

tidy(m)
augment(m)
glance(m)

## End(Not run)

}

[Package broom version 0.7.12 Index]