tidy.orcutt {broom}	R Documentation

Tidy a(n) orcutt object

Description

This method wraps tidy.lm().

Usage

## S3 method for class 'orcutt'
tidy(x, ...)

Arguments

x
An orcutt object returned from orcutt::cochrane.orcutt().

...
Arguments passed on to tidy.lm

x

An lm object created by stats::lm().

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

quick

Logical indiciating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

exponentiate

Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

Value

tidy returns the same information as tidy.lm(), though without confidence interval options.

See Also

tidy(), tidy.lm()

orcutt::cochrane.orcutt()

Other orcutt tidiers: glance.orcutt

Examples

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

if (require("orcutt", quietly = TRUE)) {
  co <- cochrane.orcutt(reg)
  co

  tidy(co)
  glance(co)
}

[Package broom version 0.5.0 Index]