tidy.pam {broom}	R Documentation

Tidy a(n) pam object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'pam'
tidy(x, col.names = paste0("x", 1:ncol(x$medoids)), ...)

Arguments

x
An pam object returned from cluster::pam()

col.names
Column names in the input data frame. Defaults to the names of the variables in x.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

For examples, see the pam vignette.

Value

A tibble::tibble() with columns:

size
Size of each cluster.

max.diss
Maximal dissimilarity between the observations in the cluster and that cluster's medoid.

avg.diss
Average dissimilarity between the observations in the cluster and that cluster's medoid.

diameter
Diameter of the cluster.

separation
Separation of the cluster.

avg.width
Average silhouette width of the cluster.

cluster
A factor describing the cluster from 1:k.

See Also

tidy(), cluster::pam()

Other pam tidiers: augment.pam(), glance.pam()

Examples

## Not run: 
library(dplyr)
library(ggplot2)
library(cluster)
library(modeldata)
data(hpc_data)

x <- hpc_data[, 2:5]
p <- pam(x, k = 4)

tidy(p)
glance(p)
augment(p, x)

augment(p, x) %>%
  ggplot(aes(compounds, input_fields)) +
  geom_point(aes(color = .cluster)) +
  geom_text(aes(label = cluster), data = tidy(p), size = 10)

## End(Not run)   
  

[Package broom version 0.7.12 Index]