Tidy a(n) speedlm object

This method wraps tidy.lm().

# S3 method for speedlm
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

Arguments

x
A speedlm object returned from speedglm::speedlm().

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Examples

mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars, fitted = TRUE)

tidy(mod)

#> # A tibble: 3 x 5
#>   term        estimate std.error statistic  p.value
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)   19.7       5.25       3.76 7.65e- 4
#> 2 wt            -5.05      0.484    -10.4  2.52e-11
#> 3 qsec           0.929     0.265      3.51 1.50e- 3

glance(mod)

#> # A tibble: 1 x 11
#>   r.squared adj.r.squared statistic  p.value    df logLik   AIC   BIC deviance
#>       <dbl>         <dbl>     <dbl>    <dbl> <int>  <dbl> <dbl> <dbl>    <dbl>
#> 1     0.826         0.814      69.0 9.39e-12     3  -74.4  157.  163.     195.
#> # … with 2 more variables: df.residual <int>, nobs <int>

augment(mod)

#> # A tibble: 32 x 6
#>    .rownames           mpg    wt  qsec .fitted  .resid
#>    <chr>             <dbl> <dbl> <dbl>   <dbl>   <dbl>
#>  1 Mazda RX4          21    2.62  16.5    21.8  0.815 
#>  2 Mazda RX4 Wag      21    2.88  17.0    21.0  0.0482
#>  3 Datsun 710         22.8  2.32  18.6    25.3  2.53  
#>  4 Hornet 4 Drive     21.4  3.22  19.4    21.6  0.181 
#>  5 Hornet Sportabout  18.7  3.44  17.0    18.2 -0.504 
#>  6 Valiant            18.1  3.46  20.2    21.1  2.97  
#>  7 Duster 360         14.3  3.57  15.8    16.4  2.14  
#>  8 Merc 240D          24.4  3.19  20      22.2 -2.17  
#>  9 Merc 230           22.8  3.15  22.9    25.1  2.32  
#> 10 Merc 280           19.2  3.44  18.3    19.4  0.185 
#> # … with 22 more rows