Tidy a(n) survdiff object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for survdiff
tidy(x, strata = FALSE, ...)

Arguments

x
An survdiff object returned from survival::survdiff().

strata
logical, whether to include strata in the output

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row for each time point and columns:

...

The initial columns correspond to the grouping factors on the right hand side of the model formula.

obs

weighted observed number of events in each group

exp

weighted expected number of events in each group

N

number of subjects in each group

See also

Examples

library(survival)

s <- survdiff(
  Surv(time, status) ~ pat.karno + strata(inst),
  data = lung
)

tidy(s)

#> # A tibble: 8 x 4
#>   pat.karno     N   obs    exp
#>   <chr>     <dbl> <dbl>  <dbl>
#> 1 30            2     1  0.692
#> 2 40            2     1  1.10 
#> 3 50            4     4  1.17 
#> 4 60           30    27 16.3  
#> 5 70           41    31 26.4  
#> 6 80           50    38 41.9  
#> 7 90           60    38 47.2  
#> 8 100          35    21 26.2  

glance(s)

#> # A tibble: 1 x 3
#>   statistic    df p.value
#>       <dbl> <dbl>   <dbl>
#> 1      21.4     7 0.00326