tidy a table object

A table, typically created by the table function, contains a contingency table of frequencies across multiple vectors. This directly calls the as.data.frame.table method, which melts it into a data frame with one column for each variable and a Freq column.

# S3 method for table
tidy(x, ...)

Arguments

x
An object of class "table"

...
Extra arguments (not used)

See also

Examples

tab <- with(airquality, table(cut(Temp, quantile(Temp)), Month))
tidy(tab)

#>       Var1 Month Freq
#> 1  (56,72]     5   24
#> 2  (72,79]     5    5
#> 3  (79,85]     5    1
#> 4  (85,97]     5    0
#> 5  (56,72]     6    3
#> 6  (72,79]     6   15
#> 7  (79,85]     6    7
#> 8  (85,97]     6    5
#> 9  (56,72]     7    0
#> 10 (72,79]     7    2
#> 11 (79,85]     7   19
#> 12 (85,97]     7   10
#> 13 (56,72]     8    1
#> 14 (72,79]     8    9
#> 15 (79,85]     8    7
#> 16 (85,97]     8   14
#> 17 (56,72]     9   10
#> 18 (72,79]     9   10
#> 19 (79,85]     9    5
#> 20 (85,97]     9    5