Tidy a(n) Gam object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for Gam
tidy(x, ...)

Arguments

x
A Gam object returned from a call to gam::gam().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

Tidy gam objects created by calls to mgcv::gam() with tidy.gam().

See also

Value

A tibble::tibble() with columns:

df

Degrees of freedom used by this term in the model.

meansq

Mean sum of squares. Equal to total sum of squares divided by degrees of freedom.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

sumsq

Sum of squares explained by this term.

term

The name of the regression term.

Examples

library(gam)

#> Loading required package: splines

#> Loading required package: foreach

#> Loaded gam 1.20

g <- gam(mpg ~ s(hp, 4) + am + qsec, data = mtcars)

tidy(g)

#> # A tibble: 4 x 6
#>   term         df    sumsq   meansq statistic   p.value
#>   <chr>     <dbl>    <dbl>    <dbl>     <dbl>     <dbl>
#> 1 s(hp, 4)    1   678.     678.      94.4      5.73e-10
#> 2 am          1   113.     113.      15.7      5.52e- 4
#> 3 qsec        1     0.0263   0.0263   0.00366  9.52e- 1
#> 4 Residuals  25.0 180.       7.19    NA       NA       

glance(g)

#> # A tibble: 1 x 7
#>      df logLik   AIC   BIC deviance df.residual  nobs
#>   <dbl>  <dbl> <dbl> <dbl>    <dbl>       <dbl> <int>
#> 1  7.00  -76.0  162.  169.     180.        25.0    32