Tidy atomic vectors

Vector tidiers are deprecated and will be removed from an upcoming release of broom.

# S3 method for numeric
tidy(x, ...)

# S3 method for character
tidy(x, ...)

# S3 method for logical
tidy(x, ...)

Arguments

x
An object of class "numeric", "integer", "character", or "logical". Most likely a named vector

...
Extra arguments (not used)

Details

Turn atomic vectors into data frames, where the names of the vector (if they exist) are a column and the values of the vector are a column.

Examples

# NOT RUN {
x <- 1:5
names(x) <- letters[1:5]
tidy(x)
# }