Tidiers for aareg objects

These tidy the coefficients of Aalen additive regression objects.

# S3 method for aareg
tidy(x, ...)

# S3 method for aareg
glance(x, ...)

Arguments

x
an "aareg" object

...
extra arguments (not used)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.aareg returns one row for each coefficient, with the columns

term

name of coefficient

estimate

estimate of the slope

statistic

test statistic for coefficient

std.error

standard error of statistic

robust.se

robust version of standard error estimate (only when x was called with dfbeta = TRUE)

z

z score

p.value

p-value

glance returns a one-row data frame containing
statistic

chi-squared statistic

p.value

p-value based on chi-squared statistic

df

degrees of freedom used by coefficients

Examples

if (require("survival", quietly = TRUE)) {
    afit <- aareg(Surv(time, status) ~ age + sex + ph.ecog, data=lung,
                  dfbeta=TRUE)
    summary(afit)
    tidy(afit)
}

#> # A tibble: 4 x 7
#>   term        estimate  statistic std.error robust.se statistic.z  p.value
#>   <chr>          <dbl>      <dbl>     <dbl>     <dbl>       <dbl>    <dbl>
#> 1 Intercept  0.00505    0.00587   0.00474   0.00477          1.23 0.219   
#> 2 age        0.0000401  0.0000715 0.0000723 0.0000700        1.02 0.307   
#> 3 sex       -0.00316   -0.00403   0.00122   0.00123         -3.28 0.00103 
#> 4 ph.ecog    0.00301    0.00367   0.00102   0.00102          3.62 0.000299