Tidying methods for ANOVA and AOV objects

Tidies the result of an analysis of variance into an ANOVA table. Only a tidy method is provided, not an augment or glance method.

# S3 method for anova
tidy(x, ...)

# S3 method for aov
tidy(x, ...)

# S3 method for aovlist
tidy(x, ...)

Arguments

x
An object of class anova, aov, or aovlist.

...
Additional arguments (not used).

Value

A `tibble`` with columns

term

Term within the model, or "Residuals"

df

Degrees of freedom used by this term in the model

sumsq

Sum of squares explained by this term

meansq

Mean of sum of squares among degrees of freedom

statistic

F statistic

p.value

P-value from F test

In the case of an aovlist object, there is also a stratum column describing the error stratum
Details

Note that the "term" column of an ANOVA table can come with leading or trailing whitespace, which this tidying method trims.

Note that Anova from car (perhaps counter-intuively) outputs an object of class anova for generalized linear models. These objects are also supported.

Examples

a <- anova(lm(mpg ~ wt + qsec + disp, mtcars))
tidy(a)

#> # A tibble: 4 x 6
#>   term         df     sumsq    meansq  statistic   p.value
#>   <chr>     <int>     <dbl>     <dbl>      <dbl>     <dbl>
#> 1 wt            1 848.      848.      121.        1.08e-11
#> 2 qsec          1  82.9      82.9      11.9       1.82e- 3
#> 3 disp          1   0.00102   0.00102   0.000147  9.90e- 1
#> 4 Residuals    28 195.        6.98     NA        NA       

a <- aov(mpg ~ wt + qsec + disp, mtcars)
tidy(a)

#> # A tibble: 4 x 6
#>   term         df     sumsq    meansq  statistic   p.value
#>   <chr>     <dbl>     <dbl>     <dbl>      <dbl>     <dbl>
#> 1 wt            1 848.      848.      121.        1.08e-11
#> 2 qsec          1  82.9      82.9      11.9       1.82e- 3
#> 3 disp          1   0.00102   0.00102   0.000147  9.90e- 1
#> 4 Residuals    28 195.        6.98     NA        NA       

al <- aov(mpg ~ wt + qsec + Error(disp / am), mtcars)
tidy(al)

#> # A tibble: 5 x 7
#>   stratum term         df   sumsq  meansq statistic  p.value
#>   <chr>   <chr>     <dbl>   <dbl>   <dbl>     <dbl>    <dbl>
#> 1 disp    wt            1 809.    809.        NA    NA      
#> 2 disp:am wt            1   0.389   0.389     NA    NA      
#> 3 Within  wt            1  87.2    87.2       12.0   0.00176
#> 4 Within  qsec          1  34.2    34.2        4.72  0.0387 
#> 5 Within  Residuals    27 195.      7.24      NA    NA