Augment a(n) glmRob object

This augment method wraps augment.lm().

# S3 method for glmRob
augment(x, data = stats::model.frame(x),
  newdata = NULL, type.predict, type.residuals, ...)

Arguments

x
An lm object created by stats::lm().

data
A data.frame() or tibble::tibble() containing the original data that was used to produce the object x. Defaults to stats::model.frame(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata
A data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

type.predict
Type of predictions to use when x is a glm object. Passed to stats::predict.glm().

type.residuals
Type of residuals to use when x is a glm object. Passed to stats::residuals.glm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

For tidiers for robust models from the MASS package see tidy.rlm().

See also

Examples

library(robust)

#> Loading required package: fit.models

m <- lmRob(mpg ~ wt, data = mtcars)

tidy(m)

#> # A tibble: 2 x 5
#>   term        estimate std.error statistic  p.value
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)    35.6       3.58      9.93 5.37e-11
#> 2 wt             -4.91      1.09     -4.49 9.67e- 5

augment(m)

#> # A tibble: 32 x 6
#>    .rownames           mpg    wt .fitted .se.fit  .resid
#>  * <chr>             <dbl> <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Mazda RX4          21    2.62    22.7   1.12  -1.68  
#>  2 Mazda RX4 Wag      21    2.88    21.4   1.00  -0.431 
#>  3 Datsun 710         22.8  2.32    24.2   1.32  -1.36  
#>  4 Hornet 4 Drive     21.4  3.22    19.8   0.957  1.64  
#>  5 Hornet Sportabout  18.7  3.44    18.7   1.00   0.0445
#>  6 Valiant            18.1  3.46    18.6   1.01  -0.457 
#>  7 Duster 360         14.3  3.57    18.0   1.06  -3.72  
#>  8 Merc 240D          24.4  3.19    19.9   0.955  4.52  
#>  9 Merc 230           22.8  3.15    20.1   0.955  2.72  
#> 10 Merc 280           19.2  3.44    18.7   1.00   0.545 
#> # ... with 22 more rows

glance(m)

#> # A tibble: 1 x 4
#>   r.squared deviance sigma df.residual
#>       <dbl>    <dbl> <dbl>       <int>
#> 1     0.567     136.  2.95          30

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
glance(gm)

#> # A tibble: 1 x 3
#>   deviance null.deviance df.residual
#>      <dbl>         <dbl>       <int>
#> 1     19.2          44.4          30