Tidy a(n) loess object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for loess
augment(x, data = stats::model.frame(x), newdata, ...)

Arguments

x
A loess objects returned by stats::loess().

data
A data.frame() or tibble::tibble() containing the original data that was used to produce the object x. Defaults to stats::model.frame(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata
A data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

...
Arguments passed on to stats:::predict.loess

object

an object fitted by loess.

newdata

an optional data frame in which to look for variables with which to predict, or a matrix or vector containing exactly the variables needs for prediction. If missing, the original data points are used.

se

should standard errors be computed?

na.action

function determining what should be done with missing values in data frame newdata. The default is to predict NA.

Value

When newdata is not supplied augment.loess returns one row for each observation with three columns added to the original data:

.fitted

Fitted values of model

.se.fit

Standard errors of the fitted values

.resid

Residuals of the fitted values

When newdata is supplied augment.loess returns one row for each observation with one additional column:
.fitted

Fitted values of model

.se.fit

Standard errors of the fitted values

Details

When the modeling was performed with na.action = "na.omit" (as is the typical default), rows with NA in the initial data are omitted entirely from the augmented data frame. When the modeling was performed with na.action = "na.exclude", one should provide the original data as a second argument, at which point the augmented data will contain those rows (typically with NAs in place of the new columns). If the original data is not provided to augment() and na.action = "na.exclude", a warning is raised and the incomplete rows are dropped.

See also

Examples

lo <- loess(mpg ~ wt, mtcars)
augment(lo)

#> # A tibble: 32 x 6
#>    .rownames           mpg    wt .fitted .se.fit  .resid
#>  * <chr>             <dbl> <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Mazda RX4          21    2.62    22.2   1.01  -1.24  
#>  2 Mazda RX4 Wag      21    2.88    21.1   1.07  -0.0786
#>  3 Datsun 710         22.8  2.32    24.6   1.03  -1.84  
#>  4 Hornet 4 Drive     21.4  3.22    19.5   0.861  1.91  
#>  5 Hornet Sportabout  18.7  3.44    17.9   0.755  0.779 
#>  6 Valiant            18.1  3.46    17.8   0.747  0.290 
#>  7 Duster 360         14.3  3.57    17.2   0.721 -2.89  
#>  8 Merc 240D          24.4  3.19    19.6   0.866  4.78  
#>  9 Merc 230           22.8  3.15    19.8   0.868  2.98  
#> 10 Merc 280           19.2  3.44    17.9   0.755  1.28  
#> # ... with 22 more rows

# with all columns of original data
augment(lo, mtcars)

#> # A tibble: 32 x 15
#>    .rownames    mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>  * <chr>      <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4   21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2 Mazda RX4~  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3 Datsun 710  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4 Hornet 4 ~  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5 Hornet Sp~  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6 Valiant     18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7 Duster 360  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8 Merc 240D   24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9 Merc 230    22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10 Merc 280    19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ... with 22 more rows, and 3 more variables: .fitted <dbl>, .se.fit <dbl>,
#> #   .resid <dbl>

# with a new dataset
augment(lo, newdata = head(mtcars))

#> # A tibble: 6 x 14
#>   .rownames     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> * <chr>       <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Mazda RX4    21       6   160   110  3.9   2.62  16.5     0     1     4     4
#> 2 Mazda RX4 ~  21       6   160   110  3.9   2.88  17.0     0     1     4     4
#> 3 Datsun 710   22.8     4   108    93  3.85  2.32  18.6     1     1     4     1
#> 4 Hornet 4 D~  21.4     6   258   110  3.08  3.22  19.4     1     0     3     1
#> 5 Hornet Spo~  18.7     8   360   175  3.15  3.44  17.0     0     0     3     2
#> 6 Valiant      18.1     6   225   105  2.76  3.46  20.2     1     0     3     1
#> # ... with 2 more variables: .fitted <dbl>, .se.fit <dbl>