Tidiers for biglm and bigglm object

Tidiers for biglm object from the "biglm" package, which contains a linear model object that is limited in memory usage. Generally the behavior is as similar to the lm_tidiers() as is possible. Currently no augment is defined.

# S3 method for biglm
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)

# S3 method for biglm
glance(x, ...)

Arguments

x
a "biglm" object

conf.int
whether to include a confidence interval

conf.level
confidence level of the interval, used only if conf.int=TRUE

exponentiate
whether to exponentiate the coefficient estimates and confidence intervals (typical for logistic regression)

quick
whether to compute a smaller and faster version, containing only the term and estimate columns.

...
extra arguments (not used)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.biglm returns one row for each coefficient, with columns

term

The term in the linear model being estimated and tested

estimate

The estimated coefficient

std.error

The standard error from the linear model

p.value

two-sided p-value

If conf.int=TRUE, it also includes columns for conf.low and conf.high, computed with confint(). glance.biglm returns a one-row data frame, with columns
r.squared

The percent of variance explained by the model

AIC

the Akaike Information Criterion

deviance

deviance

df.residual

residual degrees of freedom

Examples

if (require("biglm", quietly = TRUE)) {
    bfit <- biglm(mpg ~ wt + disp, mtcars)
    tidy(bfit)
    tidy(bfit, conf.int = TRUE)
    tidy(bfit, conf.int = TRUE, conf.level = .9)

    glance(bfit)

    # bigglm: logistic regression
    bgfit <- bigglm(am ~ mpg, mtcars, family = binomial())
    tidy(bgfit)
    tidy(bgfit, exponentiate = TRUE)
    tidy(bgfit, conf.int = TRUE)
    tidy(bgfit, conf.int = TRUE, conf.level = .9)
    tidy(bgfit, conf.int = TRUE, conf.level = .9, exponentiate = TRUE)

    glance(bgfit)
}

#> # A tibble: 1 x 3
#>   r.squared   AIC deviance
#>       <dbl> <dbl>    <dbl>
#> 1     0.175  33.7     29.7