boot_tidiers {broom}	R Documentation

Tidying methods for bootstrap computations

Description

Tidying methods for "boot" objects from the "boot" package.

Usage

## S3 method for class 'boot'
tidy(x, conf.int = FALSE, conf.level = 0.95,
  conf.method = "perc", ...)

Arguments

x
boot object

conf.int
whether to include a confidence interval

conf.level
confidence level for CI

conf.method
method for computing confidence intervals (see boot.ci)

...
extra arguments (not used)

Value

The tidy method returns a data frame with one row per bootstrapped statistic that was calculated, and the following columns:

term
Name of the computed statistic, if present

statistic
The original values of the statistic

bias
The bias of the original statistic value

std.error
Standard error of the statistic

If weights were provided to the boot function, an estimate column is included showing the weighted bootstrap estimate, and the standard error is of that estimate.

If there are no original statistics in the "boot" object, such as with a call to tsboot with orig.t = FALSE, the original and statistic columns are omitted, and only estimate and std.error columns shown.

Examples

if (require("boot")) {
   clotting <- data.frame(
          u = c(5,10,15,20,30,40,60,80,100),
          lot1 = c(118,58,42,35,27,25,21,19,18),
          lot2 = c(69,35,26,21,18,16,13,12,12))

   g1 <- glm(lot2 ~ log(u), data = clotting, family = Gamma)
   
   bootfun <- function(d, i) {
      coef(update(g1, data= d[i,]))
   }
   bootres <- boot(clotting, bootfun, R = 999)
   tidy(g1, conf.int=TRUE)
   tidy(bootres, conf.int=TRUE)
}

[Package broom version 0.4.5 Index]