Tidying method for a bootstrapped temporal exponential random graph model

This method tidies the coefficients of a bootstrapped temporal exponential random graph model estimated with the xergm. It simply returns the coefficients and their confidence intervals.

# S3 method for btergm
tidy(x, conf.level = 0.95, exponentiate = FALSE,
  quick = FALSE, ...)

Arguments

x
A btergm::btergm() object.

conf.level
confidence level of the bootstrapped interval

exponentiate
whether to exponentiate the coefficient estimates and confidence intervals

quick
whether to compute a smaller and faster version, containing only the term and estimate columns.

...
extra arguments (currently not used)

Value

A data.frame without rownames.

tidy.btergm returns one row for each coefficient, with four columns:

term

The term in the model being estimated and tested

estimate

The estimated coefficient

conf.low

The lower bound of the confidence interval

conf.high

The lower bound of the confidence interval

Details

There is no augment or glance method for ergm objects.

See also

Examples

if (require("xergm")) {
    set.seed(1)
    # Using the same simulated example as the xergm package
    # Create 10 random networks with 10 actors
    networks <- list()
    for(i in 1:10){
        mat <- matrix(rbinom(100, 1, .25), nrow = 10, ncol = 10)
        diag(mat) <- 0
        nw <- network::network(mat)
        networks[[i]] <- nw
    }
    # Create 10 matrices as covariates
    covariates <- list()
    for (i in 1:10) {
        mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
        covariates[[i]] <- mat
    }
    # Fit a model where the propensity to form ties depends
    # on the edge covariates, controlling for the number of
    # in-stars
    suppressWarnings(btfit <- btergm(networks ~ edges + istar(2) +
                       edgecov(covariates), R = 100))

    # Show terms, coefficient estimates and errors
    tidy(btfit)

    # Show coefficients as odds ratios with a 99% CI
    tidy(btfit, exponentiate = TRUE, conf.level = 0.99)
}

#> Loading required package: xergm

#> Loading required package: xergm.common

#> Loading required package: ergm