durbinWatsonTest_tidiers {broom}	R Documentation

Tidy/glance a(n) durbinWatsonTest object

Description

For models that have only a single component, the tidy() and glance() methods are identical. Please see the documentation for both of those methods.

Usage

## S3 method for class 'durbinWatsonTest'
tidy(x, ...)

## S3 method for class 'durbinWatsonTest'
glance(x, ...)

Arguments

x
An object of class durbinWatsonTest created by a call to car::durbinWatsonTest().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

alternative
Alternative hypothesis (character).

autocorrelation
Autocorrelation.

p.value
The two-sided p-value associated with the observed statistic.

statistic
Test statistic for Durbin-Watson test.

method
Always 'Durbin-Watson Test'.

See Also

tidy(), glance(), car::durbinWatsonTest()

Other car tidiers: leveneTest_tidiers

Examples

if (requireNamespace("car", quietly = TRUE)) {

dw <- car::durbinWatsonTest(lm(mpg ~ wt, data = mtcars))
tidy(dw)
glance(dw) # same output for all durbinWatsonTests

}

[Package broom version 0.7.12 Index]