Tidy/glance a(n) durbinWatsonTest object

For models that have only a single component, the tidy() and glance() methods are identical. Please see the documentation for both of those methods.

# S3 method for durbinWatsonTest
tidy(x, ...)

# S3 method for durbinWatsonTest
glance(x, ...)

Arguments

x
An object of class durbinWatsonTest created by a call to car::durbinWatsonTest().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

alternative

Alternative hypothesis (character).

autocorrelation

Autocorrelation.

p.value

The two-sided p-value associated with the observed statistic.

statistic

Test statistic for Durbin-Watson test.

method

Always `Durbin-Watson Test`.

Examples

dw <- car::durbinWatsonTest(lm(mpg ~ wt, data = mtcars))
tidy(dw)

#> # A tibble: 1 x 5
#>   statistic p.value autocorrelation method             alternative
#>       <dbl>   <dbl>           <dbl> <chr>              <chr>      
#> 1      1.25   0.014           0.363 Durbin-Watson Test two.sided  

glance(dw) # same output for all durbinWatsonTests

#> # A tibble: 1 x 5
#>   statistic p.value autocorrelation method             alternative
#>       <dbl>   <dbl>           <dbl> <chr>              <chr>      
#> 1      1.25   0.014           0.363 Durbin-Watson Test two.sided