finish_glance {broom}	R Documentation

(Deprecated) Add logLik, AIC, BIC, and other common measurements to a glance of a prediction

Description

This function is now deprecated in favor of using custom logic and the appropriate nobs() method.

Usage

finish_glance(ret, x)

Arguments

ret
a one-row data frame (a partially complete glance)

x
the prediction model

Value

a one-row data frame with additional columns added, such as

logLik
log likelihoods

AIC
Akaike Information Criterion

BIC
Bayesian Information Criterion

deviance
deviance

df.residual
residual degrees of freedom

See Also

Other deprecated: bootstrap(), confint_tidy(), data.frame_tidiers, fix_data_frame(), summary_tidiers, tidy.density(), tidy.dist(), tidy.ftable(), tidy.numeric()

[Package broom version 0.7.12 Index]