(Deprecated) Add logLik, AIC, BIC, and other common measurements to a glance of a prediction

This function is now deprecated in favor of using custom logic and the appropriate nobs() method.

finish_glance(ret, x)

Arguments

ret
a one-row data frame (a partially complete glance)

x
the prediction model

Value

a one-row data frame with additional columns added, such as

logLik

log likelihoods

AIC

Akaike Information Criterion

BIC

Bayesian Information Criterion

deviance

deviance

df.residual

residual degrees of freedom

See also