Ensure an object is a data frame, with rownames moved into a column

This function is deprecated as of broom 0.7.0 and will be removed from a future release. Please see tibble::as_tibble.

Usage

fix_data_frame(x, newnames = NULL, newcol = "term")

Arguments

x

a data.frame or matrix

newnames

new column names, not including the rownames

newcol

the name of the new rownames column

Value

a data.frame, with rownames moved into a column and new column names assigned

See also