Tidying methods for a GARCH model (tseries package)

This tidies the result of a coefficient of the GARCH model implemented in tseries package.

# S3 method for garch
tidy(x, ...)

# S3 method for garch
glance(x, method = c("box-ljung-test", "jarque-bera-test"),
  ...)

Arguments

x
garch object

...
extra arguments (not used)

method
character which specifies the hypothesis test to be shown in glance. The garch function reports 2 hypothesis tests: Jarque-Bera to residuals and Box-Ljung to squared residuals.

Value

A data.frame with one row for each coefficient, with five columns:

term

The term in the linear model being estimated and tested

estimate

The estimated coefficient

std.error

The standard error

statistic

test statistic

p.value

p-value

A data.frame with one row, with seven columns:
statistic

Test statistic used to compute the p-value

p.value

P-value

parameter

Parameter field in the htest, typically degrees of freedom

method

Method used to compute the statistic as a string

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

Examples

if (require("tseries", quietly = TRUE)) {
    data(EuStockMarkets)
    dax <- diff(log(EuStockMarkets))[,"DAX"]
    dax.garch <- garch(dax)
    dax.garch
    tidy(dax.garch)
    glance(dax.garch)
    smp1 = window(dax, end = c(1997, frequency(dax)))
    smp2 = window(dax, start = c(1998, 1))
    dax.garch <- garch(smp1)
}

#> Error: package or namespace load failed for 'tseries' in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]):
#>  there is no package called 'zoo'