Glance at a(n) Mclust object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for Mclust
glance(x, ...)

Arguments

x
An Mclust object return from mclust::Mclust().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

BIC

Bayesian Information Criterion for the model.

df

Degrees of freedom used by the model.

logLik

The log-likelihood of the model. [stats::logLik()] may be a useful reference.

n

Needs custom info.

model

A string denoting the model type with optimal BIC

G

Number mixture components in optimal model

hypvol

If the other model contains a noise component, the value of the hypervolume parameter. Otherwise `NA`.

Examples

library(dplyr)
library(mclust)
set.seed(27)

centers <- tibble::tibble(
  cluster = factor(1:3),
  num_points = c(100, 150, 50),  # number points in each cluster
  x1 = c(5, 0, -3),              # x1 coordinate of cluster center
  x2 = c(-1, 1, -2)              # x2 coordinate of cluster center
)

points <- centers %>%
  mutate(
    x1 = purrr::map2(num_points, x1, rnorm),
    x2 = purrr::map2(num_points, x2, rnorm)
  ) %>%
  select(-num_points, -cluster) %>%
  tidyr::unnest(x1, x2)

#> Error in select(., -num_points, -cluster): unused arguments (-num_points, -cluster)

m <- mclust::Mclust(points)

#> Error in as.vector(x, mode): cannot coerce type 'closure' to vector of type 'any'

tidy(m)

#> Error in tidy(m): object 'm' not found

augment(m, points)

#> Error in augment(m, points): object 'm' not found

glance(m)

#> Error in glance(m): object 'm' not found