Glance at a(n) betareg object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

# S3 method for betareg
glance(x, ...)

Arguments

x

A betareg object produced by a call to betareg::betareg().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

See also

Value

A tibble::tibble() with exactly one row and columns:

AIC

Akaike's Information Criterion for the model.

BIC

Bayesian Information Criterion for the model.

df.null

Degrees of freedom used by the null model.

df.residual

Residual degrees of freedom.

logLik

The log-likelihood of the model. [stats::logLik()] may be a useful reference.

nobs

Number of observations used.

pseudo.r.squared

Like the R squared statistic, but for situations when the R squared statistic isn't defined.

Examples

# load libraries for models and data
library(betareg)

# load dats
data("GasolineYield", package = "betareg")

# fit model
mod <- betareg(yield ~ batch + temp, data = GasolineYield)

mod
#> 
#> Call:
#> betareg(formula = yield ~ batch + temp, data = GasolineYield)
#> 
#> Coefficients (mean model with logit link):
#> (Intercept)       batch1       batch2       batch3       batch4  
#>    -6.15957      1.72773      1.32260      1.57231      1.05971  
#>      batch5       batch6       batch7       batch8       batch9  
#>     1.13375      1.04016      0.54369      0.49590      0.38579  
#>        temp  
#>     0.01097  
#> 
#> Phi coefficients (precision model with identity link):
#> (phi)  
#> 440.3  
#> 

# summarize model fit with tidiers
tidy(mod)
#> # A tibble: 12 × 6
#>    component term        estimate  std.error statistic   p.value
#>    <chr>     <chr>          <dbl>      <dbl>     <dbl>     <dbl>
#>  1 mean      (Intercept)  -6.16     0.182       -33.8  3.44e-250
#>  2 mean      batch1        1.73     0.101        17.1  2.59e- 65
#>  3 mean      batch2        1.32     0.118        11.2  3.34e- 29
#>  4 mean      batch3        1.57     0.116        13.5  8.81e- 42
#>  5 mean      batch4        1.06     0.102        10.4  4.06e- 25
#>  6 mean      batch5        1.13     0.104        11.0  6.52e- 28
#>  7 mean      batch6        1.04     0.106         9.81 1.03e- 22
#>  8 mean      batch7        0.544    0.109         4.98 6.29e-  7
#>  9 mean      batch8        0.496    0.109         4.55 5.30e-  6
#> 10 mean      batch9        0.386    0.119         3.25 1.14e-  3
#> 11 mean      temp          0.0110   0.000413     26.6  1.26e-155
#> 12 precision (phi)       440.     110.            4.00 6.29e-  5
tidy(mod, conf.int = TRUE)
#> # A tibble: 12 × 8
#>    component term        estimate  std.error statistic   p.value conf.low
#>    <chr>     <chr>          <dbl>      <dbl>     <dbl>     <dbl>    <dbl>
#>  1 mean      (Intercept)  -6.16     0.182       -33.8  3.44e-250  -6.52  
#>  2 mean      batch1        1.73     0.101        17.1  2.59e- 65   1.53  
#>  3 mean      batch2        1.32     0.118        11.2  3.34e- 29   1.09  
#>  4 mean      batch3        1.57     0.116        13.5  8.81e- 42   1.34  
#>  5 mean      batch4        1.06     0.102        10.4  4.06e- 25   0.859 
#>  6 mean      batch5        1.13     0.104        11.0  6.52e- 28   0.931 
#>  7 mean      batch6        1.04     0.106         9.81 1.03e- 22   0.832 
#>  8 mean      batch7        0.544    0.109         4.98 6.29e-  7   0.330 
#>  9 mean      batch8        0.496    0.109         4.55 5.30e-  6   0.282 
#> 10 mean      batch9        0.386    0.119         3.25 1.14e-  3   0.153 
#> 11 mean      temp          0.0110   0.000413     26.6  1.26e-155   0.0102
#> 12 precision (phi)       440.     110.            4.00 6.29e-  5 225.    
#> # … with 1 more variable: conf.high <dbl>
tidy(mod, conf.int = TRUE, conf.level = .99)
#> # A tibble: 12 × 8
#>    component term        estimate  std.error statistic   p.value  conf.low
#>    <chr>     <chr>          <dbl>      <dbl>     <dbl>     <dbl>     <dbl>
#>  1 mean      (Intercept)  -6.16     0.182       -33.8  3.44e-250  -6.63   
#>  2 mean      batch1        1.73     0.101        17.1  2.59e- 65   1.47   
#>  3 mean      batch2        1.32     0.118        11.2  3.34e- 29   1.02   
#>  4 mean      batch3        1.57     0.116        13.5  8.81e- 42   1.27   
#>  5 mean      batch4        1.06     0.102        10.4  4.06e- 25   0.796  
#>  6 mean      batch5        1.13     0.104        11.0  6.52e- 28   0.867  
#>  7 mean      batch6        1.04     0.106         9.81 1.03e- 22   0.767  
#>  8 mean      batch7        0.544    0.109         4.98 6.29e-  7   0.263  
#>  9 mean      batch8        0.496    0.109         4.55 5.30e-  6   0.215  
#> 10 mean      batch9        0.386    0.119         3.25 1.14e-  3   0.0803 
#> 11 mean      temp          0.0110   0.000413     26.6  1.26e-155   0.00990
#> 12 precision (phi)       440.     110.            4.00 6.29e-  5 157.     
#> # … with 1 more variable: conf.high <dbl>

augment(mod)
#> # A tibble: 32 × 6
#>    yield batch  temp .fitted .resid   .cooksd
#>    <dbl> <fct> <dbl>   <dbl>  <dbl>     <dbl>
#>  1 0.122 1       205  0.101   1.59  0.0791   
#>  2 0.223 1       275  0.195   1.66  0.0917   
#>  3 0.347 1       345  0.343   0.211 0.00155  
#>  4 0.457 1       407  0.508  -2.88  0.606    
#>  5 0.08  2       218  0.0797  0.109 0.0000168
#>  6 0.131 2       273  0.137  -0.365 0.00731  
#>  7 0.266 2       347  0.263   0.260 0.00523  
#>  8 0.074 3       212  0.0943 -1.77  0.0805   
#>  9 0.182 3       272  0.167   1.02  0.0441   
#> 10 0.304 3       340  0.298   0.446 0.0170   
#> # … with 22 more rows

glance(mod)
#> # A tibble: 1 × 7
#>   pseudo.r.squared df.null logLik   AIC   BIC df.residual  nobs
#>              <dbl>   <dbl>  <dbl> <dbl> <dbl>       <int> <int>
#> 1            0.962      30   84.8 -146. -128.          20    32