glance.crr {broom}	R Documentation

Glance at a(n) crr object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'crr'
glance(x, ...)

Arguments

x
A crr object returned from cmprsk::crr().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

converged
Logical indicating if the model fitting procedure was succesful and converged.

df
Degrees of freedom used by the model.

logLik
The log-likelihood of the model. [stats::logLik()] may be a useful reference.

nobs
Number of observations used.

statistic
Test statistic.

See Also

glance(), cmprsk::crr()

Other cmprsk tidiers: tidy.crr()

Examples

if (requireNamespace("cmprsk", quietly = TRUE)) {

library(cmprsk)
lrf_time <- rexp(100) #time to loco-regional failure (lrf)
lrf_event <- sample(0:2, 100, replace = TRUE) 
trt <- sample(0:1, 100, replace = TRUE)
strt <- sample(1:2, 100, replace = TRUE)
x <- crr(lrf_time, lrf_event, cbind(trt, strt))
tidy(x, conf.int = TRUE)
glance(x)

}

[Package broom version 0.7.12 Index]