glance.fixest {broom}	R Documentation

Glance at a(n) fixest object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'fixest'
glance(x, ...)

Arguments

x
A fixest object returned from any of the fixest estimators

...
Additional arguments passed to summary and confint. Important arguments are se and cluster. Other arguments are dof, exact_dof, forceCovariance, and keepBounded. See summary.fixest.

Value

A tibble::tibble() with exactly one row and columns:

adj.r.squared
Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.

AIC
Akaike's Information Criterion for the model.

BIC
Bayesian Information Criterion for the model.

logLik
The log-likelihood of the model. [stats::logLik()] may be a useful reference.

nobs
Number of observations used.

pseudo.r.squared
Like the R squared statistic, but for situations when the R squared statistic isn't defined.

r.squared
R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

sigma
Estimated standard error of the residuals.

within.r.squared
R squared within fixed-effect groups.

Note

All columns listed below will be returned, but some will be NA, depending on the type of model estimated. sigma, r.squared, adj.r.squared, and within.r.squared will be NA for any model other than feols. pseudo.r.squared will be NA for feols.

Examples

if (requireNamespace("fixest", quietly = TRUE)) {

library(fixest)

gravity <- feols(log(Euros) ~ log(dist_km) | Origin + Destination + Product + Year, trade)

tidy(gravity)
glance(gravity)
augment(gravity, trade)

## To get robust or clustered SEs, users can either:
# 1) Or, specify the arguments directly in the tidy() call

tidy(gravity, conf.int = TRUE, cluster = c("Product", "Year"))

tidy(gravity, conf.int = TRUE, se = "threeway")

# 2) Feed tidy() a summary.fixest object that has already accepted these arguments

gravity_summ <- summary(gravity, cluster = c("Product", "Year"))
tidy(gravity_summ, conf.int = TRUE)

# Approach (1) is preferred.

}

[Package broom version 0.7.12 Index]