Glance at a(n) ivreg object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

# S3 method for ivreg
glance(x, diagnostics = FALSE, ...)

Arguments

x
An ivreg object created by a call to AER::ivreg().

diagnostics
Logical indicating whether to include statistics and p-values for Sargan, Wu-Hausman and weak instrument tests. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble with columns

r.squared

The percent of variance explained by the model

adj.r.squared

r.squared adjusted based on the degrees of freedom

sigma

The square root of the estimated residual variance

statistic

Wald test statistic

p.value

p-value from the Wald test

df

Degrees of freedom used by the coefficients

df.residual

residual degrees of freedom

If diagnostics = TRUE, will also return the following columns:
statistic.Sargan

Statistic for Sargan test

p.value.Sargan

P-value for Sargan test

statistic.Wu.Hausman

Statistic for Wu-Hausman test

p.value.Wu.Hausman

P-value for Wu-Hausman test

statistic.weakinst

Statistic for Wu-Hausman test

p.value.weakinst

P-value for weak instruments test

See also

Examples

library(AER)

data("CigarettesSW", package = "AER")
ivr <- ivreg(
  log(packs) ~ income | population,
  data = CigarettesSW,
  subset = year == "1995"
)

summary(ivr)

#> 
#> Call:
#> ivreg(formula = log(packs) ~ income | population, data = CigarettesSW, 
#>     subset = year == "1995")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.69305 -0.12941 -0.02257  0.11723  0.58184 
#> 
#> Coefficients:
#>               Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  4.612e+00  4.454e-02 103.549   <2e-16 ***
#> income      -5.705e-10  2.334e-10  -2.445   0.0184 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 0.2293 on 46 degrees of freedom
#> Multiple R-Squared: 0.1308,	Adjusted R-squared: 0.1119 
#> Wald test: 5.976 on 1 and 46 DF,  p-value: 0.01839 
#> 

tidy(ivr)

#> # A tibble: 2 x 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)  4.61e+ 0  4.45e- 2    104.   3.74e-56
#> 2 income      -5.71e-10  2.33e-10     -2.44 1.84e- 2

tidy(ivr, conf.int = TRUE)

#> # A tibble: 2 x 7
#>   term         estimate std.error statistic  p.value conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 (Intercept)  4.61e+ 0  4.45e- 2    104.   3.74e-56  4.52e+0  4.70e+ 0
#> 2 income      -5.71e-10  2.33e-10     -2.44 1.84e- 2 -1.03e-9 -1.13e-10

tidy(ivr, conf.int = TRUE, exponentiate = TRUE)

#> Warning: Exponentiating coefficients, but model did not use a log or logit link function.

#> # A tibble: 2 x 7
#>   term        estimate std.error statistic  p.value conf.low conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 (Intercept)  101.     4.45e- 2    104.   3.74e-56   92.2     110.   
#> 2 income         1.000  2.33e-10     -2.44 1.84e- 2    1.000     1.000

augment(ivr)

#> # A tibble: 48 x 6
#>    .rownames log.packs.    income population .fitted  .resid
#>    <chr>          <dbl>     <dbl>      <dbl>   <dbl>   <dbl>
#>  1 49              4.62  83903280    4262731    4.56  0.0522
#>  2 50              4.71  45995496    2480121    4.59  0.124 
#>  3 51              4.28  88870496    4306908    4.56 -0.285 
#>  4 52              4.04 771470144   31493524    4.17 -0.131 
#>  5 53              4.41  92946544    3738061    4.56 -0.145 
#>  6 54              4.38 104315120    3265293    4.55 -0.177 
#>  7 55              4.82  18237436     718265    4.60  0.223 
#>  8 56              4.53 333525344   14185403    4.42  0.112 
#>  9 57              4.58 159800448    7188538    4.52  0.0591
#> 10 58              4.53  60170928    2840860    4.58 -0.0512
#> # ... with 38 more rows

glance(ivr)

#> # A tibble: 1 x 7
#>   r.squared adj.r.squared sigma statistic p.value    df df.residual
#>       <dbl>         <dbl> <dbl>     <dbl>   <dbl> <int>       <int>
#> 1     0.131         0.112 0.229      5.98  0.0184     2          46