glance.muhaz {broom}	R Documentation

Glance at a(n) muhaz object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'muhaz'
glance(x, ...)

Arguments

x
A muhaz object returned by muhaz::muhaz().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

max.hazard
Maximal estimated hazard.

max.time
The maximum observed event or censoring time.

min.hazard
Minimal estimated hazard.

min.time
The minimum observed event or censoring time.

nobs
Number of observations used.

See Also

glance(), muhaz::muhaz()

Other muhaz tidiers: tidy.muhaz()

Examples

if (requireNamespace("muhaz", quietly = TRUE)) {
  if (requireNamespace("survival", quietly = TRUE)) {

library(muhaz)
library(survival)

x <- muhaz(ovarian$futime, ovarian$fustat)
tidy(x)
glance(x)

  }
}

[Package broom version 0.7.12 Index]