Glance at a(n) multinom object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

# S3 method for multinom
glance(x, ...)

Arguments

x

A multinom object returned from nnet::multinom().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with exactly one row and columns:

AIC

Akaike's Information Criterion for the model.

deviance

Deviance of the model.

edf

The effective degrees of freedom.

nobs

Number of observations used.

Examples

if (requireNamespace("nnet", quietly = TRUE)) {

library(nnet)
library(MASS)

example(birthwt)
bwt.mu <- multinom(low ~ ., bwt)
tidy(bwt.mu)
glance(bwt.mu)

#* This model is a truly terrible model
#* but it should show you what the output looks
#* like in a multinomial logistic regression

fit.gear <- multinom(gear ~ mpg + factor(am), data = mtcars)
tidy(fit.gear)
glance(fit.gear)

}
#> 
#> brthwt> bwt <- with(birthwt, {
#> brthwt+ race <- factor(race, labels = c("white", "black", "other"))
#> brthwt+ ptd <- factor(ptl > 0)
#> brthwt+ ftv <- factor(ftv)
#> brthwt+ levels(ftv)[-(1:2)] <- "2+"
#> brthwt+ data.frame(low = factor(low), age, lwt, race, smoke = (smoke > 0),
#> brthwt+            ptd, ht = (ht > 0), ui = (ui > 0), ftv)
#> brthwt+ })
#> 
#> brthwt> options(contrasts = c("contr.treatment", "contr.poly"))
#> 
#> brthwt> glm(low ~ ., binomial, bwt)
#> 
#> Call:  glm(formula = low ~ ., family = binomial, data = bwt)
#> 
#> Coefficients:
#> (Intercept)          age          lwt    raceblack    raceother    smokeTRUE  
#>     0.82302     -0.03723     -0.01565      1.19241      0.74068      0.75553  
#>     ptdTRUE       htTRUE       uiTRUE         ftv1        ftv2+  
#>     1.34376      1.91317      0.68020     -0.43638      0.17901  
#> 
#> Degrees of Freedom: 188 Total (i.e. Null);  178 Residual
#> Null Deviance:	    234.7 
#> Residual Deviance: 195.5 	AIC: 217.5
#> # weights:  12 (11 variable)
#> initial  value 131.004817 
#> iter  10 value 98.029803
#> final  value 97.737759 
#> converged
#> # weights:  12 (6 variable)
#> initial  value 35.155593 
#> iter  10 value 14.156582
#> iter  20 value 14.031881
#> iter  30 value 14.025659
#> iter  40 value 14.021414
#> iter  50 value 14.019824
#> iter  60 value 14.019278
#> iter  70 value 14.018601
#> iter  80 value 14.018282
#> iter  80 value 14.018282
#> iter  90 value 14.017126
#> final  value 14.015374 
#> converged
#> # A tibble: 1 × 4
#>     edf deviance   AIC  nobs
#>   <dbl>    <dbl> <dbl> <int>
#> 1     6     28.0  40.0    32