glance.orcutt {broom}	R Documentation

Glance at a(n) orcutt object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'orcutt'
glance(x, ...)

Arguments

x
An orcutt object returned from orcutt::cochrane.orcutt().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

adj.r.squared
Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.

dw.original
Durbin-Watson statistic of original fit.

dw.transformed
Durbin-Watson statistic of transformed fit.

nobs
Number of observations used.

number.interaction
Number of interactions.

p.value.original
P-value of original Durbin-Watson statistic.

p.value.transformed
P-value of autocorrelation after transformation.

r.squared
R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

rho
Spearman's rho autocorrelation

See Also

glance(), orcutt::cochrane.orcutt()

Other orcutt tidiers: tidy.orcutt()

Examples

if (requireNamespace("orcutt", quietly = TRUE)) {

library(orcutt)

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

co <- cochrane.orcutt(reg)
co

tidy(co)
glance(co)

}

[Package broom version 0.7.12 Index]