glance.orcutt {broom}	R Documentation

Glance at a(n) orcutt object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Usage

## S3 method for class 'orcutt'
glance(x, ...)

Arguments

x
An orcutt object returned from orcutt::cochrane.orcutt().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

r.squared
R-squared

adj.r.squared
Adjusted R-squared

rho
Spearman's rho autocorrelation

number.interaction
Number of interactions

dw.original
Durbin-Watson statistic of original fit

p.value.original
P-value of original Durbin-Watson statistic

dw.transformed
Durbin-Watson statistic of transformed fit

p.value.transformed
P-value of autocorrelation after transformation

See Also

glance(), orcutt::cochrane.orcutt()

Other orcutt tidiers: tidy.orcutt

[Package broom version 0.5.0 Index]