glance.plm {broom}	R Documentation

Glance at a(n) plm object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'plm'
glance(x, ...)

Arguments

x
A plm objected returned by plm::plm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with exactly one row and columns:

adj.r.squared
Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.

deviance
Deviance of the model.

df.residual
Residual degrees of freedom.

nobs
Number of observations used.

p.value
P-value corresponding to the test statistic.

r.squared
R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.

statistic
F-statistic

See Also

glance(), plm::plm()

Other plm tidiers: augment.plm(), tidy.plm()

Examples

if (requireNamespace("plm", quietly = TRUE)) {

library(plm)

data("Produc", package = "plm")
zz <- plm(log(gsp) ~ log(pcap) + log(pc) + log(emp) + unemp,
  data = Produc, index = c("state", "year")
)

summary(zz)

tidy(zz)
tidy(zz, conf.int = TRUE)
tidy(zz, conf.int = TRUE, conf.level = 0.9)

augment(zz)
glance(zz)

}

[Package broom version 0.7.12 Index]