glance.rlm {broom}	R Documentation

Glance at a(n) rlm object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

## S3 method for class 'rlm'
glance(x, ...)

Arguments

x
An rlm object returned by MASS::rlm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

Value

A tibble::tibble() with exactly one row and columns:

AIC
Akaike's Information Criterion for the model.

BIC
Bayesian Information Criterion for the model.

converged
Logical indicating if the model fitting procedure was succesful and converged.

deviance
Deviance of the model.

logLik
The log-likelihood of the model. [stats::logLik()] may be a useful reference.

nobs
Number of observations used.

sigma
Estimated standard error of the residuals.

See Also

glance(), MASS::rlm()

Other rlm tidiers: augment.rlm(), tidy.rlm()

Examples

# load libraries for models and data
library(MASS)

# fit model
r <- rlm(stack.loss ~ ., stackloss)

# summarize model fit with tidiers
tidy(r)
augment(r)
glance(r)

[Package broom version 1.0.5 Index]