Glance at a(n) rq object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

# S3 method for rq
glance(x, ...)

Arguments

x

An rq object returned from quantreg::rq().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

Only models with a single tau value may be passed. For multiple values, please use a purrr::map() workflow instead, e.g.

taus %>%
  map(function(tau_val) rq(y ~ x, tau = tau_val)) %>%
  map_dfr(glance)

See also

Value

A tibble::tibble() with exactly one row and columns:

AIC

Akaike's Information Criterion for the model.

BIC

Bayesian Information Criterion for the model.

df.residual

Residual degrees of freedom.

logLik

The log-likelihood of the model. [stats::logLik()] may be a useful reference.

tau

Quantile.

Examples

# feel free to ignore the following line—it allows {broom} to supply 
# examples without requiring the model-supplying package to be installed.
if (requireNamespace("quantreg", quietly = TRUE)) {

# load modeling library and data
library(quantreg)

data(stackloss)

# median (l1) regression fit for the stackloss data. 
mod1 <- rq(stack.loss ~ stack.x, .5)  

# weighted sample median
mod2 <- rq(rnorm(50) ~ 1, weights = runif(50))

# summarize model fit with tidiers
tidy(mod1)
glance(mod1)
augment(mod1)

tidy(mod2)
glance(mod2)
augment(mod2)

# varying tau to generate an rqs object
mod3 <- rq(stack.loss ~ stack.x, tau = c(.25, .5))

tidy(mod3)
augment(mod3)

# glance cannot handle rqs objects like `mod3`--use a purrr 
# `map`-based workflow instead

}
#> # A tibble: 42 × 5
#>    stack.loss stack.x[,"Air.Flo… [,"Water.Temp"] [,"Acid.Conc."] .tau     .resid
#>         <dbl>              <dbl>           <dbl>           <dbl> <chr>     <dbl>
#>  1         42                 80              27              89 0.25   1.10e+ 1
#>  2         42                 80              27              89 0.5    5.06e+ 0
#>  3         37                 80              27              88 0.25   6.00e+ 0
#>  4         37                 80              27              88 0.5   -1.42e-14
#>  5         37                 75              25              90 0.25   1.05e+ 1
#>  6         37                 75              25              90 0.5    5.43e+ 0
#>  7         28                 62              24              87 0.25   9.00e+ 0
#>  8         28                 62              24              87 0.5    7.63e+ 0
#>  9         18                 62              22              87 0.25   1.00e+ 0
#> 10         18                 62              22              87 0.5   -1.22e+ 0
#> # … with 32 more rows, and 1 more variable: .fitted <dbl>