glance.speedlm {broom}	R Documentation

Glance at a(n) speedlm object

Description

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modelling function. This includes the name of the modelling function or any arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Usage

## S3 method for class 'speedlm'
glance(x, ...)

Arguments

x
A speedlm object returned from speedglm::speedlm().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

r.squared
The percent of variance explained by the model

adj.r.squared
r.squared adjusted based on the degrees of freedom

statistic
F-statistic

p.value
p-value from the F test, describing whether the full regression is significant

df
Degrees of freedom used by the coefficients

logLik
the data's log-likelihood under the model

AIC
the Akaike Information Criterion

BIC
the Bayesian Information Criterion

deviance
deviance

df.residual
residual degrees of freedom

See Also

speedglm::speedlm()

Other speedlm tidiers: augment.speedlm, tidy.speedlm

[Package broom version 0.5.0 Index]