Glance at a(n) survexp object

Glance accepts a model object and returns a tibble::tibble() with exactly one row of model summaries. The summaries are typically goodness of fit measures, p-values for hypothesis tests on residuals, or model convergence information.

Glance never returns information from the original call to the modeling function. This includes the name of the modeling function or any arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations to appropriate methods and gathers the results together. Sometimes a goodness of fit measure will be undefined. In these cases the measure will be reported as NA.

Glance returns the same number of columns regardless of whether the model matrix is rank-deficient or not. If so, entries in columns that no longer have a well-defined value are filled in with an NA of the appropriate type.

Usage

# S3 method for class 'survexp'
glance(x, ...)

Arguments

x

An survexp object returned from survival::survexp().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

See also

Value

A tibble::tibble() with exactly one row and columns:

n.max

Maximum number of subjects at risk.

n.start

Initial number of subjects at risk.

timepoints

Number of timepoints.

Examples

# load libraries for models and data
library(survival)

# fit model
sexpfit <- survexp(
  futime ~ 1,
  rmap = list(
    sex = "male",
    year = accept.dt,
    age = (accept.dt - birth.dt)
  ),
  method = "conditional",
  data = jasa
)

# summarize model fit with tidiers
tidy(sexpfit)
#> # A tibble: 88 × 3
#>     time estimate n.risk
#>    <dbl>    <dbl>  <int>
#>  1     0    1        102
#>  2     1    1.000    102
#>  3     2    1.000     99
#>  4     4    1.000     96
#>  5     5    1.000     94
#>  6     7    1.000     92
#>  7     8    1.000     91
#>  8    10    1.000     90
#>  9    11    1.000     89
#> 10    15    1.000     88
#> # ℹ 78 more rows
glance(sexpfit)
#> # A tibble: 1 × 3
#>   n.max n.start timepoints
#>   <int>   <int>      <int>
#> 1   102     102         88