glance_optim {broom}	R Documentation

Tidy a(n) optim object masquerading as list

Description

Broom tidies a number of lists that are effectively S3 objects without a class attribute. For example, stats::optim(), svd() and akima::interp() produce consistent output, but because they do not have a class attribute, they cannot be handled by S3 dispatch.

These functions look at the elements of a list and determine if there is an appropriate tidying method to apply to the list. Those tidiers are themselves are implemented as functions of the form tidy_<function> or glance_<function> and are not exported (but they are documented!).

If no appropriate tidying method is found, throws an error.

Usage

glance_optim(x, ...)

Arguments

x
A list returned from stats::optim().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

value
minimized or maximized output value

function.count
number of calls to fn

gradient.count
number of calls to gr

convergence
convergence code representing the error state

See Also

glance(), optim()

Other list tidiers: list_tidiers, tidy_irlba, tidy_optim, tidy_svd, tidy_xyz

[Package broom version 0.5.0 Index]