Tidiers for LASSO or elasticnet regularized fits

Tidying methods for regularized fits produced by glmnet, summarizing the estimates across values of the penalty parameter lambda.

# S3 method for glmnet
tidy(x, return_zeros = FALSE, ...)

# S3 method for glmnet
glance(x, ...)

Arguments

x
a "glmnet" object

return_zeros
should coefficients with value zero be included in the results (default FALSE)?

...
extra arguments (not used)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy produces a data.frame with one row per combination of coefficient (including the intercept) and value of lambda for which the estimate is nonzero, with the columns:

term

coefficient name (V1...VN by default, along with "(Intercept)")

step

which step of lambda choices was used

estimate

estimate of coefficient

lambda

value of penalty parameter lambda

dev.ratio

fraction of null deviance explained at each value of lambda

glance returns a one-row data.frame with the values
nulldev

null deviance

npasses

total passes over the data across all lambda values

Details

Note that while this representation of GLMs is much easier to plot and combine than the default structure, it is also much more memory-intensive. Do not use for extremely large, sparse matrices.

No augment method is yet provided even though the model produces predictions, because the input data is not tidy (it is a matrix that may be very wide) and therefore combining predictions with it is not logical. Furthermore, predictions make sense only with a specific choice of lambda.

Examples

if (require("glmnet", quietly = TRUE)) {
    set.seed(2014)
    x <- matrix(rnorm(100*20),100,20)
    y <- rnorm(100)
    fit1 <- glmnet(x,y)

    head(tidy(fit1))
    glance(fit1)

    library(dplyr)
    library(ggplot2)

    tidied <- tidy(fit1) %>% filter(term != "(Intercept)")

    ggplot(tidied, aes(step, estimate, group = term)) + geom_line()
    ggplot(tidied, aes(lambda, estimate, group = term)) +
        geom_line() + scale_x_log10()

    ggplot(tidied, aes(lambda, dev.ratio)) + geom_line()

    # works for other types of regressions as well, such as logistic
    g2 <- sample(1:2, 100, replace=TRUE)
    fit2 <- glmnet(x, g2, family="binomial")
    head(tidy(fit2))
}

#> # A tibble: 6 x 5
#>   term         step estimate lambda dev.ratio
#>   <chr>       <dbl>    <dbl>  <dbl>     <dbl>
#> 1 (Intercept)     1  -0.241   0.177  1.78e-15
#> 2 (Intercept)     2  -0.241   0.161  1.57e- 2
#> 3 V13             2   0.0620  0.161  1.57e- 2
#> 4 (Intercept)     3  -0.241   0.147  2.88e- 2
#> 5 V13             3   0.119   0.147  2.88e- 2
#> 6 (Intercept)     4  -0.241   0.134  3.98e- 2