Tidying methods for an htest object

Tidies hypothesis test objects, such as those from cor.test, t.test, and wilcox.test, into a one-row data frame.

# S3 method for htest
tidy(x, ...)

# S3 method for htest
glance(x, ...)

Arguments

x
An object of class "htest"

...
extra arguments (not used)

Value

Both tidy and glance return the same output, a one-row data frame with one or more of the following columns:

estimate

Estimate of the effect size

statistic

Test statistic used to compute the p-value

p.value

P-value

parameter

Parameter field in the htest, typically degrees of freedom

conf.low

Lower bound on a confidence interval

conf.high

Upper bound on a confidence interval

estimate1

Sometimes two estimates are computed, such as in a two-sample t-test

estimate2

Sometimes two estimates are computed, such as in a two-sample t-test

method

Method used to compute the statistic as a string

alternative

Alternative hypothesis as a string

Which columns are included depends on the hypothesis test used.
Details

No augment method is provided for "htest", since there is no sense in which a hypothesis test generates one value for each observation.

Examples

tt <- t.test(rnorm(10))
tidy(tt)

#>     estimate  statistic   p.value parameter  conf.low conf.high
#> 1 -0.2178381 -0.5332314 0.6067881         9 -1.141985 0.7063086
#>              method alternative
#> 1 One Sample t-test   two.sided

glance(tt)  # same output for all htests

#>     estimate  statistic   p.value parameter  conf.low conf.high
#> 1 -0.2178381 -0.5332314 0.6067881         9 -1.141985 0.7063086
#>              method alternative
#> 1 One Sample t-test   two.sided

tt <- t.test(mpg ~ am, data = mtcars)
tidy(tt)

#>    estimate estimate1 estimate2 statistic     p.value parameter  conf.low
#> 1 -7.244939  17.14737  24.39231 -3.767123 0.001373638  18.33225 -11.28019
#>   conf.high                  method alternative
#> 1 -3.209684 Welch Two Sample t-test   two.sided

wt <- wilcox.test(mpg ~ am, data = mtcars)

#> Warning: cannot compute exact p-value with ties

tidy(wt)

#>   statistic     p.value                                            method
#> 1        42 0.001871391 Wilcoxon rank sum test with continuity correction
#>   alternative
#> 1   two.sided

ct <- cor.test(mtcars$wt, mtcars$mpg)
tidy(ct)

#>     estimate statistic      p.value parameter   conf.low  conf.high
#> 1 -0.8676594 -9.559044 1.293959e-10        30 -0.9338264 -0.7440872
#>                                 method alternative
#> 1 Pearson's product-moment correlation   two.sided