Tidy a kernel density estimate object from the ks package

Tidy a kernel density estimate object, into a table with one row for each point in the estimated grid, and one column for each dimension (along with an estimate column with the estimated density).

# S3 method for kde
tidy(x, ...)

Arguments

x
A "ks" object from the kde package

...
Extra arguments, not used

Value

A data frame with one row for each point in the estimated grid. The result contains one column (named x1, x2, etc) for each dimension, and an estimate column containing the estimated density.

Examples

if (require("ks", quietly = TRUE)) {
  dat <- replicate(2, rnorm(100))
  k <- kde(dat)

  td <- tidy(k)
  head(td)

  library(ggplot2)
  ggplot(td, aes(x1, x2, fill = estimate)) +
    geom_tile() +
    theme_void()

  # also works with 3 dimensions
  dat3 <- replicate(3, rnorm(100))
  k3 <- kde(dat3)

  td3 <- tidy(k3)
  head(td3)
}

#> # A tibble: 6 x 4
#>     obs variable value estimate
#>   <int> <chr>    <dbl>    <dbl>
#> 1     1 x1       -4.77        0
#> 2     2 x1       -4.59        0
#> 3     3 x1       -4.41        0
#> 4     4 x1       -4.23        0
#> 5     5 x1       -4.05        0
#> 6     6 x1       -3.87        0