loess_tidiers {broom}	R Documentation

Augmenting methods for loess models

Description

This method augments the original data with information on the fitted values and residuals, and optionally the standard errors.

Usage

## S3 method for class 'loess'
augment(x, data = stats::model.frame(x), newdata, ...)

Arguments

x
A "loess" object

data
Original data, defaults to the extracting it from the model

newdata
If provided, performs predictions on the new data

...
extra arguments

Details

When the modeling was performed with na.action = "na.omit" (as is the typical default), rows with NA in the initial data are omitted entirely from the augmented data frame. When the modeling was performed with na.action = "na.exclude", one should provide the original data as a second argument, at which point the augmented data will contain those rows (typically with NAs in place of the new columns). If the original data is not provided to augment and na.action = "na.exclude", a warning is raised and the incomplete rows are dropped.

Value

When newdata is not supplied augment.loess returns one row for each observation with three columns added to the original data:

.fitted
Fitted values of model

.se.fit
Standard errors of the fitted values

.resid
Residuals of the fitted values

When newdata is supplied augment.loess returns one row for each observation with one additional column:

.fitted
Fitted values of model

.se.fit
Standard errors of the fitted values

See Also

na.action

Examples

lo <- loess(mpg ~ wt, mtcars)
augment(lo)

# with all columns of original data
augment(lo, mtcars)

# with a new dataset
augment(lo, newdata = head(mtcars))

[Package broom version 0.4.5 Index]