Tidying methods for Mclust objects

These methods summarize the results of Mclust clustering into three tidy forms. tidy describes the size, mixing probability, mean and variability of each class, augment adds the class assignments and their probabilities to the original data, and glance summarizes the model parameters of the clustering.

# S3 method for Mclust
tidy(x, ...)

# S3 method for Mclust
augment(x, data, ...)

# S3 method for Mclust
glance(x, ...)

Arguments

x
Mclust object

...
extra arguments, not used

data
Original data (required for augment)

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy returns one row per component, with

component

A factor describing the cluster from 1:k (or 0:k in presence of a noise term in x)

size

The size of each component

proportion

The mixing proportion of each component

variance

In case of one-dimensional and spherical models, the variance for each component, omitted otherwise. NA for noise component

mean

The mean for each component. In case of two- or more dimensional models, a column with the mean is added for each dimension. NA for noise component

augment returns the original data with two extra columns:
.class

The class assigned by the Mclust algorithm

.uncertainty

The uncertainty associated with the classification

glance returns a one-row data.frame with the columns
model

A character string denoting the model at which the optimal BIC occurs

n

The number of observations in the data

G

The optimal number of mixture components

BIC

The optimal BIC value

logLik

The log-likelihood corresponding to the optimal BIC

df

The number of estimated parameters

hypvol

The hypervolume parameter for the noise component if required, otherwise set to NA

See also

Examples

library(dplyr)
library(ggplot2)
library(mclust)

#> Package 'mclust' version 5.4.1
#> Type 'citation("mclust")' for citing this R package in publications.

#> 
#> Attaching package: 'mclust'

#> The following object is masked from 'package:psych':
#> 
#>     sim

set.seed(2016)
centers <- data.frame(cluster=factor(1:3), size=c(100, 150, 50),
                      x1=c(5, 0, -3), x2=c(-1, 1, -2))
points <- centers %>% group_by(cluster) %>%
 do(data.frame(x1=rnorm(.$size[1], .$x1[1]),
               x2=rnorm(.$size[1], .$x2[1]))) %>%
 ungroup()

m = Mclust(points %>% dplyr::select(x1, x2))

tidy(m)

#> # A tibble: 3 x 6
#>   component  size proportion variance mean.x1 mean.x2
#>       <int> <int>      <dbl>    <dbl>   <dbl>   <dbl>
#> 1         1   100      0.333    0.985  4.89    -0.841
#> 2         2   151      0.503    0.985  0.0665   1.01 
#> 3         3    49      0.164    0.985 -3.07    -2.02 

head(augment(m, points))

#> # A tibble: 6 x 5
#>   cluster    x1      x2 .class .uncertainty
#>   <fct>   <dbl>   <dbl> <fct>         <dbl>
#> 1 1        4.09 -0.371  1          2.44e- 4
#> 2 1        6.00 -2.31   1          5.31e-10
#> 3 1        4.94 -2.02   1          1.62e- 7
#> 4 1        5.30  0.534  1          3.57e- 6
#> 5 1        2.21 -0.675  2          4.27e- 1
#> 6 1        4.72 -0.0123 1          2.17e- 5

glance(m)

#> # A tibble: 1 x 7
#>   model     n     G    BIC logLik    df hypvol
#>   <chr> <int> <int>  <dbl>  <dbl> <dbl>  <dbl>
#> 1 EII     300     3 -2324. -1136.     9     NA