Tidying methods for mixed effects models

nlme tidiers will soon be deprecated in broom and there is no ongoing development of these functions at this time. nlme tidiers are being developed in the broom.mixed package, which is not yet on CRAN.

# S3 method for lme
tidy(x, effects = "random", ...)

# S3 method for lme
augment(x, data = x$data, newdata, ...)

# S3 method for lme
glance(x, ...)

Arguments

x
An object of class lme, such as those from lme or nlme

effects
Either "random" (default) or "fixed"

...
extra arguments (not used)

data
original data this was fitted on; if not given this will attempt to be reconstructed

newdata
new data to be used for prediction; optional

Value

All tidying methods return a data.frame without rownames. The structure depends on the method chosen.

tidy returns one row for each estimated effect, either random or fixed depending on the effects parameter. If effects = "random", it contains the columns

group

the group within which the random effect is being estimated

level

level within group

term

term being estimated

estimate

estimated coefficient

If effects="fixed", tidy returns the columns
term

fixed term being estimated

estimate

estimate of fixed effect

std.error

standard error

statistic

t-statistic

p.value

P-value computed from t-statistic

augment returns one row for each original observation, with columns (each prepended by a .) added. Included are the columns
.fitted

predicted values

.resid

residuals

.fixed

predicted values with no random effects

glance returns one row with the columns
sigma

the square root of the estimated residual variance

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

deviance

returned as NA. To quote Brian Ripley on R-help: McCullagh & Nelder (1989) would be the authoritative reference, but the 1982 first edition manages to use 'deviance' in three separate senses on one page.

Details

These methods tidy the coefficients of mixed effects models of the lme class from functions of the nlme package.

When the modeling was performed with na.action = "na.omit" (as is the typical default), rows with NA in the initial data are omitted entirely from the augmented data frame. When the modeling was performed with na.action = "na.exclude", one should provide the original data as a second argument, at which point the augmented data will contain those rows (typically with NAs in place of the new columns). If the original data is not provided to augment() and na.action = "na.exclude", a warning is raised and the incomplete rows are dropped.

See also

Examples

if (require("nlme") & require("lme4")) {
    # example regressions are from lme4 documentation, but used for nlme
    lmm1 <- lme(Reaction ~ Days, random=~ Days|Subject, sleepstudy)
    tidy(lmm1)
    tidy(lmm1, effects = "fixed")
    head(augment(lmm1, sleepstudy))
    glance(lmm1)

    startvec <- c(Asym = 200, xmid = 725, scal = 350)
    nm1 <- nlme(circumference ~ SSlogis(age, Asym, xmid, scal),
                  data = Orange,
                  fixed = Asym + xmid + scal ~1,
                  random = Asym ~1,
                  start = startvec)
    tidy(nm1)
    tidy(nm1, effects = "fixed")
    head(augment(nm1, Orange))
    glance(nm1)
}

#> Loading required package: nlme

#> 
#> Attaching package: 'nlme'

#> The following object is masked from 'package:lme4':
#> 
#>     lmList

#> The following object is masked from 'package:dplyr':
#> 
#>     collapse

#> Warning: deviance undefined for REML fit

#> # A tibble: 1 x 5
#>   sigma logLik   AIC   BIC deviance
#>   <dbl>  <dbl> <dbl> <dbl> <lgl>   
#> 1  7.85  -132.  273.  281. NA