null_tidiers {broom}	R Documentation

Tidiers for NULL inputs

Description

tidy(NULL), glance(NULL) and augment(NULL) all return an empty tibble::tibble. This empty tibble can be treated a tibble with zero rows, making it convenient to combine with other tibbles using functions like purrr::map_df() on lists of potentially NULL objects.

Usage

## S3 method for class ''NULL''
tidy(x, ...)

## S3 method for class ''NULL''
glance(x, ...)

## S3 method for class ''NULL''
augment(x, ...)

Arguments

x
The value NULL.

...
Additional arguments (not used).

Value

An empty tibble::tibble.

See Also

tibble::tibble

[Package broom version 0.7.12 Index]