Tidiers for Cochrane Orcutt object

Tidies a Cochrane Orcutt object, which estimates autocorrelation and beta coefficients in a linear fit.

# S3 method for orcutt
tidy(x, ...)

# S3 method for orcutt
glance(x, ...)

Arguments

x
An "orcutt" object returned by cochrane.orcutt

...
Extra arguments passed on to tidy.lm()

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy returns the same information as tidy.lm(), though without confidence interval options.

glance

glance returns a one-row data frame with the following columns:

r.squared

R-squared

adj.r.squared

Adjusted R-squared

rho

Spearman's rho autocorrelation

number.interaction

Number of interactions

dw.original

Durbin-Watson statistic of original fit

p.value.original

P-value of original Durbin-Watson statistic

dw.transformed

Durbin-Watson statistic of transformed fit

p.value.transformed

P-value of autocorrelation after transformation

Examples

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

#> # A tibble: 4 x 5
#>   term          estimate std.error statistic  p.value
#>   <chr>            <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)  19.8         5.94      3.33   0.00244 
#> 2 wt           -5.03        1.22     -4.11   0.000310
#> 3 qsec          0.927       0.342     2.71   0.0114  
#> 4 disp         -0.000128    0.0106   -0.0121 0.990   

if (require("orcutt", quietly = TRUE)) {
  co <- cochrane.orcutt(reg)
  co

  tidy(co)
  glance(co)
}