helper function to process a tidied ergm object

Optionally exponentiates the coefficients, and optionally adds a confidence interval, to a tidied ergm object.

process_ergm(ret, x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE)

Arguments

ret
data frame with a tidied version of a coefficient matrix

x
an "ergm" object

conf.int
whether to include a confidence interval

conf.level
confidence level of the interval, used only if conf.int=TRUE

exponentiate
whether to exponentiate the coefficient estimates and confidence intervals (typical for logistic regression)