Tidying methods for quantile regression models

These methods tidy the coefficients of a quantile regression model into a summary, augment the original data with information on the fitted values and residuals, and construct a glance of the model's statistics.

# S3 method for nlrq
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

# S3 method for nlrq
glance(x, ...)

# S3 method for nlrq
augment(x, ...)

# S3 method for rq
tidy(x, se.type = "rank", conf.int = TRUE, conf.level = 0.9,
  alpha = 1 - conf.level, ...)

# S3 method for rq
glance(x, ...)

# S3 method for rq
augment(x, data = model.frame(x), newdata = NULL, ...)

# S3 method for rqs
tidy(x, se.type = "rank", conf.int = TRUE, conf.level = 0.9,
  alpha = 1 - conf.level, ...)

# S3 method for rqs
augment(x, data = model.frame(x), newdata, ...)

Arguments

x
model object returned by rq or nlrq

conf.int
boolean; should confidence intervals be calculated, ignored if se.type = "rank"

conf.level
confidence level for intervals

…
other arguments passed on to summary.rq

se.type
Type of standard errors to calculate; see summary.rq

alpha
confidence level when se.type = "rank"; defaults to the same as conf.level although the specification is inverted

data
Original data, defaults to extracting it from the model

newdata
If provided, new data frame to use for predictions

Value

tidy.nlrq returns one row for each coefficient in the model, with five columns:

term

The term in the nonlinear model being estimated and tested

estimate

The estimated coefficient

std.error

The standard error from the linear model

statistic

t-statistic

p.value

two-sided p-value

glance.rq returns one row for each quantile (tau) with the columns:
tau

quantile estimated

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

df.residual

residual degrees of freedom

augment.rqs returns a row for each original observation with the following columns added:
.resid

Residuals

.fitted

Fitted quantiles of the model

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen. tidy.rq returns a tibble with one row for each coefficient. The columns depend upon the confidence interval method selected. glance.rq returns one row for each quantile (tau) with the columns:
tau

quantile estimated

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

df.residual

residual degrees of freedom

augment.rq returns a row for each original observation with the following columns added:
.resid

Residuals

.fitted

Fitted quantiles of the model

.tau

Quantile estimated

Depending on the arguments passed on to predict.rq via … a confidence interval is also calculated on the fitted values resulting in columns:
.conf.low

Lower confidence interval value

.conf.high

Upper confidence interval value

See predict.rq for details on additional arguments to specify confidence intervals. predict.rq does not provide confidence intervals when newdata is provided. tidy.rqs returns a data frame with one row for each coefficient at each quantile that was estimated. The columns depend upon the confidence interval method selected. augment.rqs returns a row for each original observation and each estimated quantile (tau) with the following columns added:
.resid

Residuals

.fitted

Fitted quantiles of the model

.tau

Quantile estimated

predict.rqs does not return confidence interval estimates.
Details

This simply calls augment.nls on the "nlrq" object.

If se.type = "rank" confidence intervals are calculated by summary.rq. When only a single predictor is included in the model, no confidence intervals are calculated and the confidence limits are set to NA. If se.type != 'rank' and conf.int = TRUE, confidence intervals are standard t based intervals.

Only models with a single tau value may be passed. For multiple values, please use a purrr::map() workflow instead, e.g.

taus 
  map(function(tau_val) rq(y ~ x, tau = tau_val)) 
  map_dfr(glance)