Tidy a(n) SpatialPolygonsDataFrame object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Note that the sf package now defines tidy spatial objects and is the recommend approach to spatial data. sp tidiers are likely to be deprecated in the near future in favor of sf::st_as_sf(). Development of sp tidiers has halted in broom.

# S3 method for SpatialPolygonsDataFrame
tidy(x, region = NULL, ...)

# S3 method for SpatialPolygons
tidy(x, ...)

# S3 method for Polygons
tidy(x, ...)

# S3 method for Polygon
tidy(x, ...)

# S3 method for SpatialLinesDataFrame
tidy(x, ...)

# S3 method for Lines
tidy(x, ...)

# S3 method for Line
tidy(x, ...)

Arguments

x
A SpatialPolygonsDataFrame, SpatialPolygons, Polygons, Polygon, SpatialLinesDataFrame, Lines or Line object.

region
name of variable used to split up regions

...
not used by this method