Tidying methods for a speedlm model

These methods tidy the coefficients of a "speedlm" object into a summary, augment the original data with information on the fitted values and residuals, and construct a one-row glance of the model's statistics.

# S3 method for speedlm
tidy(x, ...)

# S3 method for speedlm
glance(x, ...)

# S3 method for speedlm
augment(x, data = stats::model.frame(x), newdata = data,
  ...)

Arguments

x
speedlm object

...
extra arguments (not used)

data
data frame to augment

newdata
new data to use for predictions, optional

Value

All tidying methods return a data.frame without rownames, whose structure depends on the method chosen.

tidy.speedlm returns the tidied output of the lm with one row for each term in the formula. The columns match those in lm_tidiers.

glance.speedlm returns a one-row data.frame with the columns

r.squared

The percent of variance explained by the model

adj.r.squared

r.squared adjusted based on the degrees of freedom

statistic

F-statistic

p.value

p-value from the F test, describing whether the full regression is significant

df

Degrees of freedom used by the coefficients

logLik

the data's log-likelihood under the model

AIC

the Akaike Information Criterion

BIC

the Bayesian Information Criterion

deviance

deviance

df.residual

residual degrees of freedom

augment.speedlm returns one row for each observation, with just one column added:
.fitted

Fitted values of model

See also

Examples

if (require("speedglm", quietly = TRUE)) {
    mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars)
    tidy(mod)
    glance(mod)
    augment(mod)
}

#> 
#> Attaching package: 'speedglm'

#> The following object is masked from 'package:boot':
#> 
#>     control

#> # A tibble: 32 x 5
#>    .rownames           mpg    wt  qsec .fitted
#>    <chr>             <dbl> <dbl> <dbl>   <dbl>
#>  1 Mazda RX4          21    2.62  16.5    21.8
#>  2 Mazda RX4 Wag      21    2.88  17.0    21.0
#>  3 Datsun 710         22.8  2.32  18.6    25.3
#>  4 Hornet 4 Drive     21.4  3.22  19.4    21.6
#>  5 Hornet Sportabout  18.7  3.44  17.0    18.2
#>  6 Valiant            18.1  3.46  20.2    21.1
#>  7 Duster 360         14.3  3.57  15.8    16.4
#>  8 Merc 240D          24.4  3.19  20      22.2
#>  9 Merc 230           22.8  3.15  22.9    25.1
#> 10 Merc 280           19.2  3.44  18.3    19.4
#> # ... with 22 more rows