summary_tidiers {broom}	R Documentation

(Deprecated) Tidy summaryDefault objects

Description

Tidiers for summaryDefault objects have been deprecated as of broom 0.7.0 in favor of skimr::skim().

Usage

## S3 method for class 'summaryDefault'
tidy(x, ...)

## S3 method for class 'summaryDefault'
glance(x, ...)

Arguments

x
A summaryDefault object, created by calling summary() on a vector.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

minimum
Minimum value in original vector.

q1
First quartile of original vector.

median
Median of original vector.

mean
Mean of original vector.

q3
Third quartile of original vector.

maximum
Maximum value in original vector.

na
Number of NA values in original vector. Column present only when original vector had at least one NA entry.

See Also

Other deprecated: bootstrap(), confint_tidy(), data.frame_tidiers, finish_glance(), fix_data_frame(), tidy.density(), tidy.dist(), tidy.ftable(), tidy.numeric()

Other deprecated: bootstrap(), confint_tidy(), data.frame_tidiers, finish_glance(), fix_data_frame(), tidy.density(), tidy.dist(), tidy.ftable(), tidy.numeric()

Examples

v <- rnorm(1000)
s <- summary(v)
s

tidy(s)
glance(s)

v2 <- c(v,NA)
tidy(summary(v2))

[Package broom version 0.7.12 Index]