(Deprecated) Tidy summaryDefault objects

Tidiers for summaryDefault objects have been deprecated as of broom 0.7.0 in favor of skimr::skim().

# S3 method for summaryDefault
tidy(x, ...)

# S3 method for summaryDefault
glance(x, ...)

Arguments

x

A summaryDefault object, created by calling summary() on a vector.

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A one-row tibble::tibble with columns:

minimum

Minimum value in original vector.

q1

First quartile of original vector.

median

Median of original vector.

mean

Mean of original vector.

q3

Third quartile of original vector.

maximum

Maximum value in original vector.

na

Number of NA values in original vector. Column present only when original vector had at least one NA entry.

See also

Examples

v <- rnorm(1000)
s <- summary(v)
s
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -2.80978 -0.62832  0.00921  0.01613  0.66460  3.24104 

tidy(s)
#> Warning: `tidy.summaryDefault()` is deprecated. Please use `skimr::skim()` instead.
#> # A tibble: 1 × 6
#>   minimum     q1  median   mean    q3 maximum
#>     <dbl>  <dbl>   <dbl>  <dbl> <dbl>   <dbl>
#> 1   -2.81 -0.628 0.00921 0.0161 0.665    3.24
glance(s)
#> Warning: `tidy.summaryDefault()` is deprecated. Please use `skimr::skim()` instead.
#> # A tibble: 1 × 6
#>   minimum     q1  median   mean    q3 maximum
#>     <dbl>  <dbl>   <dbl>  <dbl> <dbl>   <dbl>
#> 1   -2.81 -0.628 0.00921 0.0161 0.665    3.24

v2 <- c(v,NA)
tidy(summary(v2))
#> Warning: `tidy.summaryDefault()` is deprecated. Please use `skimr::skim()` instead.
#> # A tibble: 1 × 7
#>   minimum     q1  median   mean    q3 maximum    na
#>     <dbl>  <dbl>   <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1   -2.81 -0.628 0.00921 0.0161 0.665    3.24     1