tidy.Gam {broom}	R Documentation

Tidy a(n) Gam object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'Gam'
tidy(x, ...)

Arguments

x
A Gam object returned from a call to gam::gam().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

Tidy gam objects created by calls to mgcv::gam() with tidy.gam().

Value

The tidied output of the parametric ANOVA for the GAM model as a tibble::tibble with one row for each term in the model.

See Also

tidy(), gam::gam(), tidy.anova(), tidy.gam()

Other gam tidiers: glance.Gam

Examples

library(gam)
g <- gam(mpg ~ s(hp, 4) + am + qsec, data = mtcars)
  
tidy(g)
glance(g)

[Package broom version 0.5.0 Index]