Tidy a(n) Kendall object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for Kendall
tidy(x, ...)

Arguments

x
A Kendall object returned from a call to Kendall::Kendall(), Kendall::MannKendall(), or Kendall::SeasonalMannKendall().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row and columns:

statistic

Kendall's tau statistic.

p.value

two-sided p-value.

kendall_score

Kendall score.

denominator

The denominator, which is tau=kendall_score/denominator.

var_kendall_score

Variance of the kendall_score.

See also

Examples

library(Kendall)

#> Error in library(Kendall): there is no package called ‘Kendall’

A <- c(2.5,2.5,2.5,2.5,5,6.5,6.5,10,10,10,10,10,14,14,14,16,17)
B <- c(1,1,1,1,2,1,1,2,1,1,1,1,1,1,2,2,2)

f_res <- Kendall(A, B)

#> Error in Kendall(A, B): could not find function "Kendall"

tidy(f_res)

#> Error in tidy(f_res): object 'f_res' not found

s_res <- MannKendall(B)

#> Error in MannKendall(B): could not find function "MannKendall"

tidy(s_res)

#> Error in tidy(s_res): object 's_res' not found

t_res <- SeasonalMannKendall(ts(A))

#> Error in SeasonalMannKendall(ts(A)): could not find function "SeasonalMannKendall"

tidy(t_res)

#> Error in tidy(t_res): object 't_res' not found