Tidy a(n) Kendall object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for Kendall
tidy(x, ...)

Arguments

x
A Kendall object returned from a call to Kendall::Kendall(), Kendall::MannKendall(), or Kendall::SeasonalMannKendall().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

kendall_score

Kendall score.

p.value

The two-sided p-value associated with the observed statistic.

var_kendall_score

Variance of the kendall_score.

statistic

Kendall's tau statistic

denominator

The denominator, which is tau=kendall_score/denominator.

Examples

library(Kendall)

A <- c(2.5, 2.5, 2.5, 2.5, 5, 6.5, 6.5, 10, 10, 10, 10, 10, 14, 14, 14, 16, 17)
B <- c(1, 1, 1, 1, 2, 1, 1, 2, 1, 1, 1, 1, 1, 1, 2, 2, 2)

f_res <- Kendall(A, B)
tidy(f_res)

#> # A tibble: 1 x 5
#>   statistic p.value kendall_score denominator var_kendall_score
#>       <dbl>   <dbl>         <dbl>       <dbl>             <dbl>
#> 1     0.408  0.0754            34        83.4              345.

s_res <- MannKendall(B)
tidy(s_res)

#> # A tibble: 1 x 5
#>   statistic p.value kendall_score denominator var_kendall_score
#>       <dbl>   <dbl>         <dbl>       <dbl>             <dbl>
#> 1     0.354   0.102            32        90.3               360

t_res <- SeasonalMannKendall(ts(A))
tidy(t_res)

#> # A tibble: 1 x 5
#>   statistic     p.value kendall_score denominator var_kendall_score
#>       <dbl>       <dbl>         <dbl>       <dbl>             <dbl>
#> 1     0.924 0.000000935           116        126.              559.