tidy.TukeyHSD {broom}	R Documentation

Tidy a(n) TukeyHSD object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'TukeyHSD'
tidy(x, ...)

Arguments

x
A TukeyHSD object return from stats::TukeyHSD().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

adj.p.value
P-value adjusted for multiple comparisons.

conf.high
Upper bound on the confidence interval for the estimate.

conf.low
Lower bound on the confidence interval for the estimate.

contrast
Levels being compared.

estimate
The estimated value of the regression term.

null.value
Value to which the estimate is compared.

term
The name of the regression term.

See Also

tidy(), stats::TukeyHSD()

Other anova tidiers: glance.aov(), tidy.anova(), tidy.aovlist(), tidy.aov(), tidy.manova()

Examples

fm1 <- aov(breaks ~ wool + tension, data = warpbreaks)
thsd <- TukeyHSD(fm1, "tension", ordered = TRUE)
tidy(thsd)

# may include comparisons on multiple terms
fm2 <- aov(mpg ~ as.factor(gear) * as.factor(cyl), data = mtcars)
tidy(TukeyHSD(fm2))

[Package broom version 0.7.12 Index]