tidy.aovlist {broom}	R Documentation

Tidy a(n) aovlist object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'aovlist'
tidy(x, ...)

Arguments

x
An aovlist objects, such as those created by stats::aov().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

Details

The term column of an ANOVA table can come with leading or trailing whitespace, which this tidying method trims.

For documentation on the tidier for car::leveneTest() output, see tidy.leveneTest()

Value

A tibble::tibble() with columns:

df
Degrees of freedom used by this term in the model.

meansq
Mean sum of squares. Equal to total sum of squares divided by degrees of freedom.

p.value
The two-sided p-value associated with the observed statistic.

statistic
The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

stratum
The error stratum.

sumsq
Sum of squares explained by this term.

term
The name of the regression term.

See Also

tidy(), stats::aov()

Other anova tidiers: glance.anova(), glance.aov(), tidy.TukeyHSD(), tidy.anova(), tidy.aov(), tidy.manova()

Examples

a <- aov(mpg ~ wt + qsec + Error(disp / am), mtcars)
tidy(a)

[Package broom version 1.0.5 Index]