Tidy a(n) boot object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for boot
tidy(x, conf.int = FALSE, conf.level = 0.95,
  conf.method = "perc", ...)

Arguments

x
A boot::boot() object.

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

conf.method
Passed to the type argument of boot::boot.ci(). Defaults to "perc".

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row per bootstrapped statistic and columns:

term

Name of the computed statistic, if present.

statistic

Original value of the statistic.

bias

Bias of the statistic.

std.error

Standard error of the statistic.

If weights were provided to the boot function, an estimate column is included showing the weighted bootstrap estimate, and the standard error is of that estimate. If there are no original statistics in the "boot" object, such as with a call to tsboot with orig.t = FALSE, the original and statistic columns are omitted, and only estimate and std.error columns shown.
See also

Examples

if (require("boot")) {
   clotting <- data.frame(
          u = c(5,10,15,20,30,40,60,80,100),
          lot1 = c(118,58,42,35,27,25,21,19,18),
          lot2 = c(69,35,26,21,18,16,13,12,12))

   g1 <- glm(lot2 ~ log(u), data = clotting, family = Gamma)

   bootfun <- function(d, i) {
      coef(update(g1, data= d[i,]))
   }
   bootres <- boot(clotting, bootfun, R = 999)
   tidy(g1, conf.int=TRUE)
   tidy(bootres, conf.int=TRUE)
}

#> Loading required package: boot

#> 
#> Attaching package: 'boot'

#> The following object is masked from 'package:survival':
#> 
#>     aml

#> The following object is masked from 'package:car':
#> 
#>     logit

#> # A tibble: 2 x 6
#>   term        statistic      bias std.error conf.low conf.high
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1 (Intercept)   -0.0239 -0.00169    0.00320  -0.0326   -0.0222
#> 2 log(u)         0.0236  0.000505   0.00102   0.0227    0.0264