Tidy a(n) btergm object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

This method tidies the coefficients of a bootstrapped temporal exponential random graph model estimated with the xergm. It simply returns the coefficients and their confidence intervals.

# S3 method for btergm
tidy(x, conf.level = 0.95, exponentiate = FALSE, ...)

Arguments

x
A btergm::btergm() object.

conf.level
Confidence level for confidence intervals. Defaults to 0.95.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

conf.high

Upper bound on the confidence interval for the estimate.

conf.low

Lower bound on the confidence interval for the estimate.

estimate

The estimated value of the regression term.

term

The name of the regression term.

Examples

library(btergm)

#> Loading required package: xergm.common

#> Loading required package: ergm

#> Loading required package: network

#> network: Classes for Relational Data
#> Version 1.16.0 created on 2019-11-30.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> ergm: version 3.10.4, created on 2019-06-10
#> Copyright (c) 2019, Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Carter T. Butts, University of California -- Irvine
#>                     Steven M. Goodreau, University of Washington
#>                     Pavel N. Krivitsky, University of Wollongong
#>                     Martina Morris, University of Washington
#>                     with contributions from
#>                     Li Wang
#>                     Kirk Li, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>                     Chad Klumb
#> Based on "statnet" project software (statnet.org).
#> For license and citation information see statnet.org/attribution
#> or type citation("ergm").

#> NOTE: Versions before 3.6.1 had a bug in the implementation of the bd()
#> constriant which distorted the sampled distribution somewhat. In
#> addition, Sampson's Monks datasets had mislabeled vertices. See the
#> NEWS and the documentation for more details.

#> NOTE: Some common term arguments pertaining to vertex attribute and
#> level selection have changed in 3.10.0. See terms help for more
#> details. Use ‘options(ergm.term=list(version="3.9.4"))’ to use old
#> behavior.

#> 
#> Attaching package: ‘xergm.common’

#> The following object is masked from ‘package:ergm’:
#> 
#>     gof

#> Error: package or namespace load failed for ‘btergm’:
#>  .onLoad failed in loadNamespace() for 'tcltk', details:
#>   call: fun(libname, pkgname)
#>   error: X11 library is missing: install XQuartz from xquartz.macosforge.org

set.seed(1)

# Create 10 random networks with 10 actors

networks <- list()

for (i in 1:10) {
  mat <- matrix(rbinom(100, 1, .25), nrow = 10, ncol = 10)
  diag(mat) <- 0
  nw <- network::network(mat)
  networks[[i]] <- nw
}

# Create 10 matrices as covariates

covariates <- list()

for (i in 1:10) {
  mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
  covariates[[i]] <- mat
}

# Fit a model where the propensity to form ties depends
# on the edge covariates, controlling for the number of
# in-stars
btfit <- btergm(networks ~ edges + istar(2) + edgecov(covariates), R = 100)

#> Error in btergm(networks ~ edges + istar(2) + edgecov(covariates), R = 100): could not find function "btergm"

# Show terms, coefficient estimates and errors
tidy(btfit)

#> Error in tidy(btfit): object 'btfit' not found

# Show coefficients as odds ratios with a 99% CI
tidy(btfit, exponentiate = TRUE, conf.level = 0.99)

#> Error in tidy(btfit, exponentiate = TRUE, conf.level = 0.99): object 'btfit' not found