Tidy a(n) btergm object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

This method tidies the coefficients of a bootstrapped temporal exponential random graph model estimated with the xergm. It simply returns the coefficients and their confidence intervals.

# S3 method for btergm
tidy(x, conf.level = 0.95, exponentiate = FALSE,
  quick = FALSE, ...)

Arguments

x
A btergm::btergm() object.

conf.level
Confidence level for confidence intervals. Defaults to 0.95.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

quick
Logical indiciating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

conf.high

The upper end of a confidence interval for the term under consideration. Included only if `conf.int = TRUE`.

conf.low

The lower end of a confidence interval for the term under consideration. Included only if `conf.int = TRUE`.

estimate

The estimated value of the regression term.

term

The name of the regression term.

Examples

library(xergm)

#> Loading required package: xergm.common

#> Loading required package: ergm

#> Loading required package: statnet.common

#> 
#> Attaching package: ‘statnet.common’

#> The following object is masked from ‘package:base’:
#> 
#>     order

#> Loading required package: network

#> network: Classes for Relational Data
#> Version 1.13.0.1 created on 2015-08-31.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> ergm: version 3.8.0, created on 2017-08-18
#> Copyright (c) 2017, Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Carter T. Butts, University of California -- Irvine
#>                     Steven M. Goodreau, University of Washington
#>                     Pavel N. Krivitsky, University of Wollongong
#>                     Martina Morris, University of Washington
#>                     with contributions from
#>                     Li Wang
#>                     Kirk Li, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#> Based on "statnet" project software (statnet.org).
#> For license and citation information see statnet.org/attribution
#> or type citation("ergm").

#> NOTE: Versions before 3.6.1 had a bug in the implementation of the bd()
#> constriant which distorted the sampled distribution somewhat. In
#> addition, Sampson's Monks datasets had mislabeled vertices. See the
#> NEWS and the documentation for more details.

#> 
#> Attaching package: ‘ergm’

#> The following objects are masked from ‘package:statnet.common’:
#> 
#>     colMeans.mcmc.list, sweep.mcmc.list

#> 
#> Attaching package: ‘xergm.common’

#> The following object is masked from ‘package:ergm’:
#> 
#>     gof

#> Loading required package: btergm

#> Warning: replacing previous import ‘statnet.common::colMeans.mcmc.list’ by ‘ergm::colMeans.mcmc.list’ when loading ‘ergm.count’

#> Warning: replacing previous import ‘statnet.common::sweep.mcmc.list’ by ‘ergm::sweep.mcmc.list’ when loading ‘ergm.count’

#> Warning: replacing previous import ‘statnet.common::colMeans.mcmc.list’ by ‘ergm::colMeans.mcmc.list’ when loading ‘tergm’

#> Warning: replacing previous import ‘statnet.common::sweep.mcmc.list’ by ‘ergm::sweep.mcmc.list’ when loading ‘tergm’

#> Warning: replacing previous import ‘statnet.common::colMeans.mcmc.list’ by ‘ergm::colMeans.mcmc.list’ when loading ‘statnet’

#> Warning: replacing previous import ‘statnet.common::sweep.mcmc.list’ by ‘ergm::sweep.mcmc.list’ when loading ‘statnet’

#> Error: package or namespace load failed for ‘btergm’:
#>  .onLoad failed in loadNamespace() for 'tcltk', details:
#>   call: fun(libname, pkgname)
#>   error: X11 library is missing: install XQuartz from xquartz.macosforge.org

#> Error: package ‘btergm’ could not be loaded

set.seed(1)
# Using the same simulated example as the xergm package
# Create 10 random networks with 10 actors
networks <- list()
for(i in 1:10){
    mat <- matrix(rbinom(100, 1, .25), nrow = 10, ncol = 10)
    diag(mat) <- 0
    nw <- network::network(mat)
    networks[[i]] <- nw
}
# Create 10 matrices as covariates
covariates <- list()
for (i in 1:10) {
    mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
    covariates[[i]] <- mat
}
# Fit a model where the propensity to form ties depends
# on the edge covariates, controlling for the number of
# in-stars
suppressWarnings(btfit <- btergm(networks ~ edges + istar(2) +
                   edgecov(covariates), R = 100))

#> Error in btergm(networks ~ edges + istar(2) + edgecov(covariates), R = 100): could not find function "btergm"

# Show terms, coefficient estimates and errors
tidy(btfit)

#> Error in tidy(btfit): object 'btfit' not found

# Show coefficients as odds ratios with a 99% CI
tidy(btfit, exponentiate = TRUE, conf.level = 0.99)

#> Error in tidy(btfit, exponentiate = TRUE, conf.level = 0.99): object 'btfit' not found