Tidy a(n) btergm object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

This method tidies the coefficients of a bootstrapped temporal exponential random graph model estimated with the xergm. It simply returns the coefficients and their confidence intervals.

# S3 method for btergm
tidy(x, conf.level = 0.95, exponentiate = FALSE,
  quick = FALSE, ...)

Arguments

x
A btergm::btergm() object.

conf.level
Confidence level for confidence intervals. Defaults to 0.95.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

quick
Logical indiciating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row per term in the random graph model and columns:

term

The term in the model being estimated and tested.

estimate

The estimated value of the coefficient.

conf.low

The lower bound of the confidence interval.

conf.high

The lower bound of the confidence interval.

See also

Examples

if (require("xergm")) {
    set.seed(1)
    # Using the same simulated example as the xergm package
    # Create 10 random networks with 10 actors
    networks <- list()
    for(i in 1:10){
        mat <- matrix(rbinom(100, 1, .25), nrow = 10, ncol = 10)
        diag(mat) <- 0
        nw <- network::network(mat)
        networks[[i]] <- nw
    }
    # Create 10 matrices as covariates
    covariates <- list()
    for (i in 1:10) {
        mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
        covariates[[i]] <- mat
    }
    # Fit a model where the propensity to form ties depends
    # on the edge covariates, controlling for the number of
    # in-stars
    suppressWarnings(btfit <- btergm(networks ~ edges + istar(2) +
                       edgecov(covariates), R = 100))

    # Show terms, coefficient estimates and errors
    tidy(btfit)

    # Show coefficients as odds ratios with a 99% CI
    tidy(btfit, exponentiate = TRUE, conf.level = 0.99)
}

#> Loading required package: xergm

#> Loading required package: xergm.common

#> Loading required package: ergm

#> Loading required package: statnet.common

#> 
#> Attaching package: 'statnet.common'

#> The following object is masked from 'package:base':
#> 
#>     order

#> Loading required package: network

#> network: Classes for Relational Data
#> Version 1.13.0.1 created on 2015-08-31.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> ergm: version 3.8.0, created on 2017-08-18
#> Copyright (c) 2017, Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Carter T. Butts, University of California -- Irvine
#>                     Steven M. Goodreau, University of Washington
#>                     Pavel N. Krivitsky, University of Wollongong
#>                     Martina Morris, University of Washington
#>                     with contributions from
#>                     Li Wang
#>                     Kirk Li, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#> Based on "statnet" project software (statnet.org).
#> For license and citation information see statnet.org/attribution
#> or type citation("ergm").

#> NOTE: Versions before 3.6.1 had a bug in the implementation of the bd()
#> constriant which distorted the sampled distribution somewhat. In
#> addition, Sampson's Monks datasets had mislabeled vertices. See the
#> NEWS and the documentation for more details.

#> 
#> Attaching package: 'ergm'

#> The following objects are masked from 'package:statnet.common':
#> 
#>     colMeans.mcmc.list, sweep.mcmc.list

#> 
#> Attaching package: 'xergm.common'

#> The following object is masked from 'package:ergm':
#> 
#>     gof

#> Loading required package: btergm

#> Warning: replacing previous import 'statnet.common::colMeans.mcmc.list' by 'ergm::colMeans.mcmc.list' when loading 'ergm.count'

#> Warning: replacing previous import 'statnet.common::sweep.mcmc.list' by 'ergm::sweep.mcmc.list' when loading 'ergm.count'

#> Warning: replacing previous import 'statnet.common::colMeans.mcmc.list' by 'ergm::colMeans.mcmc.list' when loading 'tergm'

#> Warning: replacing previous import 'statnet.common::sweep.mcmc.list' by 'ergm::sweep.mcmc.list' when loading 'tergm'

#> Warning: replacing previous import 'statnet.common::colMeans.mcmc.list' by 'ergm::colMeans.mcmc.list' when loading 'statnet'

#> Warning: replacing previous import 'statnet.common::sweep.mcmc.list' by 'ergm::sweep.mcmc.list' when loading 'statnet'

#> Warning: replacing previous import 'statnet.common::sweep.mcmc.list' by 'ergm::sweep.mcmc.list' when loading 'btergm'

#> Warning: replacing previous import 'statnet.common::colMeans.mcmc.list' by 'ergm::colMeans.mcmc.list' when loading 'btergm'

#> Package:  btergm
#> Version:  1.9.1
#> Date:     2018-02-14
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)
#>           Bruce A. Desmarais (Pennsylvania State University)

#> Loading required package: tnam

#> Package:  tnam
#> Version:  1.6.5
#> Date:     2017-03-31
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)

#> Loading required package: rem

#> Loading required package: GERGM

#> GERGM: Generalized Exponential Random Graph Models
#> Version 0.13.0 created on 2018-05-14.
#> copyright (c) 2018, Matthew J. Denny, Penn State University
#>                     James D. Wilson, University of San Francisco
#>                     Skyler Cranmer, Ohio State University
#>                     Bruce A. Desmarais, Penn State University
#>                     Shankar Bhamidi, University of North Carolina
#> Type help('gergm') to get started.
#> Development website: https://github.com/matthewjdenny/GERGM

#> Package:  xergm
#> Version:  1.8.2
#> Date:     2017-04-01
#> Authors:  Philip Leifeld (University of Glasgow)
#>           Skyler J. Cranmer (The Ohio State University)
#>           Bruce A. Desmarais (Pennsylvania State University)
#> 
#> Please cite the xergm package in your publications -- see citation("xergm").

#> 
#> Initial dimensions of the network and covariates:

#>                  t=1 t=2 t=3 t=4 t=5 t=6 t=7 t=8 t=9 t=10
#> networks (row)    10  10  10  10  10  10  10  10  10   10
#> networks (col)    10  10  10  10  10  10  10  10  10   10
#> covariates (row)  10  10  10  10  10  10  10  10  10   10
#> covariates (col)  10  10  10  10  10  10  10  10  10   10

#> 
#> All networks are conformable.

#> 
#> Dimensions of the network and covariates after adjustment:

#>                  t=1 t=2 t=3 t=4 t=5 t=6 t=7 t=8 t=9 t=10
#> networks (row)    10  10  10  10  10  10  10  10  10   10
#> networks (col)    10  10  10  10  10  10  10  10  10   10
#> covariates (row)  10  10  10  10  10  10  10  10  10   10
#> covariates (col)  10  10  10  10  10  10  10  10  10   10

#> 
#> Starting pseudolikelihood estimation with 100 bootstrapping replications on a single computing core...

#> Done.

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> # A tibble: 3 x 4
#>   term                    estimate conf.low conf.high
#>   <chr>                      <dbl>    <dbl>     <dbl>
#> 1 edges                      0.311    0.201     0.512
#> 2 istar2                     1.04     0.831     1.25 
#> 3 edgecov.covariates[[i]]    1.00     0.843     1.13