tidy.coeftest {broom}	R Documentation

Tidy a(n) coeftest object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'coeftest'
tidy(x, ...)

Arguments

x
A coeftest object returned from lmtest::coeftest().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble with one row for each coefficient and columns:

term
The term in the linear model being estimated and tested

estimate
The estimated coefficient

std.error
The standard error

statistic
test statistic

p.value
p-value

See Also

tidy(), lmtest::coeftest()

Examples

if (require("lmtest", quietly = TRUE)) {
    data(Mandible)
    fm <- lm(length ~ age, data=Mandible, subset=(age <= 28))

    lmtest::coeftest(fm)
    tidy(coeftest(fm))
}

[Package broom version 0.5.0 Index]