Tidy a(n) ergm object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

The methods should work with any model that conforms to the ergm class, such as those produced from weighted networks by the ergm.count package.

# S3 method for ergm
tidy(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

Arguments

x
An ergm object returned from a call to ergm::ergm().

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

exponentiate
Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

...
Additional arguments to pass to ergm::summary(). Cautionary note: Mispecified arguments may be silently ignored.

Value

A tibble::tibble with one row for each coefficient in the exponential random graph model, with columns:

term

The term in the model being estimated and tested

estimate

The estimated coefficient

std.error

The standard error

mcmc.error

The MCMC error

p.value

The two-sided p-value

References

Hunter DR, Handcock MS, Butts CT, Goodreau SM, Morris M (2008b). ergm: A Package to Fit, Simulate and Diagnose Exponential-Family Models for Networks. Journal of Statistical Software, 24(3). http://www.jstatsoft.org/v24/i03/.

See also

Examples

library(ergm)

#> Error in library(ergm): there is no package called ‘ergm’

# Using the same example as the ergm package
# Load the Florentine marriage network data
data(florentine)

#> Warning: data set ‘florentine’ not found

# Fit a model where the propensity to form ties between
# families depends on the absolute difference in wealth
gest <- ergm(flomarriage ~ edges + absdiff("wealth"))

#> Error in ergm(flomarriage ~ edges + absdiff("wealth")): could not find function "ergm"

# Show terms, coefficient estimates and errors
tidy(gest)

#> Error in tidy(gest): object 'gest' not found

# Show coefficients as odds ratios with a 99% CI
tidy(gest, exponentiate = TRUE, conf.int = TRUE, conf.level = 0.99)

#> Error in tidy(gest, exponentiate = TRUE, conf.int = TRUE, conf.level = 0.99): object 'gest' not found

# Take a look at likelihood measures and other
# control parameters used during MCMC estimation
glance(gest)

#> Error in glance(gest): object 'gest' not found

glance(gest, deviance = TRUE)

#> Error in glance(gest, deviance = TRUE): object 'gest' not found

glance(gest, mcmc = TRUE)

#> Error in glance(gest, mcmc = TRUE): object 'gest' not found