tidy.fitdistr {broom}	R Documentation

Tidy a(n) fitdistr object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'fitdistr'
tidy(x, ...)

Arguments

x
A fitdistr object returned by MASS::fitdistr().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

estimate
The estimated value of the regression term.

std.error
The standard error of the regression term.

term
The name of the regression term.

See Also

tidy(), MASS::fitdistr()

Other fitdistr tidiers: glance.fitdistr()

Examples

if (requireNamespace("MASS", quietly = TRUE)) {

set.seed(2015)
x <- rnorm(100, 5, 2)

library(MASS)
fit <- fitdistr(x, dnorm, list(mean = 3, sd = 1))

tidy(fit)
glance(fit)

}

[Package broom version 0.7.12 Index]