Tidy a(n) glmRob object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for glmRob
tidy(x, ...)

Arguments

x
A glmRob object returned from robust::glmRob().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

For tidiers for robust models from the MASS package see tidy.rlm().

See also

Examples

library(robust)

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")

tidy(gm)

#> # A tibble: 2 x 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept)    12.0       4.51      2.67 0.00759
#> 2 wt             -4.02      1.44     -2.80 0.00509

glance(gm)

#> # A tibble: 1 x 5
#>   deviance sigma null.deviance df.residual  nobs
#>      <dbl> <dbl>         <dbl>       <int> <int>
#> 1     19.2 0.800          44.4          30    32