tidy.lavaan {broom}	R Documentation

Tidy a(n) lavaan object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'lavaan'
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x
A lavaan object, such as those returned from lavaan::cfa(), and lavaan::sem().

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments passed to lavaan::parameterEstimates(). Cautionary note: Misspecified arguments may be silently ignored.

Value

A tibble::tibble() with one row for each estimated parameter and columns:

term
The result of paste(lhs, op, rhs)

op
The operator in the model syntax (e.g. ⁠~~⁠ for covariances, or ~ for regression parameters)

group
The group (if specified) in the lavaan model

estimate
The parameter estimate (may be standardized)

std.error

statistic
The z value returned by lavaan::parameterEstimates()

p.value

conf.low

conf.high

std.lv
Standardized estimates based on the variances of the (continuous) latent variables only

std.all
Standardized estimates based on both the variances of both (continuous) observed and latent variables.

std.nox
Standardized estimates based on both the variances of both (continuous) observed and latent variables, but not the variances of exogenous covariates.

See Also

tidy(), lavaan::cfa(), lavaan::sem(), lavaan::parameterEstimates()

Other lavaan tidiers: glance.lavaan()

Examples

# load libraries for models and data
library(lavaan)

cfa.fit <- cfa("F =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9",
  data = HolzingerSwineford1939, group = "school"
)

tidy(cfa.fit)

[Package broom version 1.0.5 Index]