tidy.manova {broom}	R Documentation

Tidy a(n) manova object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'manova'
tidy(x, test = "Pillai", ...)

Arguments

x
A manova object return from stats::manova().

test
One of "Pillai" (Pillai's trace), "Wilks" (Wilk's lambda), "Hotelling-Lawley" (Hotelling-Lawley trace) or "Roy" (Roy's greatest root) indicating which test statistic should be used. Defaults to "Pillai".

...
Arguments passed on to stats::summary.manova

object

An object of class "manova" or an aov object with multiple responses.

intercept

logical. If TRUE, the intercept term is included in the table.

tol

tolerance to be used in deciding if the residuals are rank-deficient: see qr.

Details

Depending on which test statistic is specified only one of pillai, wilks, hl or roy is included.

Value

A tibble::tibble() with columns:

den.df
Degrees of freedom of the denominator.

num.df
Degrees of freedom.

p.value
The two-sided p-value associated with the observed statistic.

statistic
The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

term
The name of the regression term.

pillai
Pillai's trace.

wilks
Wilk's lambda.

hl
Hotelling-Lawley trace.

roy
Roy's greatest root.

See Also

tidy(), stats::summary.manova()

Other anova tidiers: glance.aov(), tidy.TukeyHSD(), tidy.anova(), tidy.aovlist(), tidy.aov()

Examples

npk2 <- within(npk, foo <- rnorm(24))
m <- manova(cbind(yield, foo) ~ block + N * P * K, npk2)
tidy(m)

[Package broom version 0.7.12 Index]