tidy.manova {broom}	R Documentation

tidy a MANOVA object

Description

Constructs a data frame with one row for each of the terms in the model, containing the information from summary.manova.

Usage

## S3 method for class 'manova'
tidy(x, test = "Pillai", ...)

Arguments

x
object of class "manova"

test
one of "Pillai" (Pillai's trace), "Wilks" (Wilk's lambda), "Hotelling-Lawley" (Hotelling-Lawley trace) or "Roy" (Roy's greatest root) indicating which test statistic should be used. Defaults to "Pillai"

...
additional arguments passed on to summary.manova

Value

A data.frame with the columns

term
Term in design

statistic
Approximate F statistic

num.df
Degrees of freedom

p.value
P-value

Depending on which test statistic is specified, one of the following columns is also included:

pillai
Pillai's trace

wilks
Wilk's lambda

hl
Hotelling-Lawley trace

roy
Roy's greatest root

See Also

summary.manova

Examples

npk2 <- within(npk, foo <- rnorm(24))
npk2.aov <- manova(cbind(yield, foo) ~ block + N*P*K, npk2)

[Package broom version 0.4.5 Index]