tidy.muhaz {broom}	R Documentation

Tidy a(n) muhaz object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'muhaz'
tidy(x, ...)

Arguments

x
A muhaz object returned by muhaz::muhaz().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

time
Point in time.

estimate
Estimated hazard rate.

See Also

tidy(), muhaz::muhaz()

Other muhaz tidiers: glance.muhaz()

Examples

if (requireNamespace("muhaz", quietly = TRUE)) {
  if (requireNamespace("survival", quietly = TRUE)) {

library(muhaz)
library(survival)

x <- muhaz(ovarian$futime, ovarian$fustat)
tidy(x)
glance(x)

  }
}

[Package broom version 0.7.12 Index]