Tidy a(n) orcutt object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for orcutt
tidy(x, ...)

Arguments

x
An orcutt object returned from orcutt::cochrane.orcutt().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

See also

Value

A tibble::tibble() with columns:

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

Examples

library(orcutt)

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

#> # A tibble: 4 x 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept) 19.8         5.94      3.33   0.00244 
#> 2 wt          -5.03        1.22     -4.11   0.000310
#> 3 qsec         0.927       0.342     2.71   0.0114  
#> 4 disp        -0.000128    0.0106   -0.0121 0.990   

co <- cochrane.orcutt(reg)
co

#> Cochrane-orcutt estimation for first order autocorrelation 
#>  
#> Call:
#> lm(formula = mpg ~ wt + qsec + disp, data = mtcars)
#> 
#>  number of interaction: 7
#>  rho 0.26819
#> 
#> Durbin-Watson statistic 
#> (original):    1.49575 , p-value: 4.063e-02
#> (transformed): 2.05696 , p-value: 5.21e-01
#>  
#>  coefficients: 
#> (Intercept)          wt        qsec        disp 
#>   21.814858   -4.852590    0.797029   -0.001359 

tidy(co)

#> # A tibble: 4 x 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept) 21.8        6.63       3.29  0.00279
#> 2 wt          -4.85       1.33      -3.65  0.00112
#> 3 qsec         0.797      0.370      2.15  0.0402 
#> 4 disp        -0.00136    0.0110    -0.123 0.903  

glance(co)

#> # A tibble: 1 x 9
#>   r.squared adj.r.squared   rho number.interaction dw.original p.value.original
#>       <dbl>         <dbl> <dbl>              <dbl>       <dbl>            <dbl>
#> 1     0.799         0.777 0.268                  7        1.50           0.0406
#> # … with 3 more variables: dw.transformed <dbl>, p.value.transformed <dbl>,
#> #   nobs <int>