Tidy a(n) orcutt object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

# S3 method for class 'orcutt'
tidy(x, ...)

Arguments

x

An orcutt object returned from orcutt::cochrane.orcutt().

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

See also

Value

A tibble::tibble() with columns:

estimate

The estimated value of the regression term.

p.value

The two-sided p-value associated with the observed statistic.

statistic

The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

std.error

The standard error of the regression term.

term

The name of the regression term.

Examples

# load libraries for models and data
library(orcutt)

# fit model and summarize results
reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)
#> # A tibble: 4 × 5
#>   term         estimate std.error statistic  p.value
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept) 19.8         5.94      3.33   0.00244 
#> 2 wt          -5.03        1.22     -4.11   0.000310
#> 3 qsec         0.927       0.342     2.71   0.0114  
#> 4 disp        -0.000128    0.0106   -0.0121 0.990   

co <- cochrane.orcutt(reg)
tidy(co)
#> # A tibble: 4 × 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept) 21.8        6.63       3.29  0.00279
#> 2 wt          -4.85       1.33      -3.65  0.00112
#> 3 qsec         0.797      0.370      2.15  0.0402 
#> 4 disp        -0.00136    0.0110    -0.123 0.903  
glance(co)
#> # A tibble: 1 × 9
#>   r.squared adj.r.squared   rho number.interaction dw.original
#>       <dbl>         <dbl> <dbl>              <dbl>       <dbl>
#> 1     0.799         0.777 0.268                  7        1.50
#> # ℹ 4 more variables: p.value.original <dbl>, dw.transformed <dbl>,
#> #   p.value.transformed <dbl>, nobs <int>