tidy.regsubsets {broom}	R Documentation

Tidy a(n) regsubsets object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'regsubsets'
tidy(x, ...)

Arguments

x
A regsubsets object created by leaps::regsubsets().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

r.squared
R squared statistic, or the percent of variation explained by the model.

adj.r.squared
Adjusted R squared statistic

BIC
Bayesian information criterion for the component.

mallows_cp
Mallow's Cp statistic.

See Also

tidy(), leaps::regsubsets()

Examples

if (requireNamespace("leaps", quietly = TRUE)) {

all_fits <- leaps::regsubsets(hp ~ ., mtcars)
tidy(all_fits)

}

[Package broom version 0.7.12 Index]