Tidy a(n) lmrob object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

# S3 method for lmrob
tidy(x, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x

A lmrob object returned from robustbase::lmrob().

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...

Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Two exceptions here are:

tidy() methods will warn when supplied an exponentiate argument if it will be ignored.

augment() methods will warn when supplied a newdata argument if it will be ignored.

Details

For tidiers for robust models from the MASS package see tidy.rlm().

See also

Examples

if (requireNamespace("robustbase", quietly = TRUE)) {

# load libraries for models and data
library(robustbase)

data(coleman)
set.seed(0)

m <- lmrob(Y ~ ., data = coleman)
tidy(m)
augment(m)
glance(m)

data(carrots)

Rfit <- glmrob(cbind(success, total - success) ~ logdose + block,
  family = binomial, data = carrots, method = "Mqle",
  control = glmrobMqle.control(tcc = 1.2)
)

tidy(Rfit)
augment(Rfit)

}
#> # A tibble: 24 × 4
#>    `cbind(success, total - succe…`[,"success"] [,""] logdose block .fitted
#>                                          <int> <int>   <dbl> <fct>   <dbl>
#>  1                                          10    25    1.52 B1     -0.726
#>  2                                          16    26    1.64 B1     -0.972
#>  3                                           8    42    1.76 B1     -1.22 
#>  4                                           6    36    1.88 B1     -1.46 
#>  5                                           9    26    2    B1     -1.71 
#>  6                                           9    33    2.12 B1     -1.96 
#>  7                                           1    31    2.24 B1     -2.20 
#>  8                                           2    26    2.36 B1     -2.45 
#>  9                                          17    21    1.52 B2     -0.491
#> 10                                          10    30    1.64 B2     -0.737
#> # … with 14 more rows