Tidy a(n) rq object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for rq
tidy(x, se.type = "rank", conf.int = TRUE,
  conf.level = 0.95, ...)

Arguments

x
An rq object returned from quantreg::rq().

se.type
Character specifying the method to use to calculate standard errors. Passed to quantreg::summary.rq() se argument. Defaults to "rank".

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments passed to quantreg::summary.rq().

Value

A tibble::tibble() with one row for each term in the regression. The tibble has columns:

term

The name of the regression term.

estimate

The estimated value of the regression term.

std.error

The standard error of the regression term.

statistic

The value of a statistic, almost always a T-statistic, to use in a hypothesis that the regression term is non-zero.

p.value

The two-sided p-value associated with the observed statistic.

conf.low

The low end of a confidence interval for the regression term. Included only if conf.int = TRUE.

conf.high

The high end of a confidence interval for the regression term. Included only if conf.int = TRUE.

Details

If se.type = "rank" confidence intervals are calculated by summary.rq. When only a single predictor is included in the model, no confidence intervals are calculated and the confidence limits are set to NA.

See also