Tidy a(n) speedlm object

This method wraps tidy.lm().

# S3 method for speedlm
tidy(x, ...)

Arguments

x
A speedlm object returned from speedglm::speedlm().

...
Arguments passed on to tidy.lm

x

An lm object created by stats::lm().

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

quick

Logical indiciating if the only the term and estimate columns should be returned. Often useful to avoid time consuming covariance and standard error calculations. Defaults to FALSE.

exponentiate

Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

Value

A tibble::tibble() with one row for each term in the regression. The tibble has columns:

term

The name of the regression term.

estimate

The estimated value of the regression term.

std.error

The standard error of the regression term.

statistic

The value of a statistic, almost always a T-statistic, to use in a hypothesis that the regression term is non-zero.

p.value

The two-sided p-value associated with the observed statistic.

conf.low

The low end of a confidence interval for the regression term. Included only if conf.int = TRUE.

conf.high

The high end of a confidence interval for the regression term. Included only if conf.int = TRUE.

If the linear model is an mlm object (multiple linear model), there is an additional column:
response

Which response column the coefficients correspond to (typically Y1, Y2, etc)

See also

Examples

mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars)

tidy(mod)

#> # A tibble: 3 x 5
#>   term        estimate std.error statistic  p.value
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)   19.7       5.25       3.76 7.65e- 4
#> 2 wt            -5.05      0.484    -10.4  2.52e-11
#> 3 qsec           0.929     0.265      3.51 1.50e- 3

glance(mod)

#> # A tibble: 1 x 10
#>   r.squared adj.r.squared statistic  p.value    df logLik   AIC   BIC
#>       <dbl>         <dbl>     <dbl>    <dbl> <int>  <dbl> <dbl> <dbl>
#> 1     0.826         0.814      69.0 9.39e-12     3  -74.4  157.  163.
#> # ... with 2 more variables: df.residual <int>, deviance <dbl>

augment(mod)

#> # A tibble: 32 x 5
#>    .rownames           mpg    wt  qsec .fitted
#>    <chr>             <dbl> <dbl> <dbl>   <dbl>
#>  1 Mazda RX4          21    2.62  16.5    21.8
#>  2 Mazda RX4 Wag      21    2.88  17.0    21.0
#>  3 Datsun 710         22.8  2.32  18.6    25.3
#>  4 Hornet 4 Drive     21.4  3.22  19.4    21.6
#>  5 Hornet Sportabout  18.7  3.44  17.0    18.2
#>  6 Valiant            18.1  3.46  20.2    21.1
#>  7 Duster 360         14.3  3.57  15.8    16.4
#>  8 Merc 240D          24.4  3.19  20      22.2
#>  9 Merc 230           22.8  3.15  22.9    25.1
#> 10 Merc 280           19.2  3.44  18.3    19.4
#> # ... with 22 more rows