tidy.survreg {broom}	R Documentation

Tidy a(n) survreg object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'survreg'
tidy(x, conf.level = 0.95, ...)

Arguments

x
An survreg object returned from survival::survreg().

conf.level
confidence level for CI

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with one row for each term in the regression. The tibble has columns:

term
The name of the regression term.

estimate
The estimated value of the regression term.

std.error
The standard error of the regression term.

statistic
The value of a statistic, almost always a T-statistic, to use in a hypothesis that the regression term is non-zero.

p.value
The two-sided p-value associated with the observed statistic.

conf.low
The low end of a confidence interval for the regression term. Included only if conf.int = TRUE.

conf.high
The high end of a confidence interval for the regression term. Included only if conf.int = TRUE.

See Also

tidy(), survival::survreg()

Other survreg tidiers: augment.survreg, glance.survreg

Other survival tidiers: augment.coxph, augment.survreg, glance.aareg, glance.cch, glance.coxph, glance.pyears, glance.survdiff, glance.survexp, glance.survfit, glance.survreg, tidy.aareg, tidy.cch, tidy.coxph, tidy.pyears, tidy.survdiff, tidy.survexp, tidy.survfit

Examples

library(survival)

sr <- survreg(
  Surv(futime, fustat) ~ ecog.ps + rx,
  ovarian,
  dist = "exponential"
)

td <- tidy(sr)
augment(sr, ovarian)
glance(sr)

# coefficient plot
library(ggplot2)
ggplot(td, aes(estimate, term)) + 
  geom_point() +
  geom_errorbarh(aes(xmin = conf.low, xmax = conf.high), height = 0) +
  geom_vline(xintercept = 0)

[Package broom version 0.5.0 Index]