Tidy a(n) table object

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies cross models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

# S3 method for table
tidy(x, ...)

Arguments

x
A table object.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble in long-form containing frequency information for the table in a Freq column. The result is much like what you get from tidyr::gather().

Details

Directly calls tibble::as_tibble() on a table object, which does the same things as as.data.frame.table() but also gives the returned object tibble::tibble class.

See also

Examples

tab <- with(airquality, table(cut(Temp, quantile(Temp)), Month))
tidy(tab)

#> # A tibble: 20 x 3
#>    Var1    Month     n
#>    <chr>   <chr> <int>
#>  1 (56,72] 5        24
#>  2 (72,79] 5         5
#>  3 (79,85] 5         1
#>  4 (85,97] 5         0
#>  5 (56,72] 6         3
#>  6 (72,79] 6        15
#>  7 (79,85] 6         7
#>  8 (85,97] 6         5
#>  9 (56,72] 7         0
#> 10 (72,79] 7         2
#> 11 (79,85] 7        19
#> 12 (85,97] 7        10
#> 13 (56,72] 8         1
#> 14 (72,79] 8         9
#> 15 (79,85] 8         7
#> 16 (85,97] 8        14
#> 17 (56,72] 9        10
#> 18 (72,79] 9        10
#> 19 (79,85] 9         5
#> 20 (85,97] 9         5