tidy.Gam {broom}	R Documentation

Tidy a(n) Gam object

Description

Tidy summarizes information about the components of a model. A model component might be a single term in a regression, a single hypothesis, a cluster, or a class. Exactly what tidy considers to be a model component varies across models but is usually self-evident. If a model has several distinct types of components, you will need to specify which components to return.

Usage

## S3 method for class 'Gam'
tidy(x, ...)

Arguments

x
A Gam object returned from a call to gam::gam().

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.lvel = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Details

Tidy gam objects created by calls to mgcv::gam() with tidy.gam().

Value

A tibble::tibble() with columns:

df
Degrees of freedom used by this term in the model.

meansq
Mean sum of squares. Equal to total sum of squares divided by degrees of freedom.

p.value
The two-sided p-value associated with the observed statistic.

statistic
The value of a T-statistic to use in a hypothesis that the regression term is non-zero.

sumsq
Sum of squares explained by this term.

term
The name of the regression term.

See Also

tidy(), gam::gam(), tidy.anova(), tidy.gam()

Other gam tidiers: glance.Gam()

Examples

if (requireNamespace("gam", quietly = TRUE)) {

library(gam)
g <- gam(mpg ~ s(hp, 4) + am + qsec, data = mtcars)

tidy(g)
glance(g)

}

[Package broom version 0.7.12 Index]