Tidy a(n) irlba object masquerading as list

Broom tidies a number of lists that are effectively S3 objects without a class attribute. For example, stats::optim(), svd() and akima::interp() produce consistent output, but because they do not have a class attribute, they cannot be handled by S3 dispatch.

These functions look at the elements of a list and determine if there is an appropriate tidying method to apply to the list. Those tidiers are themselves are implemented as functions of the form tidy_<function> or glance_<function> and are not exported (but they are documented!).

If no appropriate tidying method is found, throws an error.

tidy_irlba(x, ...)

Arguments

x
A list returned from irlba::irlba().

...
Arguments passed on to tidy_svd

x

A list with components u, d, v returned by svd().

matrix

Character specifying which component of the PCA should be tidied.

"u", "samples", or "x": returns information about the map from the original space into principle components space.

"v", "rotation", or "variables": returns information about the map from principle components space back into the original space.

"d" or "pcs": returns information about the eigenvalues will return information about

Value

A tibble::tibble with columns depending on the component of PCA being tidied.

If matrix is "u", "samples", or "x" each row in the tidied output corresponds to the original data in PCA space. The columns are:

row

ID of the original observation (i.e. rowname from original data).

PC

Integer indicating a principle component.

value

The score of the observation for that particular principle component. That is, the location of the observation in PCA space.

If matrix is "v", "rotation", or "variables", each row in the tidied ouput corresponds to information about the principle components in the original space. The columns are:
row

The variable labels (colnames) of the data set on which PCA was performed

PC

An integer vector indicating the principal component

value

The value of the eigenvector (axis score) on the indicated principal component

If matrix is "d" or "pcs", the columns are:
PC

An integer vector indicating the principal component

std.dev

Standard deviation explained by this PC

percent

Percentage of variation explained

cumulative

Cumulative percentage of variation explained

Details

A very thin wrapper around tidy_svd().

See also